Independent Component Analysis (ICA)

Performs an Independent Component Analysis using the FastICA algorithm. Contrary to PCA, that attempts to find uncorrelated sources (through least squares minimization), ICA attempts to find independent sources, i.e., the source space that maximizes the "non-gaussianity" of all sources. Contrary to PCA, ICA does not rank each source, which makes it a poor tool for dimensionality reduction. Requires the fastICA package to be installed.

ICA(x, n = "all", ...)

Arguments

x
A dataframe or a statistical model.

n
Number of components to extract. If n = "all" =, then n is set as the number of variables minus 1 (ncol(x)-1). If n = "auto" (default) or n = NULL, the number of components is selected through n_factors. In parameters_reduction, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

...
Arguments passed to or from other methods.

Examples

ICA(iris[, 1:4])

#>             ICA1          ICA2         ICA3
#> 1   -1.396802725 -0.4138602931  0.161984020
#> 2   -1.335029693  0.7875331715  0.227035900
#> 3   -1.328550435  0.3236754294 -0.456786702
#> 4   -1.199277967  0.5974651256 -0.657545315
#> 5   -1.368671408 -0.6390302107 -0.196368576
#> 6   -1.250421585 -1.5234853349  0.013491489
#> 7   -1.207955261 -0.2109160891 -1.131274209
#> 8   -1.317313633 -0.1535885038 -0.005892003
#> 9   -1.189649910  1.0774549834 -0.894649590
#> 10  -1.329501198  0.6520667863  0.254046466
#> 11  -1.435349299 -0.9046421332  0.645761309
#> 12  -1.209693223 -0.1184866320 -0.532120622
#> 13  -1.349578857  0.8853027617  0.151935624
#> 14  -1.354336401  0.8987089528 -0.982781677
#> 15  -1.702735093 -1.6962494835  1.396062393
#> 16  -1.440679886 -2.7627065958  0.260877803
#> 17  -1.449555089 -1.5775569628  0.145021850
#> 18  -1.353335045 -0.5008379009 -0.009588718
#> 19  -1.379857616 -1.2329217743  1.036762852
#> 20  -1.305913268 -1.1952056939 -0.377510054
#> 21  -1.333420948 -0.1697206194  0.915034874
#> 22  -1.261658387 -1.0462214018 -0.437403210
#> 23  -1.452130845 -0.6499339090 -1.051530273
#> 24  -1.115475155 -0.1623155958 -0.228022802
#> 25  -1.060343096 -0.0779329112 -0.630768393
#> 26  -1.264381458  0.8037770030  0.407943734
#> 27  -1.180594895 -0.3140258125 -0.381920070
#> 28  -1.375937866 -0.4111343685  0.375774444
#> 29  -1.424934042 -0.1886903755  0.520336616
#> 30  -1.179200307  0.3642291503 -0.555434473
#> 31  -1.207331624  0.5893990678 -0.197081877
#> 32  -1.346052338 -0.3707116490  0.637654578
#> 33  -1.424128745 -1.7399281604 -0.122730308
#> 34  -1.517987190 -2.1087615224  0.366919005
#> 35  -1.286033517  0.5650891785  0.082473728
#> 36  -1.465089361  0.2777815753  0.316114932
#> 37  -1.562260169 -0.4705461296  1.181558668
#> 38  -1.383220572 -0.5412606204 -0.271468852
#> 39  -1.240220485  0.8279751765 -0.973446582
#> 40  -1.346232150 -0.1643804862  0.240781012
#> 41  -1.374199904 -0.5035638255 -0.223379143
#> 42  -1.220160923  2.3819388860 -0.116589233
#> 43  -1.241794885  0.3560513765 -1.196805745
#> 44  -1.094446733 -0.7239429282 -0.836745128
#> 45  -1.063312083 -1.2281116739 -0.680613153
#> 46  -1.262643496  0.7113475460 -0.191209853
#> 47  -1.299597573 -1.0947101791 -0.238819906
#> 48  -1.249848542  0.3479853187 -0.736342307
#> 49  -1.406430782 -0.8938501508  0.399088295
#> 50  -1.366309809  0.0688554892  0.138670169
#> 51   0.220570632 -0.5086626231  2.039811702
#> 52   0.337982662 -0.5579241505  0.453966058
#> 53   0.393310781 -0.3218505346  1.667480351
#> 54   0.269481869  1.7692264715 -0.253416408
#> 55   0.361996320  0.3886493739  1.114474809
#> 56   0.456625716  0.6354225417 -0.482881240
#> 57   0.509148412 -0.8430358620 -0.141724458
#> 58  -0.036680905  1.7643239407 -1.100237730
#> 59   0.245355242  0.3158507073  1.592613718
#> 60   0.346772925  0.7572593040 -1.578843927
#> 61   0.037116129  2.7244153721 -0.472611569
#> 62   0.334799517 -0.0725941596 -0.457392080
#> 63  -0.004726558  2.2121612831  1.606346461
#> 64   0.483198882  0.2963509183  0.154793317
#> 65   0.036706650  0.2885914615 -0.545290524
#> 66   0.158763254 -0.2808784968  1.510120011
#> 67   0.570905196  0.0003355085 -1.296058894
#> 68   0.098957853  1.0674536550  0.521719932
#> 69   0.403691692  1.8232788139  1.077415847
#> 70   0.102270215  1.4469479979  0.145925509
#> 71   0.762328416 -0.7243433413 -1.392765000
#> 72   0.092034769  0.5246650774  0.668223769
#> 73   0.571545080  1.1586727594  0.857519755
#> 74   0.397050720  0.7062680340  0.609618376
#> 75   0.153842147  0.2968809512  1.197915460
#> 76   0.188468971 -0.0341246145  1.375126579
#> 77   0.331339841  0.4702868485  1.960301410
#> 78   0.588653752 -0.2247419787  0.909785837
#> 79   0.456018328  0.1931294789 -0.197687255
#> 80  -0.170036687  1.2331000956  0.584022041
#> 81   0.082192555  1.6801839733  0.043814667
#> 82  -0.011058502  1.7536436742  0.248269996
#> 83   0.086326463  0.8664626254  0.244339636
#> 84   0.799760664  0.6591831127 -0.343196371
#> 85   0.628742229  0.0219194733 -1.789404923
#> 86   0.495550010 -1.0736576288 -0.927657902
#> 87   0.351581063 -0.3273023838  1.239899502
#> 88   0.237267238  1.7369622403  1.588437346
#> 89   0.284836330  0.1202190963 -0.821383057
#> 90   0.267907469  1.2973026716 -0.476775572
#> 91   0.422786093  1.2023900073 -0.548412776
#> 92   0.432628307  0.0468711114  0.075996326
#> 93   0.136897038  1.1159424323  0.323136628
#> 94  -0.064812222  1.9894938583 -0.741885134
#> 95   0.336981305  0.8416227032 -0.519226902
#> 96   0.262233508  0.2099226288 -0.436019894
#> 97   0.306488389  0.3589069209 -0.495913051
#> 98   0.211679181  0.3184649159  0.704569431
#> 99  -0.201187585  1.3792467472 -0.791496251
#> 100  0.257492213  0.5813509139 -0.351350879
#> 101  1.547541427 -1.4501015422 -2.113352776
#> 102  0.988000740  0.4198342539 -1.351260615
#> 103  1.094900793 -0.4941591770  0.914243629
#> 104  1.047282956  0.0485176848 -0.334094919
#> 105  1.262096198 -0.5299027975 -0.704484606
#> 106  1.298791840 -0.4534937402  1.917430570
#> 107  0.864206173  1.0817336693 -2.807517563
#> 108  1.106581420  0.0352232096  1.902457094
#> 109  1.034324440  0.9762331690  1.033550285
#> 110  1.334696553 -2.2415971766 -0.261216981
#> 111  0.825102805 -0.9224967304 -0.354520161
#> 112  0.914056392  0.3821181736  0.063012291
#> 113  0.982522839 -0.5158548577  0.305754947
#> 114  1.012177961  0.8020545214 -1.513264614
#> 115  1.204551945 -0.2510156853 -2.320803888
#> 116  1.083991117 -1.1456017577 -1.181676571
#> 117  0.938875347 -0.2225460869  0.080454118
#> 118  1.356826783 -2.4254406231  1.066211601
#> 119  1.509307612  0.3461603825  2.169028664
#> 120  0.710445606  1.9124523135  0.419656867
#> 121  1.138532036 -1.1454900418 -0.079841860
#> 122  0.988951503  0.0914428970 -2.062093783
#> 123  1.277763418  0.1081335922  2.526152896
#> 124  0.700373724  0.4258161360  0.119442376
#> 125  1.108646509 -1.1859127612 -0.341721994
#> 126  0.983788210 -0.7024242287  1.419393104
#> 127  0.678721665  0.1871283114 -0.206027630
#> 128  0.755849158 -0.2604855992 -0.708942398
#> 129  1.149554682  0.0127547788 -0.490460538
#> 130  0.798860495 -0.0835810270  2.051662925
#> 131  1.022351032  0.1463797054  2.155002132
#> 132  1.062704260 -2.3136230931  2.001350877
#> 133  1.193022362 -0.0742228291 -0.662033276
#> 134  0.668750233  0.4778228735  0.456715829
#> 135  0.933610865  1.1258977808  0.193888751
#> 136  1.107891805 -0.7058304731  1.985371059
#> 137  1.304153043 -1.6531574622 -1.921929258
#> 138  0.967006664 -0.4477160045 -0.277898478
#> 139  0.734984299 -0.2632115237 -0.922732822
#> 140  0.903033747 -0.7761266470  0.473630970
#> 141  1.190840575 -0.9884396918 -0.600198455
#> 142  0.840618980 -0.9906355836  0.229133264
#> 143  0.988000740  0.4198342539 -1.351260615
#> 144  1.267017305 -1.1076622454 -0.392280055
#> 145  1.282517232 -1.5338231927 -1.028012947
#> 146  0.949026590 -0.7195718119 -0.185415774
#> 147  0.795199180  0.8242802350  0.138346211
#> 148  0.876460581 -0.4370550235 -0.164043588
#> 149  1.190037127 -1.5824236859 -1.931264354
#> 150  0.913252944 -0.2118658205 -1.268053608