Compute a linear discriminant analysis on classified cluster groups

Computes linear discriminant analysis (LDA) on classified cluster groups, and determines the goodness of classification for each cluster group. See MASS::lda() for details.

Usage

cluster_discrimination(x, cluster_groups = NULL, ...)

Arguments

x

A data frame

cluster_groups

Group classification of the cluster analysis, which can be retrieved from the cluster_analysis() function.

...

Other arguments to be passed to or from.

See also

Examples

# Retrieve group classification from hierarchical cluster analysis
clustering <- cluster_analysis(iris[, 1:4], n = 3)

# Goodness of group classification
cluster_discrimination(clustering)
#> # Accuracy of Cluster Group Classification via Linear Discriminant Analysis (LDA)
#> 
#>  Group Accuracy
#>      1   82.98%
#>      2   94.34%
#>      3  100.00%
#> 
#> Overall accuracy of classification: 92.67%
#>