Classical Multidimensional Scaling (cMDS)

Also referred to as principal Coordinates Analysis (PCoA), Classical Multidimensional Scaling (cMDS) takes a set of dissimilarities (i.e., a distance matrix) and returns a set of points such that the distances between the points are approximately equal to the dissimilarities.

cmds(x, n = "all", distance = "euclidean", ...)

Arguments

x
A dataframe or a statistical model.

n
Number of components to extract. If n = "all" =, then n is set as the number of variables minus 1 (ncol(x)-1). If n = "auto" (default) or n = NULL, the number of components is selected through n_factors. In parameters_reduction, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

distance
The distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given.

...
Arguments passed to or from other methods.

References

Nguyen, L. H., \& Holmes, S. (2019). Ten quick tips for effective dimensionality reduction. PLOS Computational Biology, 15(6).

Examples

cmds(iris[, 1:4])

#>            CMDS1        CMDS2        CMDS3
#> 1   -2.684125626  0.319397247  0.027914828
#> 2   -2.714141687 -0.177001225  0.210464272
#> 3   -2.888990569 -0.144949426 -0.017900256
#> 4   -2.745342856 -0.318298979 -0.031559374
#> 5   -2.728716537  0.326754513 -0.090079241
#> 6   -2.280859633  0.741330449 -0.168677658
#> 7   -2.820537751 -0.089461385 -0.257892158
#> 8   -2.626144973  0.163384960  0.021879318
#> 9   -2.886382732 -0.578311754 -0.020759570
#> 10  -2.672755798 -0.113774246  0.197632725
#> 11  -2.506947091  0.645068899  0.075318009
#> 12  -2.612755231  0.014729939 -0.102150260
#> 13  -2.786109266 -0.235112000  0.206844430
#> 14  -3.223803744 -0.511394587 -0.061299672
#> 15  -2.644750390  1.178764636  0.151627524
#> 16  -2.386039034  1.338062330 -0.277776903
#> 17  -2.623527875  0.810679514 -0.138183228
#> 18  -2.648296706  0.311849145 -0.026668316
#> 19  -2.199820324  0.872839039  0.120305523
#> 20  -2.587986400  0.513560309 -0.213665172
#> 21  -2.310256215  0.391345936  0.239444043
#> 22  -2.543705229  0.432996063 -0.208457232
#> 23  -3.215939416  0.133468070 -0.292396751
#> 24  -2.302733182  0.098708855 -0.039123259
#> 25  -2.355754049 -0.037281860 -0.125021083
#> 26  -2.506668907 -0.146016880  0.253420042
#> 27  -2.468820073  0.130951489 -0.094910576
#> 28  -2.562319906  0.367718857  0.078494205
#> 29  -2.639534715  0.312039980  0.145908896
#> 30  -2.631989387 -0.196961225 -0.040771079
#> 31  -2.587398477 -0.204318491  0.077222989
#> 32  -2.409932497  0.410924264  0.145524972
#> 33  -2.648862334  0.813363820 -0.225669150
#> 34  -2.598736749  1.093145759 -0.157810813
#> 35  -2.636926878 -0.121322348  0.143049582
#> 36  -2.866241652  0.069364472  0.164332307
#> 37  -2.625238050  0.599370021  0.268350376
#> 38  -2.800684115  0.268643738 -0.093699082
#> 39  -2.980502044 -0.487958344 -0.072927046
#> 40  -2.590006314  0.229043837  0.080082303
#> 41  -2.770102426  0.263527534 -0.077247693
#> 42  -2.849368705 -0.940960574  0.349230377
#> 43  -2.997406547 -0.341926057 -0.192509212
#> 44  -2.405614485  0.188871429 -0.263867946
#> 45  -2.209489238  0.436663142 -0.298742746
#> 46  -2.714451427 -0.250208204  0.097678144
#> 47  -2.538148259  0.503771144 -0.166705637
#> 48  -2.839462168 -0.227945569 -0.083726849
#> 49  -2.543085750  0.579410022  0.017115024
#> 50  -2.703359782  0.107706082  0.089294008
#> 51   1.284825689  0.685160470  0.406568025
#> 52   0.932488532  0.318333638  0.018014187
#> 53   1.464302322  0.504262815  0.338325765
#> 54   0.183317720 -0.827959012  0.179591392
#> 55   1.088103258  0.074590675  0.307757896
#> 56   0.641669084 -0.418246872 -0.041076091
#> 57   1.095060663  0.283468270 -0.169810240
#> 58  -0.749122670 -1.004890961 -0.012302919
#> 59   1.044131826  0.228361900  0.415336085
#> 60  -0.008745404 -0.723081905 -0.281141431
#> 61  -0.507840884 -1.265971191  0.269817183
#> 62   0.511698557 -0.103981235 -0.130547750
#> 63   0.264976508 -0.550036464  0.694146830
#> 64   0.984934510 -0.124817854  0.062114408
#> 65  -0.173925372 -0.254854209 -0.090457691
#> 66   0.927860781  0.467179494  0.314620976
#> 67   0.660283762 -0.352969666 -0.328027528
#> 68   0.236104993 -0.333610767  0.271161837
#> 69   0.944733728 -0.543145551  0.499519046
#> 70   0.045226976 -0.583834377  0.235002105
#> 71   1.116283177 -0.084616852 -0.459620991
#> 72   0.357888418 -0.068925032  0.229853888
#> 73   1.298183875 -0.327787308  0.347854352
#> 74   0.921728922 -0.182737794  0.231071778
#> 75   0.714853326  0.149055944  0.321800937
#> 76   0.900174373  0.328504474  0.316209074
#> 77   1.332024437  0.244440876  0.521702780
#> 78   1.557802155  0.267495447  0.164920984
#> 79   0.813290650 -0.163350301 -0.035424505
#> 80  -0.305583778 -0.368262190  0.318491581
#> 81  -0.068126492 -0.705172132  0.244213810
#> 82  -0.189622472 -0.680286764  0.306420561
#> 83   0.136428712 -0.314032438  0.177242766
#> 84   1.380026436 -0.420954287 -0.016167128
#> 85   0.588006443 -0.484287420 -0.444433499
#> 86   0.806858313  0.194182315 -0.388963063
#> 87   1.220690882  0.407619594  0.237167010
#> 88   0.815095236 -0.372037060  0.614720843
#> 89   0.245957680 -0.268524397 -0.188366812
#> 90   0.166413217 -0.681926725  0.060009226
#> 91   0.464800288 -0.670711545  0.024306856
#> 92   0.890815198 -0.034464444  0.009946933
#> 93   0.230548024 -0.404385848  0.229410241
#> 94  -0.704531759 -1.012248228  0.105691149
#> 95   0.356981495 -0.504910093 -0.016617170
#> 96   0.331934480 -0.212654684 -0.083204291
#> 97   0.376215651 -0.293218929 -0.077996351
#> 98   0.642576008  0.017738190  0.205394967
#> 99  -0.906469865 -0.756093367  0.012599648
#> 100  0.299000842 -0.348897806 -0.010581660
#> 101  2.531192728 -0.009849109 -0.760165427
#> 102  1.415235877 -0.574916348 -0.296322527
#> 103  2.616676016  0.343903151  0.110787883
#> 104  1.971531053 -0.179727904 -0.108424662
#> 105  2.350005920 -0.040260947 -0.285389563
#> 106  3.397038736  0.550836673  0.348437556
#> 107  0.521232244 -1.192758727 -0.545659296
#> 108  2.932587069  0.355500003  0.420239936
#> 109  2.321228817 -0.243831502  0.348304395
#> 110  2.916750967  0.782791949 -0.423335418
#> 111  1.661774154  0.242228408 -0.242440190
#> 112  1.803401953 -0.215637617  0.037648168
#> 113  2.165591796  0.216275585 -0.033326642
#> 114  1.346163579 -0.776818347 -0.281902882
#> 115  1.585928224 -0.539640714 -0.629029326
#> 116  1.904456375  0.119250692 -0.479639820
#> 117  1.949689059  0.041943260 -0.044186168
#> 118  3.487055364  1.175739330 -0.133894874
#> 119  3.795645422  0.257322973  0.513767764
#> 120  1.300791713 -0.761149636  0.344995038
#> 121  2.427817913  0.378196013 -0.219119324
#> 122  1.199001105 -0.606091528 -0.511855509
#> 123  3.499920039  0.460674099  0.573182243
#> 124  1.388766132 -0.204399327  0.064522757
#> 125  2.275430504  0.334990606 -0.286150091
#> 126  2.614090474  0.560901355  0.205534524
#> 127  1.258508161 -0.179704795 -0.045847704
#> 128  1.291132059 -0.116668651 -0.231256463
#> 129  2.123608723 -0.209729477 -0.154180024
#> 130  2.388003016  0.464639805  0.449530192
#> 131  2.841672778  0.375269167  0.498898076
#> 132  3.230673661  1.374165087  0.114548205
#> 133  2.159437642 -0.217277579 -0.208763167
#> 134  1.444161242 -0.143413410  0.153233888
#> 135  1.781294810 -0.499901681  0.172875189
#> 136  3.076499932  0.688085678  0.335592292
#> 137  2.144243314  0.140064201 -0.734878937
#> 138  1.905098149  0.049300526 -0.162180236
#> 139  1.169326339 -0.164990262 -0.281835840
#> 140  2.107611143  0.372287872 -0.027291132
#> 141  2.314154705  0.183651279 -0.322693747
#> 142  1.922267801  0.409203467 -0.113586596
#> 143  1.415235877 -0.574916348 -0.296322527
#> 144  2.563013375  0.277862603 -0.292569525
#> 145  2.418746183  0.304798198 -0.504482664
#> 146  1.944109795  0.187532303 -0.177825091
#> 147  1.527166615 -0.375316983  0.121898172
#> 148  1.764345717  0.078858855 -0.130481631
#> 149  1.900941614  0.116627959 -0.723251563
#> 150  1.390188862 -0.282660938 -0.362909648