Convert data to numeric

Convert data to numeric by converting characters to factors and factors to either numeric levels or dummy variables.

convert_data_to_numeric(x, dummy_factors = TRUE, ...)

data_to_numeric(x, dummy_factors = TRUE, ...)

Arguments

x
A data.frame or a vector.

dummy_factors
Transform factors to dummy factors (all factor levels as different columns filled with a binary 0-1 value).

...
Arguments passed to or from other methods.

Value

Data.frame of numeric variables.

Examples

convert_data_to_numeric(iris)

#>     Sepal.Length Sepal.Width Petal.Length Petal.Width Species.setosa
#> 1            5.1         3.5          1.4         0.2              1
#> 2            4.9         3.0          1.4         0.2              1
#> 3            4.7         3.2          1.3         0.2              1
#> 4            4.6         3.1          1.5         0.2              1
#> 5            5.0         3.6          1.4         0.2              1
#> 6            5.4         3.9          1.7         0.4              1
#> 7            4.6         3.4          1.4         0.3              1
#> 8            5.0         3.4          1.5         0.2              1
#> 9            4.4         2.9          1.4         0.2              1
#> 10           4.9         3.1          1.5         0.1              1
#> 11           5.4         3.7          1.5         0.2              1
#> 12           4.8         3.4          1.6         0.2              1
#> 13           4.8         3.0          1.4         0.1              1
#> 14           4.3         3.0          1.1         0.1              1
#> 15           5.8         4.0          1.2         0.2              1
#> 16           5.7         4.4          1.5         0.4              1
#> 17           5.4         3.9          1.3         0.4              1
#> 18           5.1         3.5          1.4         0.3              1
#> 19           5.7         3.8          1.7         0.3              1
#> 20           5.1         3.8          1.5         0.3              1
#> 21           5.4         3.4          1.7         0.2              1
#> 22           5.1         3.7          1.5         0.4              1
#> 23           4.6         3.6          1.0         0.2              1
#> 24           5.1         3.3          1.7         0.5              1
#> 25           4.8         3.4          1.9         0.2              1
#> 26           5.0         3.0          1.6         0.2              1
#> 27           5.0         3.4          1.6         0.4              1
#> 28           5.2         3.5          1.5         0.2              1
#> 29           5.2         3.4          1.4         0.2              1
#> 30           4.7         3.2          1.6         0.2              1
#> 31           4.8         3.1          1.6         0.2              1
#> 32           5.4         3.4          1.5         0.4              1
#> 33           5.2         4.1          1.5         0.1              1
#> 34           5.5         4.2          1.4         0.2              1
#> 35           4.9         3.1          1.5         0.2              1
#> 36           5.0         3.2          1.2         0.2              1
#> 37           5.5         3.5          1.3         0.2              1
#> 38           4.9         3.6          1.4         0.1              1
#> 39           4.4         3.0          1.3         0.2              1
#> 40           5.1         3.4          1.5         0.2              1
#> 41           5.0         3.5          1.3         0.3              1
#> 42           4.5         2.3          1.3         0.3              1
#> 43           4.4         3.2          1.3         0.2              1
#> 44           5.0         3.5          1.6         0.6              1
#> 45           5.1         3.8          1.9         0.4              1
#> 46           4.8         3.0          1.4         0.3              1
#> 47           5.1         3.8          1.6         0.2              1
#> 48           4.6         3.2          1.4         0.2              1
#> 49           5.3         3.7          1.5         0.2              1
#> 50           5.0         3.3          1.4         0.2              1
#> 51           7.0         3.2          4.7         1.4              0
#> 52           6.4         3.2          4.5         1.5              0
#> 53           6.9         3.1          4.9         1.5              0
#> 54           5.5         2.3          4.0         1.3              0
#> 55           6.5         2.8          4.6         1.5              0
#> 56           5.7         2.8          4.5         1.3              0
#> 57           6.3         3.3          4.7         1.6              0
#> 58           4.9         2.4          3.3         1.0              0
#> 59           6.6         2.9          4.6         1.3              0
#> 60           5.2         2.7          3.9         1.4              0
#> 61           5.0         2.0          3.5         1.0              0
#> 62           5.9         3.0          4.2         1.5              0
#> 63           6.0         2.2          4.0         1.0              0
#> 64           6.1         2.9          4.7         1.4              0
#> 65           5.6         2.9          3.6         1.3              0
#> 66           6.7         3.1          4.4         1.4              0
#> 67           5.6         3.0          4.5         1.5              0
#> 68           5.8         2.7          4.1         1.0              0
#> 69           6.2         2.2          4.5         1.5              0
#> 70           5.6         2.5          3.9         1.1              0
#> 71           5.9         3.2          4.8         1.8              0
#> 72           6.1         2.8          4.0         1.3              0
#> 73           6.3         2.5          4.9         1.5              0
#> 74           6.1         2.8          4.7         1.2              0
#> 75           6.4         2.9          4.3         1.3              0
#> 76           6.6         3.0          4.4         1.4              0
#> 77           6.8         2.8          4.8         1.4              0
#> 78           6.7         3.0          5.0         1.7              0
#> 79           6.0         2.9          4.5         1.5              0
#> 80           5.7         2.6          3.5         1.0              0
#> 81           5.5         2.4          3.8         1.1              0
#> 82           5.5         2.4          3.7         1.0              0
#> 83           5.8         2.7          3.9         1.2              0
#> 84           6.0         2.7          5.1         1.6              0
#> 85           5.4         3.0          4.5         1.5              0
#> 86           6.0         3.4          4.5         1.6              0
#> 87           6.7         3.1          4.7         1.5              0
#> 88           6.3         2.3          4.4         1.3              0
#> 89           5.6         3.0          4.1         1.3              0
#> 90           5.5         2.5          4.0         1.3              0
#> 91           5.5         2.6          4.4         1.2              0
#> 92           6.1         3.0          4.6         1.4              0
#> 93           5.8         2.6          4.0         1.2              0
#> 94           5.0         2.3          3.3         1.0              0
#> 95           5.6         2.7          4.2         1.3              0
#> 96           5.7         3.0          4.2         1.2              0
#> 97           5.7         2.9          4.2         1.3              0
#> 98           6.2         2.9          4.3         1.3              0
#> 99           5.1         2.5          3.0         1.1              0
#> 100          5.7         2.8          4.1         1.3              0
#> 101          6.3         3.3          6.0         2.5              0
#> 102          5.8         2.7          5.1         1.9              0
#> 103          7.1         3.0          5.9         2.1              0
#> 104          6.3         2.9          5.6         1.8              0
#> 105          6.5         3.0          5.8         2.2              0
#> 106          7.6         3.0          6.6         2.1              0
#> 107          4.9         2.5          4.5         1.7              0
#> 108          7.3         2.9          6.3         1.8              0
#> 109          6.7         2.5          5.8         1.8              0
#> 110          7.2         3.6          6.1         2.5              0
#> 111          6.5         3.2          5.1         2.0              0
#> 112          6.4         2.7          5.3         1.9              0
#> 113          6.8         3.0          5.5         2.1              0
#> 114          5.7         2.5          5.0         2.0              0
#> 115          5.8         2.8          5.1         2.4              0
#> 116          6.4         3.2          5.3         2.3              0
#> 117          6.5         3.0          5.5         1.8              0
#> 118          7.7         3.8          6.7         2.2              0
#> 119          7.7         2.6          6.9         2.3              0
#> 120          6.0         2.2          5.0         1.5              0
#> 121          6.9         3.2          5.7         2.3              0
#> 122          5.6         2.8          4.9         2.0              0
#> 123          7.7         2.8          6.7         2.0              0
#> 124          6.3         2.7          4.9         1.8              0
#> 125          6.7         3.3          5.7         2.1              0
#> 126          7.2         3.2          6.0         1.8              0
#> 127          6.2         2.8          4.8         1.8              0
#> 128          6.1         3.0          4.9         1.8              0
#> 129          6.4         2.8          5.6         2.1              0
#> 130          7.2         3.0          5.8         1.6              0
#> 131          7.4         2.8          6.1         1.9              0
#> 132          7.9         3.8          6.4         2.0              0
#> 133          6.4         2.8          5.6         2.2              0
#> 134          6.3         2.8          5.1         1.5              0
#> 135          6.1         2.6          5.6         1.4              0
#> 136          7.7         3.0          6.1         2.3              0
#> 137          6.3         3.4          5.6         2.4              0
#> 138          6.4         3.1          5.5         1.8              0
#> 139          6.0         3.0          4.8         1.8              0
#> 140          6.9         3.1          5.4         2.1              0
#> 141          6.7         3.1          5.6         2.4              0
#> 142          6.9         3.1          5.1         2.3              0
#> 143          5.8         2.7          5.1         1.9              0
#> 144          6.8         3.2          5.9         2.3              0
#> 145          6.7         3.3          5.7         2.5              0
#> 146          6.7         3.0          5.2         2.3              0
#> 147          6.3         2.5          5.0         1.9              0
#> 148          6.5         3.0          5.2         2.0              0
#> 149          6.2         3.4          5.4         2.3              0
#> 150          5.9         3.0          5.1         1.8              0
#>     Species.versicolor Species.virginica
#> 1                    0                 0
#> 2                    0                 0
#> 3                    0                 0
#> 4                    0                 0
#> 5                    0                 0
#> 6                    0                 0
#> 7                    0                 0
#> 8                    0                 0
#> 9                    0                 0
#> 10                   0                 0
#> 11                   0                 0
#> 12                   0                 0
#> 13                   0                 0
#> 14                   0                 0
#> 15                   0                 0
#> 16                   0                 0
#> 17                   0                 0
#> 18                   0                 0
#> 19                   0                 0
#> 20                   0                 0
#> 21                   0                 0
#> 22                   0                 0
#> 23                   0                 0
#> 24                   0                 0
#> 25                   0                 0
#> 26                   0                 0
#> 27                   0                 0
#> 28                   0                 0
#> 29                   0                 0
#> 30                   0                 0
#> 31                   0                 0
#> 32                   0                 0
#> 33                   0                 0
#> 34                   0                 0
#> 35                   0                 0
#> 36                   0                 0
#> 37                   0                 0
#> 38                   0                 0
#> 39                   0                 0
#> 40                   0                 0
#> 41                   0                 0
#> 42                   0                 0
#> 43                   0                 0
#> 44                   0                 0
#> 45                   0                 0
#> 46                   0                 0
#> 47                   0                 0
#> 48                   0                 0
#> 49                   0                 0
#> 50                   0                 0
#> 51                   1                 0
#> 52                   1                 0
#> 53                   1                 0
#> 54                   1                 0
#> 55                   1                 0
#> 56                   1                 0
#> 57                   1                 0
#> 58                   1                 0
#> 59                   1                 0
#> 60                   1                 0
#> 61                   1                 0
#> 62                   1                 0
#> 63                   1                 0
#> 64                   1                 0
#> 65                   1                 0
#> 66                   1                 0
#> 67                   1                 0
#> 68                   1                 0
#> 69                   1                 0
#> 70                   1                 0
#> 71                   1                 0
#> 72                   1                 0
#> 73                   1                 0
#> 74                   1                 0
#> 75                   1                 0
#> 76                   1                 0
#> 77                   1                 0
#> 78                   1                 0
#> 79                   1                 0
#> 80                   1                 0
#> 81                   1                 0
#> 82                   1                 0
#> 83                   1                 0
#> 84                   1                 0
#> 85                   1                 0
#> 86                   1                 0
#> 87                   1                 0
#> 88                   1                 0
#> 89                   1                 0
#> 90                   1                 0
#> 91                   1                 0
#> 92                   1                 0
#> 93                   1                 0
#> 94                   1                 0
#> 95                   1                 0
#> 96                   1                 0
#> 97                   1                 0
#> 98                   1                 0
#> 99                   1                 0
#> 100                  1                 0
#> 101                  0                 1
#> 102                  0                 1
#> 103                  0                 1
#> 104                  0                 1
#> 105                  0                 1
#> 106                  0                 1
#> 107                  0                 1
#> 108                  0                 1
#> 109                  0                 1
#> 110                  0                 1
#> 111                  0                 1
#> 112                  0                 1
#> 113                  0                 1
#> 114                  0                 1
#> 115                  0                 1
#> 116                  0                 1
#> 117                  0                 1
#> 118                  0                 1
#> 119                  0                 1
#> 120                  0                 1
#> 121                  0                 1
#> 122                  0                 1
#> 123                  0                 1
#> 124                  0                 1
#> 125                  0                 1
#> 126                  0                 1
#> 127                  0                 1
#> 128                  0                 1
#> 129                  0                 1
#> 130                  0                 1
#> 131                  0                 1
#> 132                  0                 1
#> 133                  0                 1
#> 134                  0                 1
#> 135                  0                 1
#> 136                  0                 1
#> 137                  0                 1
#> 138                  0                 1
#> 139                  0                 1
#> 140                  0                 1
#> 141                  0                 1
#> 142                  0                 1
#> 143                  0                 1
#> 144                  0                 1
#> 145                  0                 1
#> 146                  0                 1
#> 147                  0                 1
#> 148                  0                 1
#> 149                  0                 1
#> 150                  0                 1