Compute group-meaned and de-meaned variables

demean() computes group- and de-meaned versions of a variable that can be used in regression analysis to model the between- and within-subject effect.

demean(x, ..., cluster_id, append = TRUE, suffix_dm = "_dm",
  suffix_gm = "_gm")

Arguments

x
A data frame.

...
Names of variables that should be group- and de-meaned.

cluster_id
Quoted or unquoted name of the variable that indicates the group- or cluster-ID.

append
Logical, if TRUE, new created variables including x are returned, else only the new created variables are returned.

suffix_dm, suffix_gm
String value, will be appended to the names of the group-meaned and de-meaned variables of x. By default, de-meaned variables will be suffixed with "_dm" and grouped-meaned variables with "_gm".

Value

For append = TRUE, x including the group-/de-meaned variables as new columns is returned; if append = FALSE, only the group-/de-meaned variables will be returned.

Details

demean() is intended to create group- and de-meaned variables for complex random-effect-within-between models (see Bell et al. 2018), where group-effects (random effects) and fixed effects correlate (see Bafumi and Gelman 2006)). This violation of one of the Gauss-Markov-assumptions can happen, for instance, when analysing panel data. To control for correlating predictors and group effects, it is recommended to include the group-meaned and de-meaned version of time-varying covariates in the model. By this, one can fit complex multilevel models for panel data, including time-varying predictors, time-invariant predictors and random effects. This approach is superior to classic fixed-effects models, which lack information of variation in the group-effects or between-subject effects.

A description of how to translate the formulas described in Bell et al. 2018 into R using lmer() from lme4 or glmmTMB() from glmmTMB can be found here: for lmer() and for glmmTMB().

References

Bafumi J, Gelman A. 2006. Fitting Multilevel Models When Predictors and Group Effects Correlate. In. Philadelphia, PA: Annual meeting of the American Political Science Association.

Bell A, Fairbrother M, Jones K. 2018. Fixed and Random Effects Models: Making an Informed Choice. Quality & Quantity. doi: 10.1007/s11135-018-0802-x

Examples

data(iris)
iris$ID <- sample(1:4, nrow(iris), replace = TRUE) # fake-ID
demean(iris, Sepal.Length, Petal.Length, cluster_id = ID, append = FALSE)

#>     Sepal.Length_gm Petal.Length_gm Sepal.Length_dm Petal.Length_dm
#> 1          5.848718        3.710256    -0.748717949    -2.310256410
#> 2          5.902439        4.009756    -1.002439024    -2.609756098
#> 3          5.848718        3.710256    -1.148717949    -2.410256410
#> 4          5.848718        3.710256    -1.248717949    -2.210256410
#> 5          5.848718        3.710256    -0.848717949    -2.310256410
#> 6          5.848718        3.710256    -0.448717949    -2.010256410
#> 7          6.002703        3.937838    -1.402702703    -2.537837838
#> 8          5.584848        3.300000    -0.584848485    -1.800000000
#> 9          6.002703        3.937838    -1.602702703    -2.537837838
#> 10         5.584848        3.300000    -0.684848485    -1.800000000
#> 11         5.902439        4.009756    -0.502439024    -2.509756098
#> 12         5.584848        3.300000    -0.784848485    -1.700000000
#> 13         6.002703        3.937838    -1.202702703    -2.537837838
#> 14         5.584848        3.300000    -1.284848485    -2.200000000
#> 15         6.002703        3.937838    -0.202702703    -2.737837838
#> 16         6.002703        3.937838    -0.302702703    -2.437837838
#> 17         5.902439        4.009756    -0.502439024    -2.709756098
#> 18         6.002703        3.937838    -0.902702703    -2.537837838
#> 19         6.002703        3.937838    -0.302702703    -2.237837838
#> 20         5.848718        3.710256    -0.748717949    -2.210256410
#> 21         5.848718        3.710256    -0.448717949    -2.010256410
#> 22         5.902439        4.009756    -0.802439024    -2.509756098
#> 23         6.002703        3.937838    -1.402702703    -2.937837838
#> 24         5.584848        3.300000    -0.484848485    -1.600000000
#> 25         5.848718        3.710256    -1.048717949    -1.810256410
#> 26         5.848718        3.710256    -0.848717949    -2.110256410
#> 27         5.902439        4.009756    -0.902439024    -2.409756098
#> 28         5.902439        4.009756    -0.702439024    -2.509756098
#> 29         6.002703        3.937838    -0.802702703    -2.537837838
#> 30         5.584848        3.300000    -0.884848485    -1.700000000
#> 31         6.002703        3.937838    -1.202702703    -2.337837838
#> 32         5.584848        3.300000    -0.184848485    -1.800000000
#> 33         5.848718        3.710256    -0.648717949    -2.210256410
#> 34         5.848718        3.710256    -0.348717949    -2.310256410
#> 35         5.584848        3.300000    -0.684848485    -1.800000000
#> 36         5.848718        3.710256    -0.848717949    -2.510256410
#> 37         6.002703        3.937838    -0.502702703    -2.637837838
#> 38         5.902439        4.009756    -1.002439024    -2.609756098
#> 39         5.848718        3.710256    -1.448717949    -2.410256410
#> 40         5.584848        3.300000    -0.484848485    -1.800000000
#> 41         5.902439        4.009756    -0.902439024    -2.709756098
#> 42         5.584848        3.300000    -1.084848485    -2.000000000
#> 43         5.584848        3.300000    -1.184848485    -2.000000000
#> 44         5.902439        4.009756    -0.902439024    -2.409756098
#> 45         5.584848        3.300000    -0.484848485    -1.400000000
#> 46         6.002703        3.937838    -1.202702703    -2.537837838
#> 47         5.902439        4.009756    -0.802439024    -2.409756098
#> 48         5.584848        3.300000    -0.984848485    -1.900000000
#> 49         5.584848        3.300000    -0.284848485    -1.800000000
#> 50         5.584848        3.300000    -0.584848485    -1.900000000
#> 51         5.902439        4.009756     1.097560976     0.690243902
#> 52         5.902439        4.009756     0.497560976     0.490243902
#> 53         5.902439        4.009756     0.997560976     0.890243902
#> 54         5.848718        3.710256    -0.348717949     0.289743590
#> 55         6.002703        3.937838     0.497297297     0.662162162
#> 56         5.848718        3.710256    -0.148717949     0.789743590
#> 57         5.902439        4.009756     0.397560976     0.690243902
#> 58         5.848718        3.710256    -0.948717949    -0.410256410
#> 59         5.902439        4.009756     0.697560976     0.590243902
#> 60         5.848718        3.710256    -0.648717949     0.189743590
#> 61         6.002703        3.937838    -1.002702703    -0.437837838
#> 62         5.848718        3.710256     0.051282051     0.489743590
#> 63         6.002703        3.937838    -0.002702703     0.062162162
#> 64         6.002703        3.937838     0.097297297     0.762162162
#> 65         5.848718        3.710256    -0.248717949    -0.110256410
#> 66         5.848718        3.710256     0.851282051     0.689743590
#> 67         6.002703        3.937838    -0.402702703     0.562162162
#> 68         5.902439        4.009756    -0.102439024     0.090243902
#> 69         5.584848        3.300000     0.615151515     1.200000000
#> 70         5.848718        3.710256    -0.248717949     0.189743590
#> 71         5.902439        4.009756    -0.002439024     0.790243902
#> 72         5.584848        3.300000     0.515151515     0.700000000
#> 73         6.002703        3.937838     0.297297297     0.962162162
#> 74         5.848718        3.710256     0.251282051     0.989743590
#> 75         5.584848        3.300000     0.815151515     1.000000000
#> 76         5.584848        3.300000     1.015151515     1.100000000
#> 77         6.002703        3.937838     0.797297297     0.862162162
#> 78         5.848718        3.710256     0.851282051     1.289743590
#> 79         5.848718        3.710256     0.151282051     0.789743590
#> 80         5.584848        3.300000     0.115151515     0.200000000
#> 81         5.848718        3.710256    -0.348717949     0.089743590
#> 82         5.902439        4.009756    -0.402439024    -0.309756098
#> 83         5.584848        3.300000     0.215151515     0.600000000
#> 84         6.002703        3.937838    -0.002702703     1.162162162
#> 85         5.902439        4.009756    -0.502439024     0.490243902
#> 86         6.002703        3.937838    -0.002702703     0.562162162
#> 87         5.848718        3.710256     0.851282051     0.989743590
#> 88         5.848718        3.710256     0.451282051     0.689743590
#> 89         6.002703        3.937838    -0.402702703     0.162162162
#> 90         5.902439        4.009756    -0.402439024    -0.009756098
#> 91         5.848718        3.710256    -0.348717949     0.689743590
#> 92         5.584848        3.300000     0.515151515     1.300000000
#> 93         5.902439        4.009756    -0.102439024    -0.009756098
#> 94         5.902439        4.009756    -0.902439024    -0.709756098
#> 95         5.584848        3.300000     0.015151515     0.900000000
#> 96         5.902439        4.009756    -0.202439024     0.190243902
#> 97         5.584848        3.300000     0.115151515     0.900000000
#> 98         5.902439        4.009756     0.297560976     0.290243902
#> 99         5.902439        4.009756    -0.802439024    -1.009756098
#> 100        5.902439        4.009756    -0.202439024     0.090243902
#> 101        5.902439        4.009756     0.397560976     1.990243902
#> 102        5.902439        4.009756    -0.102439024     1.090243902
#> 103        5.584848        3.300000     1.515151515     2.600000000
#> 104        5.584848        3.300000     0.715151515     2.300000000
#> 105        5.584848        3.300000     0.915151515     2.500000000
#> 106        5.848718        3.710256     1.751282051     2.889743590
#> 107        5.902439        4.009756    -1.002439024     0.490243902
#> 108        6.002703        3.937838     1.297297297     2.362162162
#> 109        6.002703        3.937838     0.697297297     1.862162162
#> 110        5.848718        3.710256     1.351282051     2.389743590
#> 111        5.848718        3.710256     0.651282051     1.389743590
#> 112        5.902439        4.009756     0.497560976     1.290243902
#> 113        5.584848        3.300000     1.215151515     2.200000000
#> 114        5.584848        3.300000     0.115151515     1.700000000
#> 115        6.002703        3.937838    -0.202702703     1.162162162
#> 116        6.002703        3.937838     0.397297297     1.362162162
#> 117        6.002703        3.937838     0.497297297     1.562162162
#> 118        5.902439        4.009756     1.797560976     2.690243902
#> 119        5.848718        3.710256     1.851282051     3.189743590
#> 120        5.584848        3.300000     0.415151515     1.700000000
#> 121        5.902439        4.009756     0.997560976     1.690243902
#> 122        5.902439        4.009756    -0.302439024     0.890243902
#> 123        6.002703        3.937838     1.697297297     2.762162162
#> 124        6.002703        3.937838     0.297297297     0.962162162
#> 125        5.848718        3.710256     0.851282051     1.989743590
#> 126        6.002703        3.937838     1.197297297     2.062162162
#> 127        5.902439        4.009756     0.297560976     0.790243902
#> 128        5.848718        3.710256     0.251282051     1.189743590
#> 129        5.902439        4.009756     0.497560976     1.590243902
#> 130        5.902439        4.009756     1.297560976     1.790243902
#> 131        6.002703        3.937838     1.397297297     2.162162162
#> 132        6.002703        3.937838     1.897297297     2.462162162
#> 133        5.584848        3.300000     0.815151515     2.300000000
#> 134        6.002703        3.937838     0.297297297     1.162162162
#> 135        6.002703        3.937838     0.097297297     1.662162162
#> 136        5.902439        4.009756     1.797560976     2.090243902
#> 137        5.902439        4.009756     0.397560976     1.590243902
#> 138        5.584848        3.300000     0.815151515     2.200000000
#> 139        6.002703        3.937838    -0.002702703     0.862162162
#> 140        5.848718        3.710256     1.051282051     1.689743590
#> 141        6.002703        3.937838     0.697297297     1.662162162
#> 142        6.002703        3.937838     0.897297297     1.162162162
#> 143        5.584848        3.300000     0.215151515     1.800000000
#> 144        5.848718        3.710256     0.951282051     2.189743590
#> 145        5.902439        4.009756     0.797560976     1.690243902
#> 146        5.848718        3.710256     0.851282051     1.489743590
#> 147        5.848718        3.710256     0.451282051     1.289743590
#> 148        5.848718        3.710256     0.651282051     1.489743590
#> 149        5.902439        4.009756     0.297560976     1.390243902
#> 150        5.902439        4.009756    -0.002439024     1.090243902