ANOVA Effect Size (Omega Squared, Eta Squared, Epsilon Squared)

Set of functions to compute (partial) indices for ANOVAs, such as omega squared, the eta squared or the epsilon squared (Kelly, 1935). They are measures of effect size, or the degree of association for a population. They are an estimate of how much variance in the response variables are accounted for by the explanatory variables.

cohens_f(model)

epsilon_squared(model)

eta_squared(model, partial = TRUE, ci = NULL, ...)

# S3 method for aov
eta_squared(model, partial = TRUE, ci = NULL,
  iterations = 1000, ...)

omega_squared(model, partial = TRUE, ci = NULL, iterations = 1000)

Arguments

model
An ANOVA object.

partial
If TRUE, return partial indices (adjusted for sample size).

ci
Confidence Interval (CI) level computed via boostrap.

...
Arguments passed to or from other methods.

iterations
The number of bootstrap replicates.

Value

Data.frame containig the effect size values.

Details

Omega Squared
Omega squared is considered as a lesser biased alternative to eta-squared, especially when sample sizes are small (Albers \& Lakens, 2018). Field (2013) suggests the following interpretation heuristics:

Omega Squared = 0 - 0.01: Very small

Omega Squared = 0.01 - 0.06: Small

Omega Squared = 0.06 - 0.14: Medium

Omega Squared > 0.14: Large

Epsilon Squared
It is one of the least common measures of effect sizes: omega squared and eta squared are used more frequently. Altough having a different name and a formula in appearance different, this index is equivalent to the adjusted R2 (Allen, 2017, p. 382).

Cohen's f
Cohen's f statistic is one appropriate effect size index to use for a oneway analysis of variance (ANOVA). Cohen's f can take on values between zero, when the population means are all equal, and an indefinitely large number as standard deviation of means increases relative to the average standard deviation within each group. Cohen has suggested that the values of 0.10, 0.25, and 0.40 represent small, medium, and large effect sizes, respectively.
References

Albers, C., \& Lakens, D. (2018). When power analyses based on pilot data are biased: Inaccurate effect size estimators and follow-up bias. Journal of experimental social psychology, 74, 187-195.

Allen, R. (2017). Statistics and Experimental Design for Psychologists: A Model Comparison Approach. World Scientific Publishing Company.

Field, A. (2013). Discovering statistics using IBM SPSS statistics. sage.

Kelley, K. (2007). Methods for the behavioral, educational, and social sciences: An R package. Behavior Research Methods, 39(4), 979-984.

Kelley, T. (1935) An unbiased correlation ratio measure. Proceedings of the National Academy of Sciences. 21(9). 554-559.

The computation of CIs is based on the implementation done by Stanley (2018) in the ApaTables package and Kelley (2007) in the MBESS package. All credits go to them.

Examples

library(parameters)

df <- iris
df$Sepal.Big <- ifelse(df$Sepal.Width >= 3, "Yes", "No")

model <- aov(Sepal.Length ~ Sepal.Big, data = df)
omega_squared(model)

#>   Parameter Omega_Sq_partial
#> 1 Sepal.Big      0.004038505
#> 2 Residuals               NA

eta_squared(model)

#>   Parameter Eta_Sq_partial
#> 1 Sepal.Big     0.01074962
#> 2 Residuals             NA

epsilon_squared(model)

#>   Parameter  Epsilon_sq
#> 1 Sepal.Big 0.004065499
#> 2 Residuals          NA

cohens_f(model)

#>   Parameter  Cohens_f
#> 1 Sepal.Big 0.1042422
#> 2 Residuals        NA

model <- anova(lm(Sepal.Length ~ Sepal.Big * Species, data = df))
omega_squared(model)

#>           Parameter Omega_Sq_partial
#> 1         Sepal.Big       0.02361791
#> 2           Species       0.65095815
#> 3 Sepal.Big:Species      -0.01224219
#> 4         Residuals               NA

eta_squared(model)

#>           Parameter Eta_Sq_partial
#> 1         Sepal.Big    0.031140626
#> 2           Species    0.661771666
#> 3 Sepal.Big:Species    0.001289171
#> 4         Residuals             NA

epsilon_squared(model)

#>           Parameter   Epsilon_sq
#> 1         Sepal.Big  0.008427078
#> 2           Species  0.649727083
#> 3 Sepal.Big:Species -0.004213375
#> 4         Residuals           NA

# Don't work for now
model <- aov(Sepal.Length ~ Sepal.Big + Error(Species), data = df)
omega_squared(model)

#> Error in omega_squared.aovlist(model): Omega squared not implemented yet for repeated-measures ANOVAs.

eta_squared(model)

#>   Parameter Eta_Sq_partial
#> 1 Sepal.Big      0.4472423
#> 2 Residuals             NA
#> 3 Sepal.Big      0.1217329
#> 4 Residuals             NA

epsilon_squared(model)

#> Error in epsilon_squared.aovlist(model): Epsilon squared not implemented yet for repeated-measures ANOVAs.