Parameter names formatting

Parameter names formatting

format_parameters(model)

Arguments

model
A statistical model.

Value

The formatted parameter names.

Notation of Interaction Terms

The notation of interaction terms in format_parameters(), and hence also in print() for model_parameters(), uses both * and :, depending on the type of interaction. Whenever main effects for all parameters of the interaction terms are present, the interaction is denoted as *. When at least one of the interaction terms is not included as main effect, the interaction is denoted as :.

m1 <- lm(mpg ~ qsec : wt + wt / drat, data = mtcars)
model_parameters(m1)
> Parameter   | Coefficient |   SE |          95% CI | t(28) |      p
> -------------------------------------------------------------------
> (Intercept) |       35.89 | 2.34 | [ 31.09, 40.68] | 15.34 | < .001
> wt          |      -10.47 | 1.96 | [-14.48, -6.46] | -5.35 | < .001
> qsec : wt   |        0.28 | 0.09 | [  0.10,  0.46] |  3.17 | 0.004
> wt : drat   |        0.17 | 0.43 | [ -0.71,  1.05] |  0.39 | 0.701

m2 <- lm(mpg ~ qsec * wt, data = mtcars)
model_parameters(m2)
> Parameter   | Coefficient |    SE |          95% CI | t(28) |     p
> -------------------------------------------------------------------
> (Intercept) |      -13.11 | 29.38 | [-73.29, 47.06] | -0.45 | 0.659
> qsec        |        2.76 |  1.63 | [ -0.58,  6.09] |  1.69 | 0.102
> wt          |        5.50 |  9.29 | [-13.53, 24.53] |  0.59 | 0.559
> qsec * wt   |       -0.59 |  0.52 | [ -1.65,  0.47] | -1.14 | 0.265

Interpretation of Interaction Terms

Note that the interpretation of model coefficients depends on many characteristics of the model. The number of parameters, and overall performance of the model, can differ or not between a * b and a / b, suggesting that sometimes interaction terms give different parameterizations of the same model, but other times it gives completely different models (depending on a or b being factors of covariates, included as main effects or not, etc.). Their interpretation depends of the full context of the model, which should not be inferred from the parameters table alone - rather, we recommend to use packages to calculate estimated marginal means or marginal effects, such as modelbased, emmeans or ggeffects.

Examples

library(parameters)

model <- lm(Sepal.Length ~ Species * Sepal.Width, data = iris)
format_parameters(model)

#>                          (Intercept)                    Speciesversicolor 
#>                        "(Intercept)"               "Species [versicolor]" 
#>                     Speciesvirginica                          Sepal.Width 
#>                "Species [virginica]"                        "Sepal.Width" 
#>        Speciesversicolor:Sepal.Width         Speciesvirginica:Sepal.Width 
#> "Species [versicolor] * Sepal.Width"  "Species [virginica] * Sepal.Width" 

model <- lm(Sepal.Length ~ Species / Petal.Length, data = iris)
format_parameters(model)

#>                           (Intercept)                     Speciesversicolor 
#>                         "(Intercept)"                "Species [versicolor]" 
#>                      Speciesvirginica            Speciessetosa:Petal.Length 
#>                 "Species [virginica]"     "Species [setosa] : Petal.Length" 
#>        Speciesversicolor:Petal.Length         Speciesvirginica:Petal.Length 
#> "Species [versicolor] : Petal.Length"  "Species [virginica] : Petal.Length" 

model <- lm(Sepal.Length ~ Petal.Length + (Species / Sepal.Width), data = iris)
format_parameters(model)

#>                          (Intercept)                         Petal.Length 
#>                        "(Intercept)"                       "Petal.Length" 
#>                    Speciesversicolor                     Speciesvirginica 
#>               "Species [versicolor]"                "Species [virginica]" 
#>            Speciessetosa:Sepal.Width        Speciesversicolor:Sepal.Width 
#>     "Species [setosa] : Sepal.Width" "Species [versicolor] : Sepal.Width" 
#>         Speciesvirginica:Sepal.Width 
#>  "Species [virginica] : Sepal.Width" 

model <- lm(Sepal.Length ~ Species + poly(Sepal.Width, 2), data = iris)
format_parameters(model)

#>                (Intercept)          Speciesversicolor 
#>              "(Intercept)"     "Species [versicolor]" 
#>           Speciesvirginica      poly(Sepal.Width, 2)1 
#>      "Species [virginica]" "Sepal.Width [1st degree]" 
#>      poly(Sepal.Width, 2)2 
#> "Sepal.Width [2nd degree]" 

model <- lm(Sepal.Length ~ Species + poly(Sepal.Width, 2, raw = TRUE), data = iris)
format_parameters(model)

#>                       (Intercept)                 Speciesversicolor 
#>                     "(Intercept)"            "Species [versicolor]" 
#>                  Speciesvirginica poly(Sepal.Width, 2, raw = TRUE)1 
#>             "Species [virginica]"        "Sepal.Width [1st degree]" 
#> poly(Sepal.Width, 2, raw = TRUE)2 
#>        "Sepal.Width [2nd degree]"