Transform a standardized vector into character

Transform a standardized vector into character, e.g., c("-1 SD", "Mean", "+1 SD").

format_standardize(x, reference = x, robust = FALSE, digits = NULL,
  ...)

Arguments

x
A standardized numeric vector.

reference
The reference vector from which to compute the mean and SD.

robust
Logical, if TRUE, centering is done by substracting the median from the variables and divide it by the median absolute deviation (MAD). If FALSE, variables are standardized by substracting the mean and divide it by the standard deviation (SD). the

digits
Number of significant digits.

...
Arguments passed to or from other methods.

Examples

format_standardize(c(-1, 0, 1))

#> [1] -1 SD Mean  +1 SD
#> Levels: +1 SD Mean -1 SD

format_standardize(c(-1, 0, 1, 2), reference = rnorm(1000))

#> [1] -1 SD Mean  +1 SD +2 SD
#> Levels: +2 SD +1 SD Mean -1 SD

format_standardize(c(-1, 0, 1, 2), reference = rnorm(1000), robust = TRUE)

#> [1] -1 MAD Median +1 MAD +2 MAD
#> Levels: +2 MAD +1 MAD Median -1 MAD