get_scores {parameters}	R Documentation

Get Scores from Principal Component or Factor Analysis (PCA/FA)

Description

get_scores() takes n_items amount of items that load the most (either by loading cutoff or number) on a component, and then computes their average. This results in a sum score for each component from the PCA/FA, which is on the same scale as the original, single items that were used to compute the PCA/FA.

Usage

get_scores(x, n_items = NULL)

Arguments

x
An object returned by principal_components() or factor_analysis().

n_items
Number of required (i.e. non-missing) items to build the sum score for an observation. If an observation has more missing values than n_items in all items of a (sub) scale, NA is returned for that observation, else, the sum score of all (sub) items is calculated. If NULL, the value is chosen to match half of the number of columns in a data frame, i.e. no more than 50% missing values are allowed.

Details

get_scores() takes the results from principal_components() or factor_analysis() and extracts the variables for each component found by the PCA/FA. Then, for each of these "subscales", row means are calculated (which equals adding up the single items and dividing by the number of items). This results in a sum score for each component from the PCA/FA, which is on the same scale as the original, single items that were used to compute the PCA/FA.

Value

A data frame with subscales, which are average sum scores for all items from each component or factor.

See Also

Functions to carry out a PCA (principal_components()) or a FA (factor_analysis()). factor_scores() extracts factor scores from an FA object.

Examples

pca <- principal_components(mtcars[, 1:7], n = 2, rotation = "varimax")

# PCA extracted two components
pca

# assignment of items to each component
closest_component(pca)

# now we want to have sum scores for each component
get_scores(pca)

# compare to manually computed sum score for 2nd component, which
# consists of items "hp" and "qsec"
(mtcars$hp + mtcars$qsec) / 2

[Package parameters version 0.28.3 Index]