Parameters from special models

Parameters from special regression models not listed under one of the previous categories yet.

Parameters from Hypothesis Testing.

Usage

# S3 method for PMCMR
model_parameters(model, ...)

# S3 method for glimML
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("conditional", "random", "dispersion", "all"),
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for averaging
model_parameters(
  model,
  ci = 0.95,
  component = c("conditional", "full"),
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for mle2
model_parameters(
  model,
  ci = 0.95,
  ci_method = NULL,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  summary = getOption("parameters_summary", FALSE),
  df_method = ci_method,
  vcov = NULL,
  vcov_args = NULL,
  verbose = TRUE,
  ...
)

# S3 method for betareg
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("conditional", "precision", "all"),
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for bfsl
model_parameters(
  model,
  ci = 0.95,
  ci_method = "residual",
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for deltaMethod
model_parameters(model, p_adjust = NULL, verbose = TRUE, ...)

# S3 method for emmGrid
model_parameters(
  model,
  ci = 0.95,
  centrality = "median",
  dispersion = FALSE,
  ci_method = "eti",
  test = c("pd", "rope"),
  rope_range = "default",
  rope_ci = 0.95,
  exponentiate = FALSE,
  p_adjust = NULL,
  parameters = NULL,
  verbose = TRUE,
  ...
)

# S3 method for emm_list
model_parameters(
  model,
  ci = 0.95,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for epi.2by2
model_parameters(model, verbose = TRUE, ...)

# S3 method for fitdistr
model_parameters(model, exponentiate = FALSE, verbose = TRUE, ...)

# S3 method for ggeffects
model_parameters(model, parameters = NULL, verbose = TRUE, ...)

# S3 method for SemiParBIV
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for glmm
model_parameters(
  model,
  ci = 0.95,
  effects = c("all", "fixed", "random"),
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  verbose = TRUE,
  ...
)

# S3 method for glmx
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "extra"),
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for ivFixed
model_parameters(model, ci = 0.95, ci_method = "wald", verbose = TRUE, ...)

# S3 method for ivprobit
model_parameters(model, ci = 0.95, ci_method = "wald", verbose = TRUE, ...)

# S3 method for lmodel2
model_parameters(
  model,
  ci = 0.95,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for logistf
model_parameters(
  model,
  ci = 0.95,
  ci_method = NULL,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  summary = getOption("parameters_summary", FALSE),
  df_method = ci_method,
  vcov = NULL,
  vcov_args = NULL,
  verbose = TRUE,
  ...
)

# S3 method for lqmm
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for marginaleffects
model_parameters(model, ci = 0.95, ...)

# S3 method for margins
model_parameters(
  model,
  ci = 0.95,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for maxLik
model_parameters(
  model,
  ci = 0.95,
  ci_method = NULL,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  summary = getOption("parameters_summary", FALSE),
  keep = NULL,
  drop = NULL,
  parameters = keep,
  verbose = TRUE,
  vcov = NULL,
  vcov_args = NULL,
  ...
)

# S3 method for maxim
model_parameters(
  model,
  ci = 0.95,
  ci_method = NULL,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  summary = getOption("parameters_summary", FALSE),
  keep = NULL,
  drop = NULL,
  parameters = keep,
  verbose = TRUE,
  vcov = NULL,
  vcov_args = NULL,
  ...
)

# S3 method for mediate
model_parameters(model, ci = 0.95, exponentiate = FALSE, verbose = TRUE, ...)

# S3 method for metaplus
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  include_studies = TRUE,
  verbose = TRUE,
  ...
)

# S3 method for meta_random
model_parameters(
  model,
  ci = 0.95,
  ci_method = "eti",
  exponentiate = FALSE,
  include_studies = TRUE,
  verbose = TRUE,
  ...
)

# S3 method for meta_fixed
model_parameters(
  model,
  ci = 0.95,
  ci_method = "eti",
  exponentiate = FALSE,
  include_studies = TRUE,
  verbose = TRUE,
  ...
)

# S3 method for meta_bma
model_parameters(
  model,
  ci = 0.95,
  ci_method = "eti",
  exponentiate = FALSE,
  include_studies = TRUE,
  verbose = TRUE,
  ...
)

# S3 method for logitor
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = TRUE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for poissonirr
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = TRUE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for negbinirr
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = TRUE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for poissonmfx
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "marginal"),
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for logitmfx
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "marginal"),
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for probitmfx
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "marginal"),
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for negbinmfx
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "marginal"),
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for betaor
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("conditional", "precision", "all"),
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for betamfx
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "precision", "marginal"),
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for mjoint
model_parameters(
  model,
  ci = 0.95,
  effects = "fixed",
  component = c("all", "conditional", "survival"),
  exponentiate = FALSE,
  p_adjust = NULL,
  keep = NULL,
  drop = NULL,
  parameters = keep,
  verbose = TRUE,
  ...
)

# S3 method for model_fit
model_parameters(
  model,
  ci = 0.95,
  effects = "fixed",
  component = "conditional",
  ci_method = "profile",
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for glht
model_parameters(model, ci = 0.95, exponentiate = FALSE, verbose = TRUE, ...)

# S3 method for mvord
model_parameters(
  model,
  ci = 0.95,
  component = c("all", "conditional", "thresholds", "correlation"),
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for pgmm
model_parameters(
  model,
  ci = 0.95,
  component = c("conditional", "all"),
  exponentiate = FALSE,
  p_adjust = NULL,
  keep = NULL,
  drop = NULL,
  parameters = keep,
  verbose = TRUE,
  ...
)

# S3 method for rqss
model_parameters(
  model,
  ci = 0.95,
  ci_method = "residual",
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  keep = NULL,
  drop = NULL,
  parameters = keep,
  verbose = TRUE,
  ...
)

# S3 method for rqs
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  verbose = TRUE,
  ...
)

# S3 method for selection
model_parameters(
  model,
  ci = 0.95,
  component = c("all", "selection", "outcome", "auxiliary"),
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for mle
model_parameters(
  model,
  ci = 0.95,
  ci_method = NULL,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  summary = getOption("parameters_summary", FALSE),
  df_method = ci_method,
  vcov = NULL,
  vcov_args = NULL,
  verbose = TRUE,
  ...
)

# S3 method for systemfit
model_parameters(
  model,
  ci = 0.95,
  ci_method = NULL,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  summary = FALSE,
  verbose = TRUE,
  ...
)

# S3 method for varest
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for t1way
model_parameters(model, keep = NULL, verbose = TRUE, ...)

# S3 method for med1way
model_parameters(model, verbose = TRUE, ...)

# S3 method for dep.effect
model_parameters(model, keep = NULL, verbose = TRUE, ...)

# S3 method for yuen
model_parameters(model, verbose = TRUE, ...)

Arguments

Value

A data frame of indices related to the model's parameters.

A data frame of indices related to the model's parameters.

A data frame of indices related to the model's parameters.

See also

Examples

library(parameters)
if (require("brglm2", quietly = TRUE)) {
  data("stemcell")
  model <- bracl(
    research ~ as.numeric(religion) + gender,
    weights = frequency,
    data = stemcell,
    type = "ML"
  )
  model_parameters(model)
}
#> # Response level: definitely
#> 
#> Parameter       | Log-Odds |   SE |         95% CI |     z |      p
#> -------------------------------------------------------------------
#> (Intercept)     |    -1.25 | 0.26 | [-1.76, -0.73] | -4.76 | < .001
#> religion        |     0.44 | 0.10 | [ 0.23,  0.64] |  4.20 | < .001
#> gender [female] |    -0.14 | 0.17 | [-0.47,  0.19] | -0.82 | 0.414 
#> 
#> # Response level: probably
#> 
#> Parameter       | Log-Odds |   SE |        95% CI |    z |     p
#> ----------------------------------------------------------------
#> (Intercept)     |     0.47 | 0.29 | [-0.10, 1.04] | 1.62 | 0.105
#> religion        |     0.26 | 0.13 | [ 0.01, 0.51] | 2.01 | 0.044
#> gender [female] |     0.19 | 0.21 | [-0.22, 0.60] | 0.90 | 0.370
#> 
#> # Response level: probably not
#> 
#> Parameter       | Log-Odds |   SE |        95% CI |     z |     p
#> -----------------------------------------------------------------
#> (Intercept)     |     0.43 | 0.39 | [-0.33, 1.18] |  1.11 | 0.268
#> religion        |     0.01 | 0.17 | [-0.33, 0.35] |  0.07 | 0.945
#> gender [female] |    -0.16 | 0.28 | [-0.71, 0.39] | -0.57 | 0.566
#> 
#> Uncertainty intervals (equal-tailed) and p-values (two-tailed) computed
#>   using a Wald z-distribution approximation.
# \donttest{
if (require("multcomp", quietly = TRUE)) {
  # multiple linear model, swiss data
  lmod <- lm(Fertility ~ ., data = swiss)
  mod <- glht(
    model = lmod,
    linfct = c(
      "Agriculture = 0",
      "Examination = 0",
      "Education = 0",
      "Catholic = 0",
      "Infant.Mortality = 0"
    )
  )
  model_parameters(mod)
}
#> 
#> Attaching package: ‘mvtnorm’
#> The following object is masked from ‘package:mclust’:
#> 
#>     dmvnorm
#> 
#> Attaching package: ‘TH.data’
#> The following object is masked from ‘package:MASS’:
#> 
#>     geyser
#> # Fixed Effects
#> 
#> Parameter             | Coefficient |   SE |         95% CI | t(41) |      p
#> ----------------------------------------------------------------------------
#> Agriculture == 0      |       -0.17 | 0.07 | [-0.36,  0.01] | -2.45 | 0.080 
#> Examination == 0      |       -0.26 | 0.25 | [-0.93,  0.41] | -1.02 | 0.785 
#> Education == 0        |       -0.87 | 0.18 | [-1.35, -0.39] | -4.76 | < .001
#> Catholic == 0         |        0.10 | 0.04 | [ 0.01,  0.20] |  2.95 | 0.023 
#> Infant Mortality == 0 |        1.08 | 0.38 | [ 0.07,  2.09] |  2.82 | 0.033 
#> 
#> p-value adjustment method: single-step
#> 
#> 
#> Uncertainty intervals (equal-tailed) and p-values (two-tailed) computed
#>   using a Wald t-distribution approximation.
if (require("PMCMRplus", quietly = TRUE)) {
  model <- kwAllPairsConoverTest(count ~ spray, data = InsectSprays)
  model_parameters(model)
}
#> Warning: Ties are present. Quantiles were corrected for ties.
#> Conover's all-pairs test
#> 
#> Group1 | Group2 | Statistic |      p | alternative | Distribution | p_adjustment
#> --------------------------------------------------------------------------------
#> B      |      A |      0.89 | 0.988  |   two.sided |            q |  single-step
#> C      |      A |    -13.58 | < .001 |   two.sided |            q |  single-step
#> C      |      B |    -14.47 | < .001 |   two.sided |            q |  single-step
#> D      |      A |     -8.87 | < .001 |   two.sided |            q |  single-step
#> D      |      B |     -9.76 | < .001 |   two.sided |            q |  single-step
#> D      |      C |      4.71 | 0.017  |   two.sided |            q |  single-step
#> E      |      A |    -10.95 | < .001 |   two.sided |            q |  single-step
#> E      |      B |    -11.84 | < .001 |   two.sided |            q |  single-step
#> E      |      C |      2.63 | 0.437  |   two.sided |            q |  single-step
#> E      |      D |     -2.09 | 0.681  |   two.sided |            q |  single-step
#> F      |      A |      1.15 | 0.964  |   two.sided |            q |  single-step
#> F      |      B |      0.26 | > .999 |   two.sided |            q |  single-step
#> F      |      C |     14.74 | < .001 |   two.sided |            q |  single-step
#> F      |      D |     10.02 | < .001 |   two.sided |            q |  single-step
#> F      |      E |     12.11 | < .001 |   two.sided |            q |  single-step
# }
if (require("WRS2") && packageVersion("WRS2") >= "1.1.3") {
  model <- t1way(libido ~ dose, data = viagra)
  model_parameters(model)
}
#> Loading required package: WRS2
#> A heteroscedastic one-way ANOVA for trimmed means
#> 
#> F    | df | df (error) |     p | Estimate |       95% CI |                         Effectsize
#> ---------------------------------------------------------------------------------------------
#> 3.00 |  2 |          4 | 0.160 |     0.79 | [0.41, 1.43] | Explanatory measure of effect size