Parameters from Generalized Additive (Mixed) Models

Extract and compute indices and measures to describe parameters of generalized additive models (GAM(M)s).

# S3 method for gam
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  ...
)

# S3 method for rqss
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("conditional", "smooth_terms", "all"),
  standardize = NULL,
  exponentiate = FALSE,
  ...
)

# S3 method for cgam
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("conditional", "smooth_terms", "all"),
  standardize = NULL,
  exponentiate = FALSE,
  ...
)

Arguments

model
A gam/gamm model.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

bootstrap
Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also bootstrap_parameters()).

iterations
The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

standardize
The method used for standardizing the parameters. Can be "refit", "posthoc", "smart", "basic" or NULL (default) for no standardization. See 'Details' in standardize_parameters. Note that robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for, say, logistic regressions, or more generally speaking: for models with log or logit link. Note: standard errors are also transformed (by multiplying the standard errors with the exponentiated coefficients), to mimic behaviour of other software packages, such as Stata.

robust
Logical, if TRUE, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors. Additional arguments like vcov_estimation or vcov_type are passed down to other methods, see standard_error_robust() for details.

p_adjust
Character vector, if not NULL, indicates the method to adjust p-values. See p.adjust for details.

...
Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like ci_method are passed down to describe_posterior.

component
Model component for which parameters should be shown. May be one of "conditional", "precision" (betareg), "scale" (ordinal), "extra" (glmx) or "all".

Value

A data frame of indices related to the model's parameters.

See also

Examples

library(parameters)
library(mgcv)

#> Loading required package: nlme

#> 
#> Attaching package: 'nlme'

#> The following object is masked from 'package:lme4':
#> 
#>     lmList

#> This is mgcv 1.8-31. For overview type 'help("mgcv-package")'.

dat <- gamSim(1, n = 400, dist = "normal", scale = 2)

#> Gu & Wahba 4 term additive model

model <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat)
model_parameters(model)

#> # Fixed Effects component
#> 
#> Parameter   | Coefficient |   SE |       95% CI |     t |     df |      p
#> -------------------------------------------------------------------------
#> (Intercept) |        8.17 | 0.09 | [7.99, 8.35] | 87.58 | 380.71 | < .001
#> 
#> # Smooth Terms component
#> 
#> Parameter        | Coefficient |      t |     df |      p
#> ---------------------------------------------------------
#> Smooth term (x0) |        3.26 |  21.58 | 380.71 | < .001
#> Smooth term (x1) |        6.59 |  45.56 | 380.71 | < .001
#> Smooth term (x2) |        7.44 | 103.13 | 380.71 | < .001
#> Smooth term (x3) |        1.00 |   1.47 | 380.71 | 0.226 
#>