Parameters from Hypothesis Testing

Parameters from Hypothesis Testing.

# S3 method for glht
model_parameters(model, ci = 0.95, exponentiate = FALSE, ...)

Arguments

model
Object of class glht (multcomp).

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for, say, logistic regressions, or more generally speaking: for models with log or logit link. Note: standard errors are also transformed (by multiplying the standard errors with the exponentiated coefficients), to mimic behaviour of other software packages, such as Stata.

...
Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like ci_method are passed down to describe_posterior.

Value

A data frame of indices related to the model's parameters.

Examples

# \donttest{
if (require("multcomp")) {
  # multiple linear model, swiss data
  lmod <- lm(Fertility ~ ., data = swiss)
  mod <- glht(
    model = lmod,
    linfct = c(
      "Agriculture = 0",
      "Examination = 0",
      "Education = 0",
      "Catholic = 0",
      "Infant.Mortality = 0"
    )
  )
  model_parameters(mod)
}

#> Loading required package: multcomp

#> Loading required package: mvtnorm

#> 
#> Attaching package: ‘mvtnorm’

#> The following object is masked from ‘package:mclust’:
#> 
#>     dmvnorm

#> Loading required package: survival

#> Loading required package: TH.data

#> Loading required package: MASS

#> 
#> Attaching package: ‘MASS’

#> The following object is masked from ‘package:multimode’:
#> 
#>     geyser

#> 
#> Attaching package: ‘TH.data’

#> The following object is masked from ‘package:MASS’:
#> 
#>     geyser

#> The following object is masked from ‘package:multimode’:
#> 
#>     geyser

#> Warning: no 'nobs' method is available

#> Parameter             | Coefficient |   SE |         95% CI | t(41) |      p
#> ----------------------------------------------------------------------------
#> Agriculture == 0      |       -0.17 | 0.07 | [-0.36,  0.01] | -2.45 | 0.079 
#> Examination == 0      |       -0.26 | 0.25 | [-0.93,  0.41] | -1.02 | 0.785 
#> Education == 0        |       -0.87 | 0.18 | [-1.36, -0.39] | -4.76 | < .001
#> Catholic == 0         |        0.10 | 0.04 | [ 0.01,  0.20] |  2.95 | 0.023 
#> Infant.Mortality == 0 |        1.08 | 0.38 | [ 0.07,  2.09] |  2.82 | 0.033 

# }