Parameters of (General) Linear Models

Extract and compute indices and measures to describe parameters of (general) linear models (GLMs).

# S3 method for lm
model_parameters(model, ci = 0.95, standardize = FALSE,
  standardize_robust = FALSE, bootstrap = FALSE, iterations = 1000,
  ...)

# S3 method for polr
model_parameters(model, ci = 0.95, bootstrap = FALSE,
  iterations = 1000, ...)

Arguments

model
Model object.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

standardize
Add standardized parameters. Can be FALSE or a character indicating the standardization method (see parameters_standardize()), such as "refit", "2sd", "smart" or "classic". The two formers are based on model refitting using a standardized version of data. It is the most accurate, altough computationnaly heavy (as it must re-fit a second model). The "smart" and "classic" are post-hoc methods, fast, but inaccurate (especially if the model includes interactions).

standardize_robust
Robust standardization. See parameters_standardize.

bootstrap
Should estimates be based on bootsrapped model? If TRUE, then arguments of Bayesian regressions apply (see also parameters_bootstrap()).

iterations
The number of bootstrap replicates. This only apply in the case of bootsrapped frequentist models.

...
Arguments passed to or from other methods (e.g., to standardize()).

Value

A data frame of indices related to the model's parameters.

Note

Standardization (argument standardize) is not supported by all model objects.

See also

Examples

library(parameters)
model <- lm(mpg ~ wt + cyl, data = mtcars)

model_parameters(model, standardize = "refit")

#> Parameter   | Coefficient |   SE |         95% CI |     t | df |      p | Coefficient (std.)
#> --------------------------------------------------------------------------------------------
#> (Intercept) |       39.69 | 1.71 | [36.18, 43.19] | 23.14 | 29 | < .001 |               0.00
#> wt          |       -3.19 | 0.76 | [-4.74, -1.64] | -4.22 | 29 | < .001 |              -0.52
#> cyl         |       -1.51 | 0.41 | [-2.36, -0.66] | -3.64 | 29 | < .01  |              -0.45

model_parameters(model, bootstrap = TRUE)

#> Parameter   | Coefficient |         95% CI |      p
#> ---------------------------------------------------
#> (Intercept) |       39.47 | [35.28, 44.09] | < .001
#> wt          |       -3.23 | [-4.99, -1.85] | < .001
#> cyl         |       -1.47 | [-2.20, -0.70] | < .001

model <- glm(vs ~ wt + cyl, data = mtcars, family = "binomial")
model_parameters(model)

#> Waiting for profiling to be done...

#> Parameter   | Coefficient |   SE |         95% CI |     z | df |     p
#> ----------------------------------------------------------------------
#> (Intercept) |       10.62 | 4.17 | [ 4.79, 22.66] |  2.55 | 29 | < .05
#> wt          |        2.10 | 1.55 | [-0.53,  6.24] |  1.36 | 29 | > .1 
#> cyl         |       -2.93 | 1.38 | [-6.92, -1.07] | -2.12 | 29 | < .05