Mixed Model Parameters

Parameters of mixed models.

# S3 method for glmmTMB
model_parameters(model, ci = 0.95,
  standardize = FALSE, standardize_robust = FALSE, bootstrap = FALSE,
  iterations = 1000, component = c("all", "conditional", "zi",
  "zero_inflated"), ...)

# S3 method for lme
model_parameters(model, ci = 0.95, standardize = FALSE,
  standardize_robust = FALSE, bootstrap = FALSE, iterations = 1000,
  ...)

# S3 method for merMod
model_parameters(model, ci = 0.95,
  standardize = FALSE, standardize_robust = FALSE, bootstrap = FALSE,
  p_method = "wald", ci_method = "wald", iterations = 1000, ...)

Arguments

model
A mixed model.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

standardize
Add standardized parameters. Can be FALSE or a character indicating the standardization method (see parameters_standardize()), such as "refit", "2sd", "smart" or "classic". The two formers are based on model refitting using a standardized version of data. It is the most accurate, altough computationnaly heavy (as it must re-fit a second model). The "smart" and "classic" are post-hoc methods, fast, but inaccurate (especially if the model includes interactions).

standardize_robust
Robust standardization. See parameters_standardize.

bootstrap
Should estimates be based on bootsrapped model? If TRUE, then arguments of Bayesian regressions apply (see also parameters_bootstrap()).

iterations
The number of draws to simulate/bootstrap.

component
Should all parameters, parameters for the conditional model, or for the zero-inflated part of the model be returned? Applies to models with zero-inflated component. component may be one of "conditional", "zi", "zero-inflated" or "all" (default). May be abbreviated.

...
Arguments passed to or from other methods.

p_method
Method for computing p values. See p_value().

ci_method
Method for computing confidence intervals (CI). See ci().

Value

A data frame of indices related to the model's parameters.

See also

Examples

library(parameters)
library(lme4)
library(glmmTMB)

model <- lmer(mpg ~ wt + (1 | gear), data = mtcars)
model_parameters(model, standardize = "refit")

#> Parameter   | Coefficient |   SE |         95% CI |     t |      p | Coefficient (std.)
#> ---------------------------------------------------------------------------------------
#> (Intercept) |       36.19 | 2.19 | [31.90, 40.48] | 16.52 | < .001 |              -0.02
#> wt          |       -5.05 | 0.64 | [-6.30, -3.79] | -7.89 | < .001 |              -0.82

model <- glmmTMB(
  count ~ spp + mined + (1 | site),
  ziformula = ~mined,
  family = poisson(),
  data = Salamanders
)
model_parameters(model)

#> # Conditional component
#> 
#> Parameter   | Coefficient |   SE |         95% CI |     z |      p
#> ------------------------------------------------------------------
#> (Intercept) |       -0.36 | 0.28 | [-0.90,  0.18] | -1.30 | > .1  
#> spp (PR)    |       -1.27 | 0.24 | [-1.74, -0.80] | -5.27 | < .001
#> spp (DM)    |        0.27 | 0.14 | [ 0.00,  0.54] |  1.95 | 0.05  
#> spp (EC-A)  |       -0.57 | 0.21 | [-0.97, -0.16] | -2.75 | < .01 
#> spp (EC-L)  |        0.67 | 0.13 | [ 0.41,  0.92] |  5.20 | < .001
#> spp (DES-L) |        0.63 | 0.13 | [ 0.38,  0.87] |  4.96 | < .001
#> spp (DF)    |        0.12 | 0.15 | [-0.17,  0.40] |  0.78 | > .1  
#> mined (no)  |        1.27 | 0.27 | [ 0.74,  1.80] |  4.72 | < .001
#> 
#> # Zero-Inflated component
#> 
#> Parameter   | Coefficient |   SE |         95% CI |     z |      p
#> ------------------------------------------------------------------
#> (Intercept) |        0.79 | 0.27 | [ 0.26,  1.32] |  2.90 | < .01 
#> mined (no)  |       -1.84 | 0.31 | [-2.46, -1.23] | -5.87 | < .001
#> 

# \donttest{
model <- lme4::lmer(mpg ~ wt + (1 | gear), data = mtcars)
model_parameters(model, standardize = "smart", bootstrap = TRUE, iterations = 50)

#> Parameter   | Coefficient |         95% CI |      p | Coefficient (std.)
#> ------------------------------------------------------------------------
#> (Intercept) |       36.07 | [32.53, 39.35] | < .001 |               0.00
#> wt          |       -5.14 | [-6.03, -4.01] | < .001 |              -0.82

# }