Parameters from multinomial or cumulative link models

Parameters from multinomial or cumulative link models

# S3 method for DirichletRegModel
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "precision"),
  standardize = NULL,
  exponentiate = FALSE,
  verbose = TRUE,
  ...
)

# S3 method for bifeAPEs
model_parameters(model, ...)

# S3 method for bracl
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for mlm
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for clm2
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "scale"),
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

Arguments

model

A model with multinomial or categorical response value.

ci

Confidence Interval (CI) level. Default to 0.95 (95%).

bootstrap

Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also bootstrap_parameters()).

iterations

The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

component

Model component for which parameters should be shown. May be one of "conditional", "precision" (betareg), "scale" (ordinal), "extra" (glmx), "marginal" (mfx), "conditional" or "full" (for MuMIn::model.avg()) or "all".

standardize

The method used for standardizing the parameters. Can be NULL (default; no standardization), "refit" (for re-fitting the model on standardized data) or one of "basic", "posthoc", "smart", "pseudo". See 'Details' in effectsize::standardize_parameters(). Important:

The "refit" method does not standardized categorical predictors (i.e. factors), which may be a different behaviour compared to other R packages (such as lm.beta) or other software packages (like SPSS). to mimic such behaviours, either use standardize="basic" or standardize the data with effectsize::standardize(force=TRUE) before fitting the model.

For mixed models, when using methods other than "refit", only the fixed effects will be returned.

Robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

exponentiate

Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for logistic regression, or more generally speaking, for models with log or logit links. Note: Delta-method standard errors are also computed (by multiplying the standard errors by the transformed coefficients). This is to mimic behaviour of other software packages, such as Stata, but these standard errors poorly estimate uncertainty for the transformed coefficient. The transformed confidence interval more clearly captures this uncertainty. For compare_parameters(), exponentiate = "nongaussian" will only exponentiate coefficients from non-Gaussian families.

verbose

Toggle warnings and messages.

...

Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like type or parallel are passed down to bootstrap_model(), and arguments like ci_method are passed down to bayestestR::describe_posterior().

p_adjust

Character vector, if not NULL, indicates the method to adjust p-values. See stats::p.adjust() for details. Further possible adjustment methods are "tukey", "scheffe", "sidak" and "none" to explicitly disable adjustment for emmGrid objects (from emmeans).

Value

A data frame of indices related to the model's parameters.

Details

Multinomial or cumulative link models, i.e. models where the response value (dependent variable) is categorical and has more than two levels, usually return coefficients for each response level. Hence, the output from model_parameters() will split the coefficient tables by the different levels of the model's response.

See also

Examples

library(parameters)
if (require("brglm2", quietly = TRUE)) {
  data("stemcell")
  model <- bracl(
    research ~ as.numeric(religion) + gender,
    weights = frequency,
    data = stemcell,
    type = "ML"
  )
  model_parameters(model)
}
#> # Response level: definitely
#> 
#> Parameter       | Log-Odds |   SE |         95% CI |     z |      p
#> -------------------------------------------------------------------
#> (Intercept)     |    -1.25 | 0.26 | [-1.76, -0.73] | -4.76 | < .001
#> religion        |     0.44 | 0.10 | [ 0.23,  0.64] |  4.20 | < .001
#> gender [female] |    -0.14 | 0.17 | [-0.47,  0.19] | -0.82 | 0.414 
#> 
#> # Response level: probably
#> 
#> Parameter       | Log-Odds |   SE |        95% CI |    z |     p
#> ----------------------------------------------------------------
#> (Intercept)     |     0.47 | 0.29 | [-0.10, 1.04] | 1.62 | 0.105
#> religion        |     0.26 | 0.13 | [ 0.01, 0.51] | 2.01 | 0.044
#> gender [female] |     0.19 | 0.21 | [-0.22, 0.60] | 0.90 | 0.370
#> 
#> # Response level: probably not
#> 
#> Parameter       | Log-Odds |   SE |        95% CI |     z |     p
#> -----------------------------------------------------------------
#> (Intercept)     |     0.43 | 0.39 | [-0.33, 1.18] |  1.11 | 0.268
#> religion        |     0.01 | 0.17 | [-0.33, 0.35] |  0.07 | 0.945
#> gender [female] |    -0.16 | 0.28 | [-0.71, 0.39] | -0.57 | 0.566
#> 
#> Uncertainty intervals (equal-tailed) and p values (two-tailed) computed using a
#>   Wald z-distribution approximation.