Model Parameters for Zero-Inflated Models

Parameters of zero-inflated models.

# S3 method for zeroinfl
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "zi", "zero_inflated"),
  standardize = NULL,
  exponentiate = FALSE,
  ...
)

Arguments

model
A model with zero-inflation component.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

bootstrap
Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also parameters_bootstrap()).

iterations
The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

component
Should all parameters, parameters for the conditional model, or for the zero-inflated part of the model be returned? Applies to models with zero-inflated component. component may be one of "conditional", "zi", "zero-inflated" or "all" (default). May be abbreviated.

standardize
The method used for standardizing the parameters. Can be "refit", "posthoc", "smart", "basic" or NULL (default) for no standardization. See 'Details' in standardize_parameters.

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for, say, logistic regressions, or more generally speaking: for models with log or logit link.

...
Arguments passed to or from other methods.

Value

A data frame of indices related to the model's parameters.

See also

Examples

library(parameters)
library(pscl)

#> Classes and Methods for R developed in the
#> Political Science Computational Laboratory
#> Department of Political Science
#> Stanford University
#> Simon Jackman
#> hurdle and zeroinfl functions by Achim Zeileis

data("bioChemists")
model <- zeroinfl(art ~ fem + mar + kid5 + ment | kid5 + phd, data = bioChemists)
model_parameters(model)

#> # Fixed Effects component
#> 
#> Parameter     | Coefficient |   SE |         95% CI |     z |  df |      p
#> --------------------------------------------------------------------------
#> (Intercept)   |        0.56 | 0.07 | [ 0.43,  0.69] |  8.26 | 907 | < .001
#> fem [Women]   |       -0.23 | 0.06 | [-0.34, -0.11] | -3.91 | 907 | < .001
#> mar [Married] |        0.14 | 0.07 | [ 0.01,  0.27] |  2.07 | 907 | 0.038 
#> kid5          |       -0.17 | 0.05 | [-0.26, -0.07] | -3.43 | 907 | < .001
#> ment          |        0.02 | 0.00 | [ 0.02,  0.03] | 10.05 | 907 | < .001
#> 
#> # Zero-Inflated component
#> 
#> Parameter   | Coefficient |   SE |         95% CI |     z |  df |     p
#> -----------------------------------------------------------------------
#> (Intercept) |       -0.93 | 0.43 | [-1.78, -0.08] | -2.14 | 907 | 0.032
#> kid5        |        0.05 | 0.22 | [-0.38,  0.47] |  0.21 | 907 | 0.831
#> phd         |       -0.25 | 0.14 | [-0.51,  0.02] | -1.84 | 907 | 0.065
#>