Count number of parameters in a model

Returns the number of parameters of a model.

n_parameters(x, ...)

# S3 method for default
n_parameters(x, ...)

# S3 method for merMod
n_parameters(x, effects = c("fixed", "random"), ...)

# S3 method for glmmTMB
n_parameters(
  x,
  effects = c("fixed", "random"),
  component = c("all", "conditional", "zi", "zero_inflated"),
  ...
)

# S3 method for zeroinfl
n_parameters(
  x,
  component = c("all", "conditional", "zi", "zero_inflated"),
  ...
)

# S3 method for gam
n_parameters(x, component = c("all", "conditional", "smooth_terms"), ...)

# S3 method for brmsfit
n_parameters(
  x,
  effects = c("all", "fixed", "random"),
  component = c("all", "conditional", "zi", "zero_inflated", "dispersion", "simplex",
    "sigma", "smooth_terms"),
  ...
)

Arguments

x
A statistical model.

...
Arguments passed to or from other methods.

effects
Should number of parameters for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

component
Should total number of parameters, number parameters for the conditional model, the zero-inflated part of the model, the dispersion term or the instrumental variables be returned? Applies to models with zero-inflated and/or dispersion formula, or to models with instrumental variable (so called fixed-effects regressions). May be abbreviated.

Value

The number of parameters in the model.

Examples

data(iris)
model <- lm(Sepal.Length ~ Sepal.Width * Species, data = iris)
n_parameters(model)

#> [1] 6