p-values for Marginal Effects Models

This function attempts to return, or compute, p-values of marginal effects models from package mfx.

# S3 method for poissonmfx
p_value(model, component = c("all", "conditional", "marginal"), ...)

# S3 method for betaor
p_value(model, component = c("all", "conditional", "precision"), ...)

# S3 method for betamfx
p_value(
  model,
  component = c("all", "conditional", "precision", "marginal"),
  ...
)

Arguments

model
A statistical model.

component
Should all parameters, parameters for the conditional model, precision-component or marginal effects be returned? component may be one of "conditional", "precision", "marginal" or "all" (default).

...
Currently not used.

Value

A data frame with at least two columns: the parameter names and the p-values. Depending on the model, may also include columns for model components etc.

Examples

if (require("mfx", quietly = TRUE)) {
  set.seed(12345)
  n <- 1000
  x <- rnorm(n)
  y <- rnegbin(n, mu = exp(1 + 0.5 * x), theta = 0.5)
  d <- data.frame(y, x)
  model <- poissonmfx(y ~ x, data = d)

  p_value(model)
  p_value(model, component = "marginal")
}
#> 
#> Attaching package: ‘zoo’
#> The following objects are masked from ‘package:base’:
#> 
#>     as.Date, as.Date.numeric
#>   Parameter             p Component
#> 1         x 9.631324e-254  marginal