Wald-test approximation for CIs and p-values

The Wald-test approximation treats t-values as Wald z. Since the t distribution converges to the z distribution as degrees of freedom increase, this is like assuming infinite degrees of freedom. While this is unambiguously anti-conservative, this approximation appears as reasonable for reasonable sample sizes (Barr et al., 2013). That is, if we take the p-value to measure the probability of a false positive, this approximation produces a higher false positive rate than the nominal 5% at p = 0.05.

ci_wald(model, ci = 0.95, dof = NULL, component = c("all",
  "conditional", "zi", "zero_inflated"))

p_value_wald(model)

Arguments

model
A statistical model.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

dof
Degrees of Freedom. If not specified, defaults to model's residual degrees of freedom (i.e. n-k, where n is the number of observations and k is the number of parameters).

component
Should all parameters, parameters for the conditional model, or for the zero-inflated part of the model be returned? Applies to models with zero-inflated component. component may be one of "conditional", "zi", "zero-inflated" or "all" (default). May be abbreviated.

Value

The p-values.

References

Barr, D. J. (2013). Random effects structure for testing interactions in linear mixed-effects models. Frontiers in psychology, 4, 328.

Examples

# \donttest{
library(lme4)
model <- lmer(Petal.Length ~ Sepal.Length + (1 | Species), data = iris)
p_value_wald(model)

#>      Parameter            p
#> 1  (Intercept) 9.561741e-01
#> 2 Sepal.Length 1.273754e-44

ci_wald(model, ci = c(0.90, 0.95))

#>      Parameter CI     CI_low   CI_high
#> 1  (Intercept) 90 -1.5294928 1.6345428
#> 2 Sepal.Length 90  0.5592401 0.7090343
#> 3  (Intercept) 95 -1.8362096 1.9412596
#> 4 Sepal.Length 95  0.5447193 0.7235551

# }