Dimensionality reduction (DR) / Features Reduction

This function performs a reduction in the parameters space (the number of variables). It starts by creating a new set of variables, based on a given method (the default method is "PCA", but other are available via the method argument, such as "cMDS", "DRR" or "ICA"). Then, it names this new dimensions using the original variables that correlates the most with it. For instance, a variable named 'V1_0.97/V4_-0.88' means that the V1 and the V4 variables correlate maximally (with respective coefficients of .97 and -.88) with this dimension. Although this function can be useful in exploratory data analysis, it's best to perform the dimension reduction step in a separate and dedicated stage, as this is a very important process in the data analysis workflow.

parameters_reduction(x, method = "PCA", n = "max", ...)

Arguments

x
A dataframe or a statistical model.

method
The features reduction method. Can be one of 'PCA', 'cMDS', 'DRR', 'ICA' (see the Details section).

n
Number of components to extract. If n = "all" =, then n is set as the number of variables minus 1 (ncol(x)-1). If n = "auto" (default) or n = NULL, the number of components is selected through n_factors. In parameters_reduction, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

...
Arguments passed to or from other methods.

Details

The different methods available are described below:

Supervised Methods

PCA: See principal_components.

cMDS / PCoA: See cmds. Classical Multidimensional Scaling (cMDS) takes a set of dissimilarities (i.e., a distance matrix) and returns a set of points such that the distances between the points are approximately equal to the dissimilarities.

DRR: See DRR. Dimensionality Reduction via Regression (DRR) is a very recent technique extending PCA (Laparra et al., 2015). Starting from a rotated PCA, it predicts redundant information from the remaining components using non-linear regression. Some of the most notable advantages of performing PCR are avoidance of multicollinearity between predictors and overfitting mitigation. PCR tends to perform well when the first principal components are enough to explain most of the variation in the predictors. Requires the DRR package to be installed.

ICA: See ICA. Performs an Independent Component Analysis using the FastICA algorithm. Contrary to PCA, that attempts to find uncorrelated sources (through least squares minimization), ICA attempts to find independent sources, i.e., the source space that maximizes the "non-gaussianity" of all sources. Contrary to PCA, ICA does not rank each source, which makes it a poor tool for dimensionality reduction. Requires the fastICA package to be installed.

References

Nguyen, L. H., \& Holmes, S. (2019). Ten quick tips for effective dimensionality reduction. PLOS Computational Biology, 15(6).

Laparra, V., Malo, J., & Camps-Valls, G. (2015). Dimensionality reduction via regression in hyperspectral imagery. IEEE Journal of Selected Topics in Signal Processing, 9(6), 1026-1036.

Examples

parameters_reduction(iris, method = "PCA", n = "max")

#>     Petal.Length_0.99/Petal.Width_0.97/Species.setosa_-0.94/Sepal.Length_0.86/Species.virginica_0.73
#> 1                                                                                        -2.80385224
#> 2                                                                                        -2.63303450
#> 3                                                                                        -2.86692343
#> 4                                                                                        -2.80865556
#> 5                                                                                        -2.90734273
#> 6                                                                                        -2.66852255
#> 7                                                                                        -2.93651068
#> 8                                                                                        -2.77249231
#> 9                                                                                        -2.82665950
#> 10                                                                                       -2.72106438
#> 11                                                                                       -2.73632905
#> 12                                                                                       -2.84462287
#> 13                                                                                       -2.74352032
#> 14                                                                                       -3.07184037
#> 15                                                                                       -2.78088959
#> 16                                                                                       -2.84363690
#> 17                                                                                       -2.77615428
#> 18                                                                                       -2.74288567
#> 19                                                                                       -2.52696011
#> 20                                                                                       -2.87789148
#> 21                                                                                       -2.52059944
#> 22                                                                                       -2.76295367
#> 23                                                                                       -3.21305147
#> 24                                                                                       -2.43228625
#> 25                                                                                       -2.76389907
#> 26                                                                                       -2.52969938
#> 27                                                                                       -2.62365124
#> 28                                                                                       -2.72742505
#> 29                                                                                       -2.70036174
#> 30                                                                                       -2.78619963
#> 31                                                                                       -2.68270913
#> 32                                                                                       -2.45248218
#> 33                                                                                       -3.11221910
#> 34                                                                                       -2.98357397
#> 35                                                                                       -2.66009781
#> 36                                                                                       -2.74527361
#> 37                                                                                       -2.63268317
#> 38                                                                                       -3.01782855
#> 39                                                                                       -2.90753868
#> 40                                                                                       -2.72297306
#> 41                                                                                       -2.81931286
#> 42                                                                                       -2.41925414
#> 43                                                                                       -3.01548118
#> 44                                                                                       -2.55568936
#> 45                                                                                       -2.70929318
#> 46                                                                                       -2.62158719
#> 47                                                                                       -2.91195011
#> 48                                                                                       -2.88953475
#> 49                                                                                       -2.78584830
#> 50                                                                                       -2.74542899
#> 51                                                                                        1.08875588
#> 52                                                                                        0.79879107
#> 53                                                                                        1.20799031
#> 54                                                                                        0.58238625
#> 55                                                                                        1.09110325
#> 56                                                                                        0.54610818
#> 57                                                                                        0.81008301
#> 58                                                                                       -0.13995573
#> 59                                                                                        0.96471812
#> 60                                                                                        0.25200214
#> 61                                                                                        0.17926438
#> 62                                                                                        0.57841352
#> 63                                                                                        0.70105405
#> 64                                                                                        0.80499637
#> 65                                                                                        0.20044628
#> 66                                                                                        0.91344557
#> 67                                                                                        0.51057957
#> 68                                                                                        0.35906725
#> 69                                                                                        1.23946504
#> 70                                                                                        0.37512194
#> 71                                                                                        0.81481832
#> 72                                                                                        0.60964551
#> 73                                                                                        1.23470229
#> 74                                                                                        0.73703449
#> 75                                                                                        0.78495582
#> 76                                                                                        0.91789757
#> 77                                                                                        1.23251030
#> 78                                                                                        1.31176412
#> 79                                                                                        0.76262781
#> 80                                                                                        0.20207164
#> 81                                                                                        0.35266601
#> 82                                                                                        0.26479151
#> 83                                                                                        0.42718451
#> 84                                                                                        1.09298448
#> 85                                                                                        0.41154107
#> 86                                                                                        0.55373814
#> 87                                                                                        1.05513594
#> 88                                                                                        1.08617198
#> 89                                                                                        0.28101470
#> 90                                                                                        0.47444375
#> 91                                                                                        0.46713768
#> 92                                                                                        0.72411719
#> 93                                                                                        0.50806369
#> 94                                                                                       -0.03646523
#> 95                                                                                        0.46983638
#> 96                                                                                        0.29647532
#> 97                                                                                        0.41141313
#> 98                                                                                        0.68591732
#> 99                                                                                       -0.11464571
#> 100                                                                                       0.43847645
#> 101                                                                                       2.22734360
#> 102                                                                                       1.69560404
#> 103                                                                                       2.51463715
#> 104                                                                                       1.90883090
#> 105                                                                                       2.25158028
#> 106                                                                                       2.95058893
#> 107                                                                                       1.07449255
#> 108                                                                                       2.59237894
#> 109                                                                                       2.37660876
#> 110                                                                                       2.53801104
#> 111                                                                                       1.83334912
#> 112                                                                                       2.04653541
#> 113                                                                                       2.25844766
#> 114                                                                                       1.78808591
#> 115                                                                                       1.94646561
#> 116                                                                                       2.02054543
#> 117                                                                                       1.92699022
#> 118                                                                                       2.65621271
#> 119                                                                                       3.41865010
#> 120                                                                                       1.79372458
#> 121                                                                                       2.37577341
#> 122                                                                                       1.54974499
#> 123                                                                                       3.07399205
#> 124                                                                                       1.82841786
#> 125                                                                                       2.10083054
#> 126                                                                                       2.30022214
#> 127                                                                                       1.69801942
#> 128                                                                                       1.56746561
#> 129                                                                                       2.19522109
#> 130                                                                                       2.23241564
#> 131                                                                                       2.70302013
#> 132                                                                                       2.55259427
#> 133                                                                                       2.25618765
#> 134                                                                                       1.64536278
#> 135                                                                                       1.72783988
#> 136                                                                                       2.98750164
#> 137                                                                                       2.00477405
#> 138                                                                                       1.82349972
#> 139                                                                                       1.49103843
#> 140                                                                                       2.22708773
#> 141                                                                                       2.36476479
#> 142                                                                                       2.26829706
#> 143                                                                                       1.69560404
#> 144                                                                                       2.38007003
#> 145                                                                                       2.34469679
#> 146                                                                                       2.25013774
#> 147                                                                                       2.02423485
#> 148                                                                                       1.96819954
#> 149                                                                                       1.84047237
#> 150                                                                                       1.52224298
#>     Species.versicolor_0.93 Sepal.Width_0.62
#> 1              -0.651958996       0.13657921
#> 2              -0.099245388      -0.82961672
#> 3              -0.265604667      -0.59840285
#> 4              -0.146224047      -0.81545922
#> 5              -0.735791021       0.25432088
#> 6              -1.147417167       1.00764061
#> 7              -0.464983694      -0.29982389
#> 8              -0.531273605      -0.08474660
#> 9               0.099061051      -1.27091921
#> 10             -0.192130117      -0.65857818
#> 11             -0.916323163       0.64596601
#> 12             -0.494420238      -0.18833076
#> 13             -0.068835182      -0.88844291
#> 14              0.030474477      -1.17088529
#> 15             -1.297458407       1.36387027
#> 16             -1.716600302       2.01922445
#> 17             -1.142198082       0.99056280
#> 18             -0.663290133       0.14147859
#> 19             -1.090412144       0.99285318
#> 20             -0.973328187       0.66075345
#> 21             -0.610199424       0.13949952
#> 22             -0.881748229       0.49398436
#> 23             -0.654255659       0.02153584
#> 24             -0.484044533      -0.17925121
#> 25             -0.498334552      -0.17552239
#> 26             -0.120934000      -0.76715101
#> 27             -0.555240650      -0.07067839
#> 28             -0.672342837       0.19477547
#> 29             -0.568126971       0.01883755
#> 30             -0.269518981      -0.58559449
#> 31             -0.185686956      -0.70333615
#> 32             -0.630252155       0.14075938
#> 33             -1.278478264       1.21988689
#> 34             -1.448652930       1.55396569
#> 35             -0.203461254      -0.65367880
#> 36             -0.321537103      -0.44089190
#> 37             -0.726970501       0.34801698
#> 38             -0.705380815       0.19549469
#> 39             -0.002545272      -1.10352020
#> 40             -0.550352674      -0.03081980
#> 41             -0.642906293       0.08328234
#> 42              0.687422180      -2.24637327
#> 43             -0.208367460      -0.76018327
#> 44             -0.680814018       0.11078884
#> 45             -0.989878409       0.68273064
#> 46             -0.091497456      -0.87864415
#> 47             -0.963301821       0.66012352
#> 48             -0.247830369      -0.64806020
#> 49             -0.897244094       0.59203921
#> 50             -0.427057739      -0.26068452
#> 51              1.078448585       1.67501815
#> 52              1.184201405       1.34781779
#> 53              1.186498068       1.46286117
#> 54              2.311299003      -0.71368567
#> 55              1.575461941       0.71934019
#> 56              1.752061538       0.27385754
#> 57              1.086428701       1.47899775
#> 58              2.365989134      -0.91016235
#> 59              1.476134051       0.93513670
#> 60              1.946865469      -0.18816229
#> 61              2.755944898      -1.53437050
#> 62              1.489333253       0.72203848
#> 63              2.352808163      -0.63041825
#> 64              1.558893488       0.67467151
#> 65              1.679972456       0.35317411
#> 66              1.242511200       1.32876091
#> 67              1.542656146       0.57306643
#> 68              1.875106059       0.12433992
#> 69              2.251470483      -0.47672047
#> 70              2.110364791      -0.33049015
#> 71              1.241689026       1.10569028
#> 72              1.682269119       0.46821749
#> 73              1.918439047       0.10928955
#> 74              1.684466856       0.49320429
#> 75              1.518206504       0.81447473
#> 76              1.364501363       1.10316564
#> 77              1.526946328       0.88476013
#> 78              1.303600255       1.19740731
#> 79              1.569250963       0.61710518
#> 80              2.004924851      -0.12687207
#> 81              2.233659726      -0.56035487
#> 82              2.246295634      -0.56952371
#> 83              1.855053328       0.12559977
#> 84              1.755913386       0.30428435
#> 85              1.580814284       0.46521282
#> 86              1.043364357       1.48034689
#> 87              1.227265749       1.34646865
#> 88              2.153447365      -0.26519342
#> 89              1.570537505       0.54618985
#> 90              2.105476815      -0.37034874
#> 91              2.008677773      -0.18650183
#> 92              1.457287166       0.84207052
#> 93              1.956659651      -0.04179924
#> 94              2.449821159      -1.02790401
#> 95              1.877966015       0.03545391
#> 96              1.561484801       0.59948673
#> 97              1.653064758       0.43271764
#> 98              1.556364642       0.70662112
#> 99              2.217503079      -0.63854926
#> 100             1.757280624       0.25677972
#> 101            -1.512241066       0.13884131
#> 102            -0.719649393      -1.22862487
#> 103            -1.309511018       0.03138337
#> 104            -1.016059647      -0.59920603
#> 105            -1.205062969      -0.29154753
#> 106            -1.414039763       0.33090357
#> 107            -0.311624682      -2.09271853
#> 108            -1.215983737      -0.03005181
#> 109            -0.683341090      -1.06163377
#> 110            -1.993990741       1.14345740
#> 111            -1.379089484       0.01210446
#> 112            -0.836733351      -0.89652514
#> 113            -1.247054725      -0.14747486
#> 114            -0.504774502      -1.62525868
#> 115            -0.879216172      -1.03245951
#> 116            -1.396613368      -0.01858529
#> 117            -1.155824107      -0.32395341
#> 118            -2.269043492       1.76734694
#> 119            -1.048051044      -0.27923636
#> 120            -0.196622743      -2.00298057
#> 121            -1.497227799       0.26812655
#> 122            -0.793123943      -1.16844955
#> 123            -1.217270279       0.04086352
#> 124            -0.801104059      -0.97242914
#> 125            -1.539318481       0.32214265
#> 126            -1.501723636       0.41821842
#> 127            -0.883631313      -0.85895693
#> 128            -1.071679203      -0.56527735
#> 129            -0.966221033      -0.70224955
#> 130            -1.270629631       0.05654382
#> 131            -1.140873306      -0.15143300
#> 132            -2.280625042       1.85259343
#> 133            -0.977552169      -0.69735017
#> 134            -0.872631285      -0.80691991
#> 135            -0.623843679      -1.24166256
#> 136            -1.449257249       0.37328187
#> 137            -1.598601938       0.28853258
#> 138            -1.239656132      -0.20621175
#> 139            -1.051295362      -0.62347361
#> 140            -1.367740117       0.07385096
#> 141            -1.366184932      -0.01076560
#> 142            -1.386488077       0.07084136
#> 143            -0.719649393      -1.22862487
#> 144            -1.480758273       0.22273865
#> 145            -1.584643028       0.34174017
#> 146            -1.246723616      -0.20441126
#> 147            -0.607917779      -1.30659724
#> 148            -1.174572067      -0.32696301
#> 149            -1.565582189       0.22116749
#> 150            -1.036130608      -0.66459205