Parameters Selection

This function performs an automated selection of the 'best' parameters, udpating and returning the "best" model. For frequentist simple GLMs, it performs an AIC-based stepwise selection. For Bayesian models, it uses the projpred package.

parameters_selection(model, ...)

# S3 method for lm
parameters_selection(model, direction = "both",
  steps = 1000, k = 2, ...)

# S3 method for merMod
parameters_selection(model, direction = "backward",
  steps = 1000, ...)

# S3 method for stanreg
parameters_selection(model, method = NULL,
  cross_validation = FALSE, ...)

Arguments

model
A statistical model.

...
Arguments passed to or from other methods.

direction
the mode of stepwise search, can be one of "both", "backward", or "forward", with a default of "both". If the scope argument is missing the default for direction is "backward". Values can be abbreviated.

steps
the maximum number of steps to be considered. The default is 1000 (essentially as many as required). It is typically used to stop the process early.

k
the multiple of the number of degrees of freedom used for the penalty. Only k = 2 gives the genuine AIC: k = log(n) is sometimes referred to as BIC or SBC.

method
The method used in the variable selection. Can be NULL (default), "forward" or "L1". See projpred::varsel.

cross_validation
Select with cross-validation.

Examples

model <- lm(mpg ~ ., data = mtcars)
parameters_selection(model)

#> 
#> Call:
#> lm(formula = mpg ~ wt + qsec + am, data = mtcars)
#> 
#> Coefficients:
#> (Intercept)           wt         qsec           am  
#>       9.618       -3.917        1.226        2.936  
#> 

model <- lm(mpg ~ cyl * disp * hp * wt, data = mtcars)
parameters_selection(model)

#> 
#> Call:
#> lm(formula = mpg ~ cyl + disp + hp + wt + cyl:disp + cyl:hp + 
#>     disp:hp + cyl:wt + disp:wt + hp:wt + cyl:disp:hp + cyl:hp:wt, 
#>     data = mtcars)
#> 
#> Coefficients:
#> (Intercept)          cyl         disp           hp           wt     cyl:disp  
#>  49.1436077   -3.6167276   -1.2955318   -0.0004854   58.8328841    0.1704703  
#>      cyl:hp      disp:hp       cyl:wt      disp:wt        hp:wt  cyl:disp:hp  
#>  -0.0134573    0.0132124   -7.4915051   -0.0167172   -0.6524341   -0.0016542  
#>   cyl:hp:wt  
#>   0.0850798  
#> 
# NOT RUN {
# lme4 -------------------------------------------
library(lme4)
model <- lmer(Sepal.Width ~ Sepal.Length * Petal.Width * Petal.Length + (1 | Species), data = iris)
parameters_selection(model)

# rstanarm -------------------------------------------
library(rstanarm)
model <- stan_glm(mpg ~ ., data = mtcars)
parameters_selection(model, cross_validation = TRUE)

model <- stan_glm(mpg ~ cyl * disp * hp, data = mtcars)
parameters_selection(model, cross_validation = FALSE)
# }