Parameters simulation

Compute simulated draws of parameters and their related indices such as Confidence Intervals (CI) and p-values. Simulating parameter draws can be seen as a (computationally faster) alternative to boostrapping.

parameters_simulate(model, iterations = 1000, centrality = "median",
  ci = 0.95, ci_method = "quantile", test = "p-value", ...)

Arguments

model
Statistical model (no Bayesian models).

iterations
The number of draws to simulate/bootstrap.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

ci
Value or vector of probability of the CI (between 0 and 1) to be estimated. Default to .89 (89%) for Bayesian models and .95 (95%) for frequentist models.

ci_method
The type of index used for Credible Interval. Can be "HDI" (default, see hdi) or "ETI" (see eti).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests. For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

...
Arguments passed to or from other methods.

Value

A data frame with simulated parameters.

Details

Technical Details

model_simulate() is a computationally faster alternative to model_bootstrap(). Simulated draws for coefficients are based on a multivariate normal distribution (MASS::mvrnorm()) with mean mu = coef(model) and variance Sigma = vcov(model).

Models with Zero-Inflation Component

For models from packages glmmTMB, pscl, GLMMadaptive and countreg, the component argument can be used to specify which parameters should be simulated. For all other models, parameters from the conditional component (fixed effects) are simulated. This may include smooth terms, but not random effects.

References

Gelman A, Hill J. Data analysis using regression and multilevel/hierarchical models. Cambridge; New York: Cambridge University Press 2007: 140-143

See also

Examples

library(parameters)
library(glmmTMB)

model <- lm(Sepal.Length ~ Species * Petal.Width + Petal.Length, data = iris)
parameters_simulate(model)

#>                       Parameter Coefficient     CI_low    CI_high           p
#> 1                   (Intercept)   3.5291651  3.2026543  3.8481421 0.000000000
#> 2             Speciesversicolor  -1.1541920 -1.8471148 -0.4826065 0.001998002
#> 3              Speciesvirginica  -2.2293532 -3.1155210 -1.4045782 0.000000000
#> 4                   Petal.Width   0.4143481 -0.4959095  1.3357613 0.353646354
#> 5                  Petal.Length   0.9406770  0.7891260  1.0900810 0.000000000
#> 6 Speciesversicolor:Petal.Width  -0.7667890 -1.7285082  0.3051246 0.171828172
#> 7  Speciesvirginica:Petal.Width  -0.3692775 -1.3951537  0.5795459 0.423576424

model <- glmmTMB(
  count ~ spp + mined + (1 | site),
  ziformula =  ~ mined,
  family = poisson(),
  data = Salamanders
)
parameters_simulate(model, centrality = "mean")

#>         Parameter Coefficient       CI_low    CI_high           p
#> 1     (Intercept)  -0.3606631 -0.898176398  0.1798020 0.189810190
#> 2           sppPR  -1.2704593 -1.731652761 -0.8178404 0.000000000
#> 3           sppDM   0.2678054  0.001488994  0.5265815 0.049950050
#> 4         sppEC-A  -0.5686862 -0.988782861 -0.1634632 0.003996004
#> 5         sppEC-L   0.6647610  0.429755608  0.9050588 0.000000000
#> 6        sppDES-L   0.6263250  0.377223191  0.8761781 0.000000000
#> 7           sppDF   0.1211258 -0.166757937  0.4074449 0.435564436
#> 8         minedno   1.2686665  0.745682566  1.7798440 0.000000000
#> 9  (Intercept)_zi   0.7805382  0.258107620  1.3211320 0.007992008
#> 10     minedno_zi  -1.8428741 -2.462464996 -1.2228668 0.000000000

parameters_simulate(model, ci = c(.8, .95), component = "zero_inflated")

#>     Parameter Coefficient CI     CI_low   CI_high           p
#> 1 (Intercept)   0.7773829 80  0.4249926  1.144516 0.003996004
#> 2 (Intercept)   0.7773829 95  0.2494981  1.360730 0.003996004
#> 3     minedno  -1.8333760 80 -2.2494140 -1.428800 0.000000000
#> 4     minedno  -1.8333760 95 -2.4820669 -1.187040 0.000000000