Parameters standardization

Compute standardized model parameters (coefficients).

parameters_standardize(model, robust = FALSE, method = "refit",
  verbose = TRUE, ...)

Arguments

model
A statistical model.

robust
Logical, if TRUE, centering is done by substracting the median from the variables and divide it by the median absolute deviation (MAD). If FALSE, variables are standardized by substracting the mean and divide it by the standard deviation (SD). the

method
The method used for standardizing the parameters. Can be "refit" (default), "2sd", "smart" or "classic".

verbose
Toggle warnings on or off.

...
Arguments passed to or from other methods.

Value

Standardized parameters.

Details

Methods:

refit: This method is based on a complete model re-fit with a standardized version of data. Hence, this method is equal to standardizing the variables before fitting the model. It is the "purest" and the most accurate (Neter et al., 1989), but it is also the most computationally costly and long (especially for Bayesian models). This method is particularly recommended for complex models that include interactions or transformations (e.g., polynomial or spline terms). The robust (default to FALSE) argument enables a robust standardization of data, i.e., based on the median and MAD instead of the mean and SD.

2sd: Same as method = "refit", however, standardization is done by dividing by two times the SD or MAD (depending on robust). This method is useful to obtain coefficients of continuous parameters comparable to coefficients related to binary predictors (see Gelman, 2008).

smart (Standardization of Model's parameters with Adjustement, Reconnaissance and Transformation): Post-hoc standardization of the parameters, aiming at emulating the results obtained by "refit". The coefficients are divided by the standard deviation (or MAD if robust) of the outcome (which becomes their expression 'unit'). Then, the coefficients related to numeric variables are additionaly multiplied by the standard deviation (or MAD if robust) of the related term, so that they correspond to changes of 1 SD of the predictor (e.g., "A change in 1 SD of x is related to a change of 0.24 of the SD of y). This does not apply to binary variables or factors, so the coefficients are still related to changes in levels.

classic: This method is similar to method = "smart", but treats all variables as continuous: it also scales the coefficient by the standard deviation of model's matrix' parameter of factors levels (transformed to integers) or binary predictors. Altough being inapropriate for these cases, this method is the one implemented by default in other softwares, such as sjstats::std_beta() or lm.beta::lm.beta().

References

Neter, J., Wasserman, W., & Kutner, M. H. (1989). Applied linear regression models.

Gelman, A. (2008). Scaling regression inputs by dividing by two standard deviations. Statistics in medicine, 27(15), 2865-2873.

Examples

library(parameters)
data(iris)

model <- lm(Sepal.Length ~ Species * Petal.Width, data = iris)
parameters_standardize(model, method = "refit")

#>                       Parameter Std_Coefficient
#> 1                   (Intercept)      0.05969491
#> 2             Speciesversicolor     -0.16597410
#> 3              Speciesvirginica      0.18985849
#> 4                   Petal.Width      0.85622714
#> 5 Speciesversicolor:Petal.Width      0.45674637
#> 6  Speciesvirginica:Petal.Width     -0.25713539

parameters_standardize(model, method = "refit", robust = TRUE)

#>                       Parameter Std_Coefficient
#> 1                   (Intercept)      0.17960924
#> 2             Speciesversicolor     -0.08429934
#> 3              Speciesvirginica      0.12439049
#> 4                   Petal.Width      0.93017273
#> 5 Speciesversicolor:Petal.Width      0.49619195
#> 6  Speciesvirginica:Petal.Width     -0.27934215

parameters_standardize(model, method = "2sd")

#>                       Parameter Std_Coefficient
#> 1                   (Intercept)      0.02046621
#> 2             Speciesversicolor     -0.05690368
#> 3              Speciesvirginica      0.06509237
#> 4                   Petal.Width      1.01050442
#> 5 Speciesversicolor:Petal.Width      0.53904414
#> 6  Speciesvirginica:Petal.Width     -0.30346673

parameters_standardize(model, method = "2sd", robust = TRUE)

#>                       Parameter Std_Coefficient
#> 1                   (Intercept)      0.09672589
#> 2             Speciesversicolor     -0.04539816
#> 3              Speciesvirginica      0.06698865
#> 4                   Petal.Width      0.93017273
#> 5 Speciesversicolor:Petal.Width      0.49619195
#> 6  Speciesvirginica:Petal.Width     -0.27934215

parameters_standardize(model, method = "smart")

#>                       Parameter Std_Coefficient
#> 1                   (Intercept)       0.0000000
#> 2             Speciesversicolor      -0.8846360
#> 3              Speciesvirginica       0.5944450
#> 4                   Petal.Width       0.8562271
#> 5 Speciesversicolor:Petal.Width       0.4567464
#> 6  Speciesvirginica:Petal.Width      -0.2571354

parameters_standardize(model, method = "smart", robust = TRUE)

#>                       Parameter Std_Coefficient
#> 1                   (Intercept)       0.0000000
#> 2             Speciesversicolor      -0.7058421
#> 3              Speciesvirginica       0.4743016
#> 4                   Petal.Width       0.9301727
#> 5 Speciesversicolor:Petal.Width       0.4961919
#> 6  Speciesvirginica:Petal.Width      -0.2793421

iris$binary <- ifelse(iris$Sepal.Width > 3, 1, 0)
model <- glm(binary ~ Species * Sepal.Length, data = iris, family = "binomial")
parameters_standardize(model, method = "refit")

#>                        Parameter Std_Coefficient
#> 1                    (Intercept)        6.786380
#> 2              Speciesversicolor       -9.544176
#> 3               Speciesvirginica       -8.145297
#> 4                   Sepal.Length        4.238719
#> 5 Speciesversicolor:Sepal.Length       -1.518858
#> 6  Speciesvirginica:Sepal.Length       -3.514712

parameters_standardize(model, method = "refit", robust = TRUE)

#>                        Parameter Std_Coefficient
#> 1                    (Intercept)        6.564564
#> 2              Speciesversicolor       -9.464693
#> 3               Speciesvirginica       -7.961370
#> 4                   Sepal.Length        5.312411
#> 5 Speciesversicolor:Sepal.Length       -1.903594
#> 6  Speciesvirginica:Sepal.Length       -4.405008

parameters_standardize(model, method = "smart")

#>                        Parameter Std_Coefficient
#> 1                    (Intercept)        0.000000
#> 2              Speciesversicolor        1.173802
#> 3               Speciesvirginica       16.656626
#> 4                   Sepal.Length        4.238719
#> 5 Speciesversicolor:Sepal.Length       -1.518858
#> 6  Speciesvirginica:Sepal.Length       -3.514712

parameters_standardize(model, method = "smart", robust = TRUE)

#>                        Parameter Std_Coefficient
#> 1                    (Intercept)        0.000000
#> 2              Speciesversicolor        1.173802
#> 3               Speciesvirginica       16.656626
#> 4                   Sepal.Length        5.312411
#> 5 Speciesversicolor:Sepal.Length       -1.903594
#> 6  Speciesvirginica:Sepal.Length       -4.405008

library(rstanarm)
model <- stan_glm(Sepal.Length ~ Species * Petal.Width, data = iris, iter = 500, refresh = 0)

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

parameters_standardize(model, method = "smart", centrality = "all")

#> Possible multicollinearity between Speciesversicolor:Petal.Width and Speciesversicolor (r = 0.71), Speciesversicolor:Petal.Width and Petal.Width (r = 0.85), Speciesvirginica:Petal.Width and Speciesversicolor:Petal.Width (r = 0.8). This might lead to inappropriate results. See 'Details' in '?rope'.

#>                       Parameter Std_Median   Std_Mean    Std_MAP
#> 1                   (Intercept)  0.0000000  0.0000000  0.0000000
#> 2             Speciesversicolor -0.8276283 -0.8274189 -0.8626834
#> 3              Speciesvirginica  0.5272624  0.5300223  0.5740361
#> 4                   Petal.Width  0.8449862  0.8544484  0.6795245
#> 5 Speciesversicolor:Petal.Width  0.4349971  0.4270448  0.5077906
#> 6  Speciesvirginica:Petal.Width -0.2107717 -0.2309449 -0.1742782

parameters_standardize(model, method = "smart", robust = TRUE, centrality = "all")

#> Possible multicollinearity between Speciesversicolor:Petal.Width and Speciesversicolor (r = 0.71), Speciesversicolor:Petal.Width and Petal.Width (r = 0.85), Speciesvirginica:Petal.Width and Speciesversicolor:Petal.Width (r = 0.8). This might lead to inappropriate results. See 'Details' in '?rope'.

#>                       Parameter Std_Median   Std_Mean    Std_MAP
#> 1                   (Intercept)  0.0000000  0.0000000  0.0000000
#> 2             Speciesversicolor -0.6603563 -0.6601892 -0.6883264
#> 3              Speciesvirginica  0.4206974  0.4228995  0.4580176
#> 4                   Petal.Width  0.9179610  0.9282403  0.7382097
#> 5 Speciesversicolor:Petal.Width  0.4725644  0.4639253  0.5516445
#> 6  Speciesvirginica:Petal.Width -0.2289744 -0.2508898 -0.1893293