Print model parameters

A print()-method for objects from model_parameters().

# S3 method for parameters_model
print(x, pretty_names = TRUE, split_components = TRUE, select = NULL, ...)

Arguments

x
An object returned by model_parameters().

pretty_names
Pretty parameters' names.

split_components
Logical, if TRUE (default), For models with multiple components (zero-inflation, smooth terms, ...), each component is printed in a separate table. If FALSE, model parameters are printed in a single table and a Component column is added to the output.

select
Character vector (or numeric index) of column names that should be printed. If NULL (default), all columns are printed.

...
Arguments passed to or from other methods.

Value

NULL

Examples

library(parameters)
if (require("glmmTMB")) {
  model <- glmmTMB(
    count ~ spp + mined + (1 | site),
    ziformula = ~mined,
    family = poisson(),
    data = Salamanders
  )
  mp <- model_parameters(model)

  print(mp, pretty_names = FALSE)

  print(mp, split_components = FALSE)

  print(mp, select = c("Parameter", "Coefficient", "CI_low", "CI_high"))
}

#> # Fixed Effects component
#> 
#> Parameter   | Coefficient |   SE |         95% CI |     z |  df |      p
#> ------------------------------------------------------------------------
#> (Intercept) |       -0.36 | 0.28 | [-0.90,  0.18] | -1.30 | 633 | 0.194 
#> sppPR       |       -1.27 | 0.24 | [-1.74, -0.80] | -5.27 | 633 | < .001
#> sppDM       |        0.27 | 0.14 | [ 0.00,  0.54] |  1.95 | 633 | 0.051 
#> sppEC-A     |       -0.57 | 0.21 | [-0.97, -0.16] | -2.75 | 633 | 0.006 
#> sppEC-L     |        0.67 | 0.13 | [ 0.41,  0.92] |  5.20 | 633 | < .001
#> sppDES-L    |        0.63 | 0.13 | [ 0.38,  0.87] |  4.96 | 633 | < .001
#> sppDF       |        0.12 | 0.15 | [-0.17,  0.40] |  0.78 | 633 | 0.435 
#> minedno     |        1.27 | 0.27 | [ 0.74,  1.80] |  4.72 | 633 | < .001
#> 
#> # Zero-Inflated component
#> 
#> Parameter   | Coefficient |   SE |         95% CI |     z |  df |      p
#> ------------------------------------------------------------------------
#> (Intercept) |        0.79 | 0.27 | [ 0.26,  1.32] |  2.90 | 633 | 0.004 
#> minedno     |       -1.84 | 0.31 | [-2.46, -1.23] | -5.87 | 633 | < .001
#> 
#> Parameter   | Coefficient |   SE |         95% CI |     z |  df |      p |     Component
#> ----------------------------------------------------------------------------------------
#> (Intercept) |       -0.36 | 0.28 | [-0.90,  0.18] | -1.30 | 633 | 0.194  |   conditional
#> spp [PR]    |       -1.27 | 0.24 | [-1.74, -0.80] | -5.27 | 633 | < .001 |   conditional
#> spp [DM]    |        0.27 | 0.14 | [ 0.00,  0.54] |  1.95 | 633 | 0.051  |   conditional
#> spp [EC-A]  |       -0.57 | 0.21 | [-0.97, -0.16] | -2.75 | 633 | 0.006  |   conditional
#> spp [EC-L]  |        0.67 | 0.13 | [ 0.41,  0.92] |  5.20 | 633 | < .001 |   conditional
#> spp [DES-L] |        0.63 | 0.13 | [ 0.38,  0.87] |  4.96 | 633 | < .001 |   conditional
#> spp [DF]    |        0.12 | 0.15 | [-0.17,  0.40] |  0.78 | 633 | 0.435  |   conditional
#> mined [no]  |        1.27 | 0.27 | [ 0.74,  1.80] |  4.72 | 633 | < .001 |   conditional
#> (Intercept) |        0.79 | 0.27 | [ 0.26,  1.32] |  2.90 | 633 | 0.004  | zero_inflated
#> mined [no]  |       -1.84 | 0.31 | [-2.46, -1.23] | -5.87 | 633 | < .001 | zero_inflated
#> # Fixed Effects component
#> 
#> Parameter   | Coefficient |         95% CI
#> ------------------------------------------
#> (Intercept) |       -0.36 | [-0.90,  0.18]
#> spp [PR]    |       -1.27 | [-1.74, -0.80]
#> spp [DM]    |        0.27 | [ 0.00,  0.54]
#> spp [EC-A]  |       -0.57 | [-0.97, -0.16]
#> spp [EC-L]  |        0.67 | [ 0.41,  0.92]
#> spp [DES-L] |        0.63 | [ 0.38,  0.87]
#> spp [DF]    |        0.12 | [-0.17,  0.40]
#> mined [no]  |        1.27 | [ 0.74,  1.80]
#> 
#> # Zero-Inflated component
#> 
#> Parameter   | Coefficient |         95% CI
#> ------------------------------------------
#> (Intercept) |        0.79 | [ 0.26,  1.32]
#> mined [no]  |       -1.84 | [-2.46, -1.23]
#>