Rescale design weights for multilevel analysis

Most functions to fit multilevel and mixed effects models only allow to specify frequency weights, but not design (i.e. sampling or probability) weights, which should be used when analyzing complex samples and survey data. rescale_weights() implements an algorithm proposed by Asparouhov (2006) and Carle (2009) to rescale design weights in survey data to account for the grouping structure of multilevel models, which then can be used for multilevel modelling.

rescale_weights(data, group, probability_weights, nest = FALSE)

Arguments

data
A data frame.

group
Variable names (as character vector), indicating the grouping structure (strata) of the survey data (level-2-cluster variable). It is also possible to create weights for multiple group variables; in such cases, each created weighting variable will be suffixed by the name of the group variable.

probability_weights
Variable indicating the probability (design or sampling) weights of the survey data (level-1-weight).

nest
Logical, if TRUE and group indicates at least two group variables, then groups are "nested", i.e. groups are now a combination from each group level of the variables in group.

Value

data, including the new weighting variables: pweights_a and pweights_b, which represent the rescaled design weights to use in multilevel models (use these variables for the weights argument).

Details

Rescaling is based on two methods: For pweights_a, the sample weights probability_weights are adjusted by a factor that represents the proportion of group size divided by the sum of sampling weights within each group. The adjustment factor for pweights_b is the sum of sample weights within each group divided by the sum of squared sample weights within each group (see Carle (2009), Appendix B).

Regarding the choice between scaling methods A and B, Carle suggests that "analysts who wish to discuss point estimates should report results based on weighting method A. For analysts more interested in residual between-group variance, method B may generally provide the least biased estimates". In general, it is recommended to fit a non-weighted model and weighted models with both scaling methods and when comparing the models, see whether the "inferential decisions converge", to gain confidence in the results.

Though the bias of scaled weights decreases with increasing group size, method A is preferred when insufficient or low group size is a concern.

The group ID and probably PSU may be used as random effects (e.g. nested design, or group and PSU as varying intercepts), depending on the survey design that should be mimicked.

References

Carle A.C. (2009). Fitting multilevel models in complex survey data with design weights: Recommendations. BMC Medical Research Methodology 9(49): 1-13

Asparouhov T. (2006). General Multi-Level Modeling with Sampling Weights. Communications in Statistics - Theory and Methods 35: 439-460

Examples

if (require("sjstats")) {
  data(nhanes_sample, package = "sjstats")
  head(rescale_weights(nhanes_sample, "SDMVSTRA", "WTINT2YR"))

  # also works with multiple group-variables...
  head(rescale_weights(nhanes_sample, c("SDMVSTRA", "SDMVPSU"), "WTINT2YR"))

  # or nested structures.
  x <- rescale_weights(
    data = nhanes_sample,
    group = c("SDMVSTRA", "SDMVPSU"),
    probability_weights = "WTINT2YR",
    nest = TRUE
  )
  head(x)
}

#> Loading required package: sjstats

#> 
#> Attaching package: 'sjstats'

#> The following object is masked from 'package:metafor':
#> 
#>     robust

#> The following object is masked from 'package:rstanarm':
#> 
#>     se

#> The following object is masked from 'package:psych':
#> 
#>     phi

#> The following object is masked from 'package:parameters':
#> 
#>     p_value

#>   total  age RIAGENDR RIDRETH1 SDMVPSU SDMVSTRA WTINT2YR pweights_a pweights_b
#> 1     1 2.20        1        3       2       31 97593.68  1.6532834  1.3583967
#> 2     7 2.08        2        3       1       29 39599.36  0.5492655  0.4875798
#> 3     3 1.48        2        1       2       42 26619.83  0.8341084  0.5891700
#> 4     4 1.32        2        4       2       33 34998.53  0.7937824  0.5941006
#> 5     1 2.00        2        1       1       41 14746.45  0.4285939  0.3068214
#> 6     6 2.20        2        4       1       38 28232.10  0.6215579  0.4759935

if (require("lme4") && require("sjstats")) {
  data(nhanes_sample, package = "sjstats")
  nhanes_sample <- rescale_weights(nhanes_sample, "SDMVSTRA", "WTINT2YR")
  glmer(
    total ~ factor(RIAGENDR) * (log(age) + factor(RIDRETH1)) + (1 | SDMVPSU),
    family = poisson(),
    data = nhanes_sample,
    weights = pweights_a
  )
}

#> Generalized linear mixed model fit by maximum likelihood (Laplace
#>   Approximation) [glmerMod]
#>  Family: poisson  ( log )
#> Formula: total ~ factor(RIAGENDR) * (log(age) + factor(RIDRETH1)) + (1 |  
#>     SDMVPSU)
#>    Data: nhanes_sample
#> Weights: pweights_a
#>       AIC       BIC    logLik  deviance  df.resid 
#>  78844.27  78920.47 -39409.14  78818.27      2582 
#> Random effects:
#>  Groups  Name        Std.Dev.
#>  SDMVPSU (Intercept) 0.1018  
#> Number of obs: 2595, groups:  SDMVPSU, 2
#> Fixed Effects:
#>                         (Intercept)                    factor(RIAGENDR)2  
#>                            2.491801                            -1.021308  
#>                            log(age)                    factor(RIDRETH1)2  
#>                            0.838726                            -0.088627  
#>                   factor(RIDRETH1)3                    factor(RIDRETH1)4  
#>                           -0.013333                             0.722511  
#>                   factor(RIDRETH1)5           factor(RIAGENDR)2:log(age)  
#>                           -0.106521                            -1.012695  
#> factor(RIAGENDR)2:factor(RIDRETH1)2  factor(RIAGENDR)2:factor(RIDRETH1)3  
#>                           -0.009086                             0.732985  
#> factor(RIAGENDR)2:factor(RIDRETH1)4  factor(RIAGENDR)2:factor(RIDRETH1)5  
#>                            0.275967                             0.542074