Parameters simulation

Compute simulated draws of parameters and their related indices such as Confidence Intervals (CI) and p-values. Simulating parameter draws can be seen as a (computationally faster) alternative to bootstrapping.

simulate_parameters(model, ...)

parameters_simulate(model, ...)

# S3 method for default
simulate_parameters(
  model,
  iterations = 1000,
  centrality = "median",
  ci = 0.95,
  ci_method = "quantile",
  test = "p-value",
  ...
)

Arguments

model
Statistical model (no Bayesian models).

...
Arguments passed to or from other methods.

iterations
The number of draws to simulate/bootstrap.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

ci
Value or vector of probability of the CI (between 0 and 1) to be estimated. Default to .89 (89%) for Bayesian models and .95 (95%) for frequentist models.

ci_method
The type of index used for Credible Interval. Can be "HDI" (default, see hdi) or "ETI" (see eti).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests. For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

Value

A data frame with simulated parameters.

Details

Technical Details

simulate_model() is a computationally faster alternative to bootstrap_model(). Simulated draws for coefficients are based on a multivariate normal distribution (MASS::mvrnorm()) with mean mu = coef(model) and variance Sigma = vcov(model).

Models with Zero-Inflation Component

For models from packages glmmTMB, pscl, GLMMadaptive and countreg, the component argument can be used to specify which parameters should be simulated. For all other models, parameters from the conditional component (fixed effects) are simulated. This may include smooth terms, but not random effects.

References

Gelman A, Hill J. Data analysis using regression and multilevel/hierarchical models. Cambridge; New York: Cambridge University Press 2007: 140-143

See also

Examples

library(parameters)
library(glmmTMB)

model <- lm(Sepal.Length ~ Species * Petal.Width + Petal.Length, data = iris)
simulate_parameters(model)

#>                       Parameter Coefficient     CI_low    CI_high           p
#> 1                   (Intercept)   3.5284178  3.2062455  3.8624354 0.000000000
#> 2             Speciesversicolor  -1.1531106 -1.8077936 -0.4978578 0.001998002
#> 3              Speciesvirginica  -2.2501388 -3.1073087 -1.3975844 0.000000000
#> 4                   Petal.Width   0.4117286 -0.4668826  1.3212562 0.355644356
#> 5                  Petal.Length   0.9418990  0.7957040  1.0889821 0.000000000
#> 6 Speciesversicolor:Petal.Width  -0.7330585 -1.8367166  0.2348297 0.159840160
#> 7  Speciesvirginica:Petal.Width  -0.3694773 -1.4165314  0.5600944 0.423576424

model <- glmmTMB(
  count ~ spp + mined + (1 | site),
  ziformula = ~mined,
  family = poisson(),
  data = Salamanders
)
simulate_parameters(model, centrality = "mean")

#>         Parameter Coefficient      CI_low    CI_high           p
#> 1     (Intercept)  -0.3618549 -0.92521263  0.1741156 0.185814186
#> 2           sppPR  -1.2602986 -1.75081929 -0.7693395 0.000000000
#> 3           sppDM   0.2678564  0.01355088  0.5480817 0.039960040
#> 4         sppEC-A  -0.5582286 -0.93701979 -0.1496809 0.005994006
#> 5         sppEC-L   0.6704696  0.43478923  0.9069486 0.000000000
#> 6        sppDES-L   0.6273870  0.38308796  0.8802319 0.000000000
#> 7           sppDF   0.1190586 -0.15457258  0.3958122 0.405594406
#> 8         minedno   1.2695018  0.73835218  1.7833134 0.000000000
#> 9  (Intercept)_zi   0.7932094  0.28927800  1.2821302 0.000000000
#> 10     minedno_zi  -1.8416667 -2.48984112 -1.2731391 0.000000000

simulate_parameters(model, ci = c(.8, .95), component = "zero_inflated")

#>     Parameter Coefficient CI     CI_low   CI_high           p
#> 1 (Intercept)   0.7975917 80  0.4290829  1.144443 0.003996004
#> 2 (Intercept)   0.7975917 95  0.2260494  1.322215 0.003996004
#> 3     minedno  -1.8386463 80 -2.2581487 -1.438269 0.000000000
#> 4     minedno  -1.8386463 95 -2.4424867 -1.235278 0.000000000