Robust estimation

standard_error_robust(), ci_robust() and p_value_robust() attempt to return indices based on robust estimation of the variance-covariance matrix, using the packages sandwich and clubSandwich.

Usage

standard_error_robust(
  model,
  vcov_estimation = "HC",
  vcov_type = NULL,
  vcov_args = NULL,
  component = "conditional",
  ...
)

p_value_robust(
  model,
  vcov_estimation = "HC",
  vcov_type = NULL,
  vcov_args = NULL,
  component = "conditional",
  method = NULL,
  ...
)

ci_robust(
  model,
  ci = 0.95,
  method = NULL,
  vcov_estimation = "HC",
  vcov_type = NULL,
  vcov_args = NULL,
  component = "conditional",
  ...
)

Arguments

model

A model.

vcov_estimation

String, indicating the suffix of the vcov*()-function from the sandwich or clubSandwich package, e.g. vcov_estimation = "CL" (which calls sandwich::vcovCL() to compute clustered covariance matrix estimators), or vcov_estimation = "HC" (which calls sandwich::vcovHC() to compute heteroskedasticity-consistent covariance matrix estimators).

vcov_type

Character vector, specifying the estimation type for the robust covariance matrix estimation (see sandwich::vcovHC() or clubSandwich::vcovCR() for details). Passed down as type argument to the related vcov*()-function from the sandwich or clubSandwich package and hence will be ignored if there is no type argument (e.g., sandwich::vcovHAC() will ignore that argument).

vcov_args

List of named vectors, used as additional arguments that are passed down to the sandwich-function specified in vcov_estimation.

component

Should all parameters or parameters for specific model components be returned?

...

Arguments passed to or from other methods. For standard_error(), if method = "robust", arguments vcov_estimation, vcov_type and vcov_args can be passed down to standard_error_robust().

method

Method for computing degrees of freedom for confidence intervals (CI) and the related p-values. Allowed are following options (which vary depending on the model class): "residual", "normal", "likelihood", "satterthwaite", "kenward", "wald", "profile", "boot", "uniroot", "ml1", "betwithin", "hdi", "quantile", "ci", "eti", "si", "bci", or "bcai". See section Confidence intervals and approximation of degrees of freedom in model_parameters() for further details.

ci

Confidence Interval (CI) level. Default to 0.95 (95%).

Value

A data frame.

Note

These functions rely on the sandwich or clubSandwich package (the latter if vcov_estimation = "CR" for cluster-robust standard errors) and will thus only work for those models supported by those packages.

See also

Examples

if (require("sandwich", quietly = TRUE)) {
  # robust standard errors, calling sandwich::vcovHC(type="HC3") by default
  model <- lm(Petal.Length ~ Sepal.Length * Species, data = iris)
  standard_error_robust(model)
}
#>                        Parameter         SE
#> 1                    (Intercept) 0.42486667
#> 2                   Sepal.Length 0.08504442
#> 3              Speciesversicolor 0.67252996
#> 4               Speciesvirginica 0.58942889
#> 5 Sepal.Length:Speciesversicolor 0.12045791
#> 6  Sepal.Length:Speciesvirginica 0.10558799
if (FALSE) {
if (require("clubSandwich", quietly = TRUE)) {
  # cluster-robust standard errors, using clubSandwich
  iris$cluster <- factor(rep(LETTERS[1:8], length.out = nrow(iris)))
  standard_error_robust(
    model,
    vcov_type = "CR2",
    vcov_args = list(cluster = iris$cluster)
  )
}
}