Get Standardized Model Parameters

Compute standardized model parameters (coefs).

standardize_parameters(model, method = "refit", robust = FALSE, ...)

Arguments

model
A statistical model.

method
The method used for standardizing the parameters. Can be "refit" (default).

robust
If TRUE, the standardization will be based on median and mad instead of mean and sd (default).

...
Arguments passed to or from other methods.

Details

Methods:

refit: This method is based on a complete model re-fit using the standardized data. It is the most accurate, especially for parameters related to interactions, but it is also the most computationnaly costly.

Examples

data <- iris

model <- lm(Sepal.Length ~ Species * Petal.Width, data = data)
standardize_parameters(model, method="refit")

#>                       Parameter Std_Estimate
#> 1                   (Intercept)   0.05969491
#> 2             Speciesversicolor  -0.16597410
#> 3              Speciesvirginica   0.18985849
#> 4                   Petal.Width   0.85622714
#> 5 Speciesversicolor:Petal.Width   0.45674637
#> 6  Speciesvirginica:Petal.Width  -0.25713539

standardize_parameters(model, method="refit", robust=TRUE)

#>                       Parameter Std_Estimate
#> 1                   (Intercept)   0.17960924
#> 2             Speciesversicolor  -0.08429934
#> 3              Speciesvirginica   0.12439049
#> 4                   Petal.Width   0.93017273
#> 5 Speciesversicolor:Petal.Width   0.49619195
#> 6  Speciesvirginica:Petal.Width  -0.27934215

standardize_parameters(model, method="full", robust=FALSE)

#>                       Parameter Std_Estimate
#> 1                   (Intercept)           NA
#> 2             Speciesversicolor   -0.8846360
#> 3              Speciesvirginica    0.5944450
#> 4                   Petal.Width    0.8562271
#> 5 Speciesversicolor:Petal.Width    0.4567464
#> 6  Speciesvirginica:Petal.Width   -0.2571354

standardize_parameters(model, method="full", robust=TRUE)

#>                       Parameter Std_Estimate
#> 1                   (Intercept)           NA
#> 2             Speciesversicolor   -0.7058421
#> 3              Speciesvirginica    0.4743016
#> 4                   Petal.Width    0.9301727
#> 5 Speciesversicolor:Petal.Width    0.4961919
#> 6  Speciesvirginica:Petal.Width   -0.2793421

data$binary <- ifelse(data$Sepal.Width > 3, 1, 0)
model <- glm(binary ~ Species * Sepal.Length, data = data, family="binomial")
standardize_parameters(model, method="refit")

#>                        Parameter Std_Estimate
#> 1                    (Intercept)     6.786380
#> 2              Speciesversicolor    -9.544176
#> 3               Speciesvirginica    -8.145297
#> 4                   Sepal.Length     4.238719
#> 5 Speciesversicolor:Sepal.Length    -1.518858
#> 6  Speciesvirginica:Sepal.Length    -3.514712

standardize_parameters(model, method="refit", robust=TRUE)

#>                        Parameter Std_Estimate
#> 1                    (Intercept)     6.564564
#> 2              Speciesversicolor    -9.464693
#> 3               Speciesvirginica    -7.961370
#> 4                   Sepal.Length     5.312411
#> 5 Speciesversicolor:Sepal.Length    -1.903594
#> 6  Speciesvirginica:Sepal.Length    -4.405008

standardize_parameters(model, method="full", robust=FALSE)

#>                        Parameter Std_Estimate
#> 1                    (Intercept)           NA
#> 2              Speciesversicolor    0.3115888
#> 3               Speciesvirginica    4.4215443
#> 4                   Sepal.Length    1.1251789
#> 5 Speciesversicolor:Sepal.Length   -0.4031848
#> 6  Speciesvirginica:Sepal.Length   -0.9329893

standardize_parameters(model, method="full", robust=TRUE)

#>                        Parameter Std_Estimate
#> 1                    (Intercept)           NA
#> 2              Speciesversicolor    0.4047491
#> 3               Speciesvirginica    5.7435181
#> 4                   Sepal.Length    1.8318193
#> 5 Speciesversicolor:Sepal.Length   -0.6563949
#> 6  Speciesvirginica:Sepal.Length   -1.5189298