Dimensionality Reduction via Regression (DRR)

Dimensionality Reduction via Regression (DRR) is a very recent technique extending PCA (Laparra et al., 2015). Starting from a rotated PCA, it predicts redundant information from the remaining components using non-linear regression. Some of the most notable advantages of performing PCR are avoidance of multicollinearity between predictors and overfitting mitigation. PCR tends to perform well when the first principal components are enough to explain most of the variation in the predictors. Requires the DRR package to be installed.

DRR(x, n = "all", ...)

Arguments

x
A dataframe or a statistical model.

n
Number of components to extract. If n = "all" =, then n is set as the number of variables minus 1 (ncol(x)-1). If n = "auto" (default) or n = NULL, the number of components is selected through n_factors. In parameters_reduction, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

...
Arguments passed to or from other methods.

References

Laparra, V., Malo, J., & Camps-Valls, G. (2015). Dimensionality reduction via regression in hyperspectral imagery. IEEE Journal of Selected Topics in Signal Processing, 9(6), 1026-1036.

Examples

# \donttest{
DRR(iris[, 1:4])

#> Warning: ndim < data dimensionality, the inverse functions will be incomplete!

#> Error in LU.dgC(a) : cs_lu(A) failed: near-singular A (or out of memory)
#> Error in .local(a, b = b, ...) : 
#>   cs_lu(A) failed: near-singular A (or out of memory)
#> Error in LU.dgC(a) : cs_lu(A) failed: near-singular A (or out of memory)
#> Error in LU.dgC(a) : cs_lu(A) failed: near-singular A (or out of memory)
#> Error in solve(Matrix(K + diag(lambda, N))) : 
#>   Lapack dgecon(): system computationally singular, reciprocal condition number = 1.1368e-17

#>             DRR1          DRR2         DRR3
#> 1   -2.684125626 -0.1771961172  0.047897729
#> 2   -2.714141687  0.2749174776  0.154849905
#> 3   -2.888990569 -0.0040678640 -0.028658841
#> 4   -2.745342856  0.3703268060 -0.095310310
#> 5   -2.728716537 -0.2503104473 -0.061408078
#> 6   -2.280859633 -0.1244200244 -0.112549993
#> 7   -2.820537751  0.0333042212 -0.278391666
#> 8   -2.626144973  0.0637116911  0.002886139
#> 9   -2.886382732  0.4326989279 -0.071386565
#> 10  -2.672755798  0.2727279447  0.141743575
#> 11  -2.506947091 -0.2535839248  0.134503750
#> 12  -2.612755231  0.2316845827 -0.149494081
#> 13  -2.786109266  0.2279035044  0.160564839
#> 14  -3.223803744  0.0592189447 -0.030053897
#> 15  -2.644750390 -0.9787299890  0.268814107
#> 16  -2.386039034 -0.8077500172 -0.179623542
#> 17  -2.623527875 -0.5797956699 -0.043586096
#> 18  -2.648296706 -0.1169968449 -0.013715141
#> 19  -2.199820324 -0.2155208363  0.186531723
#> 20  -2.587986400 -0.2318507065 -0.169602339
#> 21  -2.310256215  0.2050501869  0.236088176
#> 22  -2.543705229 -0.0900350830 -0.185212631
#> 23  -3.215939416 -0.5820780294 -0.198404695
#> 24  -2.302733182  0.5032221473 -0.098894917
#> 25  -2.355754049  0.5961851975 -0.206128691
#> 26  -2.506668907  0.5378597054  0.167816242
#> 27  -2.468820073  0.3080990013 -0.139977177
#> 28  -2.562319906 -0.0501773027  0.090797028
#> 29  -2.639534715 -0.1043964955  0.157582986
#> 30  -2.631989387  0.4155820571 -0.114549816
#> 31  -2.587398477  0.4868579535 -0.004696315
#> 32  -2.409932497  0.0949226872  0.151744173
#> 33  -2.648862334 -0.6193387157 -0.128425006
#> 34  -2.598736749 -0.8266780301 -0.044434214
#> 35  -2.636926878  0.3327640498  0.079543026
#> 36  -2.866241652 -0.1882997508  0.177110757
#> 37  -2.625238050 -0.3709603703  0.332233581
#> 38  -2.800684115 -0.2967196055 -0.062526326
#> 39  -2.980502044  0.2278922854 -0.094854996
#> 40  -2.590006314  0.0498116341  0.068273542
#> 41  -2.770102426 -0.2476104598 -0.052691343
#> 42  -2.849368705  0.8448643223  0.280297767
#> 43  -2.997406547  0.0633604969 -0.196886081
#> 44  -2.405614485  0.3215164809 -0.302758294
#> 45  -2.209489238  0.2171202581 -0.296205028
#> 46  -2.714451427  0.3476676994  0.033935021
#> 47  -2.538148259 -0.1533253402 -0.130146787
#> 48  -2.839462168  0.1454498287 -0.116446152
#> 49  -2.543085750 -0.2356121957  0.068388792
#> 50  -2.703359782  0.0061164562  0.073724865
#> 51   1.284825689 -0.7713617679  0.170728905
#> 52   0.932488532 -0.3990533106 -0.130236691
#> 53   1.464302322 -0.6116462327  0.192503218
#> 54   0.183317720  0.3701518084  0.088444029
#> 55   1.088103258 -0.1448435008  0.212722341
#> 56   0.641669084  0.2402804328 -0.067685816
#> 57   1.095060663 -0.3538391190 -0.325404335
#> 58  -0.749122670  0.2948241812 -0.131276734
#> 59   1.044131826 -0.2989271332  0.275959284
#> 60  -0.008745404  0.1528274560 -0.414151367
#> 61  -0.507840884  0.5256441977  0.168451009
#> 62   0.511698557 -0.1455683271 -0.166030347
#> 63   0.264976508  0.1440815321  0.607736870
#> 64   0.984934510  0.0506530514  0.015478025
#> 65  -0.173925372 -0.3955833238 -0.170457124
#> 66   0.927860781 -0.5486373152  0.139660598
#> 67   0.660283762  0.1840997362 -0.356837194
#> 68   0.236104993 -0.0908441529  0.201755623
#> 69   0.944733728  0.4642524185  0.534785582
#> 70   0.045226976  0.0434163706  0.112433683
#> 71   1.116283177  0.0136248503 -0.505664950
#> 72   0.357888418 -0.2769974387  0.203122391
#> 73   1.298183875  0.2399159318  0.399278128
#> 74   0.921728922  0.1002612326  0.194473824
#> 75   0.714853326 -0.2932835735  0.248219831
#> 76   0.900174373 -0.4149326190  0.172333102
#> 77   1.332024437 -0.3366160936  0.416132815
#> 78   1.557802155 -0.3807648524  0.151120211
#> 79   0.813290650  0.0550781454 -0.077513998
#> 80  -0.305583778 -0.3313179733  0.211529400
#> 81  -0.068126492  0.1040284087  0.103997364
#> 82  -0.189622472  0.0232294307  0.157619062
#> 83   0.136428712 -0.1726946587  0.101800117
#> 84   1.380026436  0.3227477421  0.069457061
#> 85   0.588006443  0.2783371551 -0.474773573
#> 86   0.806858313 -0.3044477038 -0.488262062
#> 87   1.220690882 -0.4864077954  0.054216821
#> 88   0.815095236  0.2643148881  0.603156506
#> 89   0.245957680 -0.1496336557 -0.246808688
#> 90   0.166413217  0.2136061616 -0.045515338
#> 91   0.464800288  0.3926620995 -0.017776730
#> 92   0.890815198 -0.0538603181 -0.056992629
#> 93   0.230548024 -0.0236119679  0.149721420
#> 94  -0.704531759  0.2886536541 -0.017009445
#> 95   0.356981495  0.1584021752 -0.085470610
#> 96   0.331934480 -0.1500451279 -0.128211922
#> 97   0.376215651 -0.0408925428 -0.129644923
#> 98   0.642576008 -0.1952537640  0.156265649
#> 99  -0.906469865  0.1268085128 -0.085975352
#> 100  0.299000842 -0.0351104629 -0.075544414
#> 101  2.531192728  0.2910772107 -0.559544233
#> 102  1.415235877  0.4725669880 -0.187681754
#> 103  2.616676016  0.0005878336  0.184839791
#> 104  1.971531053  0.1357465397  0.145505458
#> 105  2.350005920  0.1955999558 -0.010031249
#> 106  3.397038736  0.1875423861  0.030876112
#> 107  0.521232244  0.9488625815 -0.480802980
#> 108  2.932587069  0.2185872928  0.276950844
#> 109  2.321228817  0.3806056484  0.592541540
#> 110  2.916750967 -0.2191803158 -0.593077466
#> 111  1.661774154 -0.3540550218 -0.187868346
#> 112  1.803401953  0.1226139530  0.237795369
#> 113  2.165591796 -0.1711775815  0.234729924
#> 114  1.346163579  0.6828422105 -0.161560335
#> 115  1.585928224  0.4258336456 -0.501056949
#> 116  1.904456375 -0.1861939636 -0.257130064
#> 117  1.949689059 -0.0938792792  0.209270460
#> 118  3.487055364 -0.4479236915 -0.214366193
#> 119  3.795645422  0.3348675540  0.182535760
#> 120  1.300791713  0.6729494295  0.461718284
#> 121  2.427817913 -0.1705240820 -0.058506097
#> 122  1.199001105  0.5294586161 -0.426380483
#> 123  3.499920039  0.2644279248  0.233529270
#> 124  1.388766132  0.1051337968  0.113889922
#> 125  2.275430504 -0.2267890577 -0.065428705
#> 126  2.614090474 -0.2183446358  0.204391801
#> 127  1.258508161  0.0966935721 -0.038693759
#> 128  1.291132059  0.0296821882 -0.232854917
#> 129  2.123608723  0.2329525886  0.115522815
#> 130  2.388003016 -0.2841160614  0.586125774
#> 131  2.841672778  0.1363599607  0.411910715
#> 132  3.230673661 -0.6555982543  0.079913854
#> 133  2.159437642  0.2590867500  0.059866222
#> 134  1.444161242  0.0379699638  0.208139218
#> 135  1.781294810  0.4025808435  0.327849216
#> 136  3.076499932 -0.0297831355  0.107226373
#> 137  2.144243314 -0.1062552104 -0.449641129
#> 138  1.905098149 -0.1160449681  0.072888559
#> 139  1.169326339  0.0909016249 -0.297287641
#> 140  2.107611143 -0.3570452363  0.174973362
#> 141  2.314154705 -0.0513703000 -0.058980114
#> 142  1.922267801 -0.4704767659  0.013486002
#> 143  1.415235877  0.4725669880 -0.187681754
#> 144  2.563013375  0.0267212013 -0.163561703
#> 145  2.418746183 -0.1033764820 -0.311531525
#> 146  1.944109795 -0.2414272066  0.044552847
#> 147  1.527166615  0.2633314860  0.236143704
#> 148  1.764345717 -0.1791292058  0.031367095
#> 149  1.900941614 -0.1846528562 -0.501735319
#> 150  1.390188862  0.1832247029 -0.297374238

# }