Dimensionality Reduction via Regression (DRR)

Dimensionality Reduction via Regression (DRR) is a very recent technique extending PCA (Laparra et al., 2015). Starting from a rotated PCA, it predicts redundant information from the remaining components using non-linear regression. Some of the most notable advantages of performing PCR are avoidance of multicollinearity between predictors and overfitting mitigation. PCR tends to perform well when the first principal components are enough to explain most of the variation in the predictors. Requires the DRR package to be installed.

DRR(x, n = "all", ...)

Arguments

x
A data frame or a statistical model.

n
Number of components to extract. If n="all", then n is set as the number of variables minus 1 (ncol(x)-1). If n="auto" (default) or n=NULL, the number of components is selected through n_factors. In reduce_parameters, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

...
Arguments passed to or from other methods.

References

Laparra, V., Malo, J., & Camps-Valls, G. (2015). Dimensionality reduction via regression in hyperspectral imagery. IEEE Journal of Selected Topics in Signal Processing, 9(6), 1026-1036.

Examples

# \donttest{
DRR(iris[, 1:4])

#> Warning: ndim < data dimensionality, the inverse functions will be incomplete!

#> Error in solve(Matrix(K + diag(lambda, N))) : 
#>   Lapack routine dsytrf returned error code 21

#> Warning: Exact singularity detected during LU decomposition: U[i,i]=0, i=9.

#> Error in solve(Matrix(K + diag(lambda, N))) : 
#>   Lapack dgecon(): system computationally singular, reciprocal condition number = 0
#> Error in solve(Matrix(K + diag(lambda, N))) : 
#>   Lapack routine dsytrf returned error code 28
#> Error in LU.dgC(a) : cs_lu(A) failed: near-singular A (or out of memory)

#> Warning: Exact singularity detected during LU decomposition: U[i,i]=0, i=30.

#> Error in solve(Matrix(K + diag(lambda, N))) : 
#>   Lapack dgecon(): system computationally singular, reciprocal condition number = 0
#> Error in .local(a, b, ...) : 
#>   cs_lu(A) failed: near-singular A (or out of memory)
#> Error in LU.dgC(a) : cs_lu(A) failed: near-singular A (or out of memory)

#>             DRR1        DRR2          DRR3
#> 1   -2.684125626 -0.19645062  0.0262417047
#> 2   -2.714141687  0.25863542  0.1662626926
#> 3   -2.888990569 -0.01200858 -0.0207288617
#> 4   -2.745342856  0.35669336 -0.0580838109
#> 5   -2.728716537 -0.26529737 -0.0929933594
#> 6   -2.280859633 -0.20261143 -0.1508460859
#> 7   -2.820537751  0.02403717 -0.2799649528
#> 8   -2.626144973  0.03766065  0.0112422201
#> 9   -2.886382732  0.42475515 -0.0228614355
#> 10  -2.672755798  0.25224539  0.1494000420
#> 11  -2.506947091 -0.29686230  0.0969830086
#> 12  -2.612755231  0.20388905 -0.1329414193
#> 13  -2.786109266  0.21700394  0.1816270236
#> 14  -3.223803744  0.02405478 -0.0426426302
#> 15  -2.644750390 -1.00246016  0.1672843529
#> 16  -2.386039034 -0.87034666 -0.2454969231
#> 17  -2.623527875 -0.60618295 -0.1068642337
#> 18  -2.648296706 -0.14029897 -0.0264842386
#> 19  -2.199820324 -0.30348834  0.1289358283
#> 20  -2.587986400 -0.26302269 -0.1988373888
#> 21  -2.310256215  0.13096013  0.2648681635
#> 22  -2.543705229 -0.12765508 -0.1948174770
#> 23  -3.215939416 -0.61611922 -0.2528488973
#> 24  -2.302733182  0.42805921 -0.0118931331
#> 25  -2.355754049  0.52886919 -0.1208507606
#> 26  -2.506668907  0.49453772  0.2124379484
#> 27  -2.468820073  0.25876696 -0.0830612572
#> 28  -2.562319906 -0.08502949  0.0889481548
#> 29  -2.639534715 -0.12876469  0.1468187989
#> 30  -2.631989387  0.39027335 -0.0931068702
#> 31  -2.587398477  0.45560363  0.0249732926
#> 32  -2.409932497  0.03612553  0.1661946041
#> 33  -2.648862334 -0.64257300 -0.1944781326
#> 34  -2.598736749 -0.85636185 -0.1342368372
#> 35  -2.636926878  0.30807310  0.0929039134
#> 36  -2.866241652 -0.19639416  0.1548235371
#> 37  -2.625238050 -0.39712767  0.2869252381
#> 38  -2.800684115 -0.30685081 -0.1005713862
#> 39  -2.980502044  0.21770097 -0.0563816668
#> 40  -2.590006314  0.01892181  0.0804996547
#> 41  -2.770102426 -0.25948218 -0.0854682274
#> 42  -2.849368705  0.83647081  0.3053516739
#> 43  -2.997406547  0.05226504 -0.1670034405
#> 44  -2.405614485  0.26202913 -0.2392453768
#> 45  -2.209489238  0.13018232 -0.2723860264
#> 46  -2.714451427  0.33141420  0.0579193712
#> 47  -2.538148259 -0.19178564 -0.1507527070
#> 48  -2.839462168  0.13680753 -0.0953812026
#> 49  -2.543085750 -0.27332557  0.0358307461
#> 50  -2.703359782 -0.01118702  0.0662670627
#> 51   1.284825689 -0.72097135  0.3171640832
#> 52   0.932488532 -0.38583834 -0.0619837589
#> 53   1.464302322 -0.55496700  0.2504638526
#> 54   0.183317720  0.30895768  0.1380304587
#> 55   1.088103258 -0.11284694  0.2780009463
#> 56   0.641669084  0.21886525 -0.0122302447
#> 57   1.095060663 -0.32105623 -0.2642473391
#> 58  -0.749122670  0.29853413 -0.0282319305
#> 59   1.044131826 -0.27204329  0.3347704532
#> 60  -0.008745404  0.08534962 -0.3595669676
#> 61  -0.507840884  0.49625297  0.2347490911
#> 62   0.511698557 -0.18048447 -0.1194903232
#> 63   0.264976508  0.08805408  0.6124820346
#> 64   0.984934510  0.07036675  0.0511593357
#> 65  -0.173925372 -0.45944862 -0.1212603960
#> 66   0.927860781 -0.53599848  0.2363926579
#> 67   0.660283762  0.16473477 -0.3064990020
#> 68   0.236104993 -0.14883157  0.2173770256
#> 69   0.944733728  0.47899072  0.4417219006
#> 70   0.045226976 -0.02329435  0.1281787725
#> 71   1.116283177  0.04876119 -0.4397007874
#> 72   0.357888418 -0.32589081  0.2236614278
#> 73   1.298183875  0.29117718  0.3840873168
#> 74   0.921728922  0.11213649  0.2004036931
#> 75   0.714853326 -0.30647837  0.2906778814
#> 76   0.900174373 -0.40574116  0.2412719073
#> 77   1.332024437 -0.28341898  0.4433444387
#> 78   1.557802155 -0.32576872  0.1577745158
#> 79   0.813290650  0.05355623 -0.0641256732
#> 80  -0.305583778 -0.38542948  0.2683534593
#> 81  -0.068126492  0.03676821  0.1624920098
#> 82  -0.189622472 -0.03979185  0.2265812074
#> 83   0.136428712 -0.23627148  0.1097824105
#> 84   1.380026436  0.37798587  0.0379138078
#> 85   0.588006443  0.25117270 -0.4077568065
#> 86   0.806858313 -0.30675064 -0.4343773800
#> 87   1.220690882 -0.44096768  0.1283405480
#> 88   0.815095236  0.26301247  0.5703419788
#> 89   0.245957680 -0.20696754 -0.2238993666
#> 90   0.166413217  0.15149899 -0.0253960359
#> 91   0.464800288  0.35323842  0.0202571729
#> 92   0.890815198 -0.04583297 -0.0149332959
#> 93   0.230548024 -0.08196072  0.1558939383
#> 94  -0.704531759  0.28683980  0.0868578587
#> 95   0.356981495  0.10943378 -0.0522988622
#> 96   0.331934480 -0.20104634 -0.0863606057
#> 97   0.376215651 -0.08824926 -0.0704377202
#> 98   0.642576008 -0.21656972  0.1887853502
#> 99  -0.906469865  0.14459925  0.0083853786
#> 100  0.299000842 -0.08866244 -0.0443906508
#> 101  2.531192728  0.32069695 -0.6018313473
#> 102  1.415235877  0.52874699 -0.2207315785
#> 103  2.616676016  0.04134207  0.1347176874
#> 104  1.971531053  0.15665565 -0.0164109584
#> 105  2.350005920  0.20818253 -0.1227101894
#> 106  3.397038736  0.24570909  0.2575739988
#> 107  0.521232244  0.91488708 -0.4789962649
#> 108  2.932587069  0.29711418  0.3125198444
#> 109  2.321228817  0.39166168  0.3902393923
#> 110  2.916750967 -0.14189774 -0.4220870420
#> 111  1.661774154 -0.30467901 -0.1850105057
#> 112  1.803401953  0.15975663  0.0958455932
#> 113  2.165591796 -0.16184106  0.1425360399
#> 114  1.346163579  0.73672868 -0.2487443080
#> 115  1.585928224  0.47967246 -0.5530875235
#> 116  1.904456375 -0.15903715 -0.2993679398
#> 117  1.949689059 -0.07100959  0.1363429390
#> 118  3.487055364 -0.40539662 -0.1211737235
#> 119  3.795645422  0.31392354  0.4301530743
#> 120  1.300791713  0.72437380  0.3748718676
#> 121  2.427817913 -0.15205797 -0.1431362082
#> 122  1.199001105  0.57296300 -0.4912365754
#> 123  3.499920039  0.30452108  0.4703400131
#> 124  1.388766132  0.16065047  0.0986466546
#> 125  2.275430504 -0.21739120 -0.1726601348
#> 126  2.614090474 -0.17794180  0.1681616756
#> 127  1.258508161  0.14520124 -0.0062406623
#> 128  1.291132059  0.08049103 -0.1950977152
#> 129  2.123608723  0.24366494 -0.0785856151
#> 130  2.388003016 -0.26896821  0.4779955003
#> 131  2.841672778  0.20641295  0.4037454727
#> 132  3.230673661 -0.57721048  0.1040262139
#> 133  2.159437642  0.26857895 -0.1402087022
#> 134  1.444161242  0.09456002  0.1812245498
#> 135  1.781294810  0.44185413  0.2111373991
#> 136  3.076499932  0.05459312  0.2962091024
#> 137  2.144243314 -0.09631012 -0.5379129252
#> 138  1.905098149 -0.08895060  0.0115985918
#> 139  1.169326339  0.13157056 -0.2495520784
#> 140  2.107611143 -0.34562178  0.0615296678
#> 141  2.314154705 -0.04063285 -0.1285784696
#> 142  1.922267801 -0.44503799 -0.0412329823
#> 143  1.415235877  0.52874699 -0.2207315785
#> 144  2.563013375  0.06034415 -0.2027302523
#> 145  2.418746183 -0.08572029 -0.3786187804
#> 146  1.944109795 -0.21804360 -0.0009952005
#> 147  1.527166615  0.31925350  0.1732225152
#> 148  1.764345717 -0.13826094  0.0000604849
#> 149  1.900941614 -0.15715346 -0.5436996264
#> 150  1.390188862  0.23878395 -0.3251049122

# }