Independent Component Analysis (ICA)

Performs an Independent Component Analysis using the FastICA algorithm. Contrary to PCA, that attempts to find uncorrelated sources (through least squares minimization), ICA attempts to find independent sources, i.e., the source space that maximizes the "non-gaussianity" of all sources. Contrary to PCA, ICA does not rank each source, which makes it a poor tool for dimensionality reduction. Requires the fastICA package to be installed.

ICA(x, n = "all", ...)

Arguments

x
A dataframe or a statistical model.

n
Number of components to extract. If n = "all" =, then n is set as the number of variables minus 1 (ncol(x)-1). If n = "auto" (default) or n = NULL, the number of components is selected through n_factors. In parameters_reduction, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

...
Arguments passed to or from other methods.

Examples

ICA(iris[, 1:4])

#>              ICA1         ICA2         ICA3
#> 1   -0.4138620600  1.396800510  0.161998612
#> 2    0.7875328143  1.335031243  0.227028029
#> 3    0.3236624837  1.328553813 -0.456786052
#> 4    0.5974489177  1.199283276 -0.657550358
#> 5   -0.6390384903  1.368670212 -0.196349972
#> 6   -1.5234893418  1.250416332  0.013525814
#> 7   -0.2109410626  1.207960179 -1.131264300
#> 8   -0.1535930960  1.317313140 -0.005882607
#> 9    1.0774344356  1.189658035 -0.894663532
#> 10   0.6520669461  1.329502152  0.254041066
#> 11  -0.9046351097  1.435343003  0.645785144
#> 12  -0.1185005624  1.209695471 -0.532112410
#> 13   0.8853009700  1.349581114  0.151926022
#> 14   0.8986862192  1.354344357 -0.982791503
#> 15  -1.6962295334  1.702722364  1.396102157
#> 16  -2.7627066879  1.440669182  0.260935930
#> 17  -1.5775592218  1.449548997  0.145058164
#> 18  -0.5008426839  1.353333388 -0.009572739
#> 19  -1.2329073512  1.379848254  1.036792464
#> 20  -1.1952171003  1.305911080 -0.377481506
#> 21  -0.1697082833  1.333415811  0.915044648
#> 22  -1.0462337621  1.261657005 -0.437377631
#> 23  -0.6499582430  1.452133872 -1.051511052
#> 24  -0.1623235975  1.115475739 -0.228014251
#> 25  -0.0779481523  1.060345973 -0.630761672
#> 26   0.8037802225  1.264382161  0.407935211
#> 27  -0.3140368739  1.180595729 -0.381908396
#> 28  -0.4111321218  1.375934595  0.375788882
#> 29  -0.1886856297  1.424930808  0.520347196
#> 30   0.3642148937  1.179204314 -0.555435314
#> 31   0.5893913240  1.207334612 -0.197086730
#> 32  -0.3707044712  1.346047899  0.637668121
#> 33  -1.7399352706  1.424123435 -0.122691126
#> 34  -2.1087599221  1.517978185  0.366965457
#> 35   0.5650863222  1.286035030  0.082469715
#> 36   0.2777824182  1.465088732  0.316117109
#> 37  -0.4705296573  1.562252680  1.181575129
#> 38  -0.5412703345  1.383220083 -0.271451979
#> 39   0.8279530034  1.240228154 -0.973455671
#> 40  -0.1643806279  1.346230391  0.240790750
#> 41  -0.5035726221  1.374199303 -0.223363009
#> 42   2.3819325825  1.220169612 -0.116627080
#> 43   0.3560250792  1.241802060 -1.196806124
#> 44  -0.7239620838  1.094448438 -0.836726325
#> 45  -1.2281278442  1.063311294 -0.680585207
#> 46   0.7113397222  1.262646870 -0.191216680
#> 47  -1.0947190063  1.299595036 -0.238793243
#> 48   0.3479674855  1.249853396 -0.736342497
#> 49  -0.8938475779  1.406425752  0.399111786
#> 50   0.0688533961  1.366309353  0.138675707
#> 51  -0.5086242554 -0.220582527  2.039819983
#> 52  -0.5579146283 -0.337986823  0.453974662
#> 53  -0.3218184454 -0.393320185  1.667484326
#> 54   1.7692227151 -0.269474584 -0.253450378
#> 55   0.3886711583 -0.362000550  1.114465838
#> 56   0.6354151908 -0.456621148 -0.482895233
#> 57  -0.8430367425 -0.509150575 -0.141711448
#> 58   1.7643035257  0.036692392 -1.100270084
#> 59   0.3158809122 -0.245362102  1.592606671
#> 60   0.7572313682 -0.346762481 -1.578859619
#> 61   2.7244067823 -0.037104501 -0.472661996
#> 62  -0.0726014549 -0.334797485 -0.457392410
#> 63   2.2121908898  0.004726084  1.606305689
#> 64   0.2963554175 -0.483198645  0.154785445
#> 65   0.2885815305 -0.036702951 -0.545296029
#> 66  -0.2808501078 -0.158771734  1.510124400
#> 67   0.0003135506 -0.570898737 -1.296061745
#> 68   1.0674636129 -0.098956819  0.521699754
#> 69   1.8233000565 -0.403690855  1.077380212
#> 70   1.4469510368 -0.102266017  0.145898316
#> 71  -0.7243664309 -0.762323942 -1.392755441
#> 72   0.5246777135 -0.092036313  0.668213635
#> 73   1.1586905184 -0.571545409  0.857495540
#> 74   0.7062806274 -0.397051354  0.609603373
#> 75   0.2969035659 -0.153847105  1.197909219
#> 76  -0.0340986138 -0.188475938  1.375126269
#> 77   0.4703241268 -0.331348007  1.960291086
#> 78  -0.2247231974 -0.588659050  0.909787048
#> 79   0.1931273854 -0.456016686 -0.197693088
#> 80   1.2331102868  0.170037974  0.584000148
#> 81   1.6801850608 -0.082187055  0.043783273
#> 82   1.7536482147  0.011063233  0.248237711
#> 83   0.8664674250 -0.086324731  0.244323227
#> 84   0.6591795057 -0.799756711 -0.343212512
#> 85   0.0218886142 -0.628733239 -1.789408460
#> 86  -1.0736730492 -0.495549043 -0.927640571
#> 87  -0.3272783217 -0.351588355  1.239903786
#> 88   1.7369923404 -0.237269240  1.588404132
#> 89   0.1202049184 -0.284831828 -0.821386693
#> 90   1.2972947908 -0.267900678 -0.476800831
#> 91   1.2023813326 -0.422779268 -0.548437056
#> 92   0.0468739853 -0.432628526  0.075993306
#> 93   1.1159488571 -0.136894850  0.323115366
#> 94   1.9894799560  0.064822690 -0.741921500
#> 95   0.8416142748 -0.336975853 -0.519244101
#> 96   0.2099154805 -0.262230621 -0.436025072
#> 97   0.3588988187 -0.306484696 -0.495921196
#> 98   0.3184786295 -0.211681607  0.704562503
#> 99   1.3792314660  0.201196223 -0.791520683
#> 100  0.5813453108 -0.257488483 -0.351362882
#> 101 -1.4501352478 -1.547535833 -2.113333744
#> 102  0.4198126976 -0.987992579 -1.351273280
#> 103 -0.4941385904 -1.094907030  0.914247286
#> 104  0.0485150881 -1.047281126 -0.334101032
#> 105 -0.5299114933 -1.262094492 -0.704481122
#> 106 -0.4534539595 -1.298802937  1.917432461
#> 107  1.0816848362 -0.864188503 -2.807541817
#> 108  0.0352620598 -1.106590779  1.902450930
#> 109  0.9762557493 -1.034326266  1.033527129
#> 110 -2.2415974506 -1.334702881 -0.261182293
#> 111 -0.9225004598 -0.825104179 -0.354507260
#> 112  0.3821224466 -0.914055405  0.063000690
#> 113 -0.5158458749 -0.982526118  0.305759564
#> 114  0.8020300597 -1.012167692 -1.513284448
#> 115 -0.2510543841 -1.204541236 -2.320805260
#> 116 -1.1456198597 -1.083989128 -1.181660845
#> 117 -0.2225414077 -0.938876505  0.080453544
#> 118 -2.4254163422 -1.356840351  1.066249569
#> 119  0.3462055193 -1.509317241  2.169014760
#> 120  1.9124624702 -0.710441187  0.419618059
#> 121 -1.1454876388 -1.138535538 -0.079826408
#> 122  0.0914082352 -0.988940918 -2.062100397
#> 123  0.1081845268 -1.277775636  2.526144535
#> 124  0.4258207224 -0.700372869  0.119431034
#> 125 -1.1859152894 -1.108648842 -0.341705648
#> 126 -0.7023947047 -0.983797673  1.419401156
#> 127  0.1871268221 -0.678720001 -0.206034462
#> 128 -0.2604961003 -0.755846512 -0.708941360
#> 129  0.0127496467 -1.149552194 -0.490466501
#> 130 -0.0835404726 -0.798871002  2.051660486
#> 131  0.1464229262 -1.022361271  2.154994338
#> 132 -2.3135825681 -1.062722106  2.001388248
#> 133 -0.0742309773 -1.193019316 -0.662037851
#> 134  0.4778335720 -0.668750882  0.456703685
#> 135  1.1259045339 -0.933607999  0.193863336
#> 136 -0.7057900892 -1.107904100  1.985378555
#> 137 -1.6531884660 -1.304149094 -1.921905269
#> 138 -0.4477178379 -0.967006803 -0.277895040
#> 139 -0.2632260385 -0.734980597 -0.922731630
#> 140 -0.7761148389 -0.903038748  0.473640783
#> 141 -0.9884467070 -1.190840949 -0.600186160
#> 142 -0.9906284967 -0.840623494  0.229147344
#> 143  0.4198126976 -0.987992579 -1.351273280
#> 144 -1.1076651670 -1.267019121 -0.392265941
#> 145 -1.5338377851 -1.282517331 -1.027991051
#> 146 -0.7195720009 -0.949028115 -0.185407232
#> 147  0.8242854927 -0.795197064  0.138327048
#> 148 -0.4370550655 -0.876461251 -0.164039895
#> 149 -1.5824552503 -1.190032890 -1.931241101
#> 150 -0.2118860967 -0.913247347 -1.268054250