Independent Component Analysis (ICA)

Performs an Independent Component Analysis using the FastICA algorithm. Contrary to PCA, that attempts to find uncorrelated sources (through least squares minimization), ICA attempts to find independent sources, i.e., the source space that maximizes the "non-gaussianity" of all sources. Contrary to PCA, ICA does not rank each source, which makes it a poor tool for dimensionality reduction. Requires the fastICA package to be installed.

ICA(x, n = "all", ...)

Arguments

x
A dataframe or a statistical model.

n
Number of components to extract. If n = "all" =, then n is set as the number of variables minus 1 (ncol(x)-1). If n = "auto" (default) or n = NULL, the number of components is selected through n_factors. In parameters_reduction, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

...
Arguments passed to or from other methods.

Examples

ICA(iris[, 1:4])

#>             ICA1          ICA2         ICA3
#> 1   -1.396797178  0.4138647155  0.162020549
#> 2   -1.335033572 -0.7875322773  0.227016194
#> 3   -1.328558890 -0.3236430204 -0.456785074
#> 4   -1.199291259 -0.5974245493 -0.657557939
#> 5   -1.368668412  0.6390509375 -0.196322002
#> 6   -1.250408435  1.5234953645  0.013577421
#> 7   -1.207967574  0.2109786094 -1.131249403
#> 8   -1.317312398  0.1535999997 -0.005868481
#> 9   -1.189670249 -1.0774035420 -0.894684494
#> 10  -1.329503585 -0.6520671865  0.254032945
#> 11  -1.435333536  0.9046245487  0.645820977
#> 12  -1.209698851  0.1185215058 -0.532100062
#> 13  -1.349584506 -0.8852982763  0.151911585
#> 14  -1.354356317 -0.8986520395 -0.982806275
#> 15  -1.702703226  1.6961995367  1.396161941
#> 16  -1.440653091  2.7627068238  0.261023321
#> 17  -1.449539837  1.5775626165  0.145112762
#> 18  -1.353330896  0.5008498741 -0.009548715
#> 19  -1.379834179  1.2328856649  1.036836983
#> 20  -1.305907792  1.1952342483 -0.377438585
#> 21  -1.333408089  0.1696897355  0.915059341
#> 22  -1.261654928  1.0462523444 -0.437339171
#> 23  -1.452138422  0.6499948278 -1.051482153
#> 24  -1.115476616  0.1623356273 -0.228001394
#> 25  -1.060350299  0.0779710668 -0.630751568
#> 26  -1.264383219 -0.8037850629  0.407922396
#> 27  -1.180596983  0.3140535037 -0.381890846
#> 28  -1.375929676  0.4111287430  0.375810587
#> 29  -1.424925944  0.1886784936  0.520363101
#> 30  -1.179210339 -0.3641934594 -0.555436579
#> 31  -1.207339105 -0.5893796814 -0.197094028
#> 32  -1.346041226  0.3706936785  0.637688482
#> 33  -1.424115450  1.7399459589 -0.122632216
#> 34  -1.517964645  2.1087575141  0.367035295
#> 35  -1.286037303 -0.5650820279  0.082463680
#> 36  -1.465087785 -0.2777836861  0.316120381
#> 37  -1.562241421  0.4705048903  1.181599877
#> 38  -1.383219346  0.5412849385 -0.271426610
#> 39  -1.240239682 -0.8279196666 -0.973469337
#> 40  -1.346227746  0.1643808401  0.240805392
#> 41  -1.374198399  0.5035858466 -0.223338753
#> 42  -1.220182674 -2.3819231044 -0.116683982
#> 43  -1.241812846 -0.3559855418 -1.196806693
#> 44  -1.094451001  0.7239908831 -0.836698053
#> 45  -1.063310107  1.2281521547 -0.680543191
#> 46  -1.262651942 -0.7113279593 -0.191226944
#> 47  -1.299591223  1.0947322767 -0.238753155
#> 48  -1.249860692 -0.3479406739 -0.736342782
#> 49  -1.406418188  0.8938437082  0.399147104
#> 50  -1.366308666 -0.0688502497  0.138684033
#> 51   0.220600409  0.5085665697  2.039832432
#> 52   0.337993078  0.5579003116  0.453987599
#> 53   0.393334322  0.3217701994  1.667490302
#> 54   0.269463632 -1.7692170660 -0.253501450
#> 55   0.362006909 -0.3887039106  1.114452349
#> 56   0.456614279 -0.6354041382 -0.482916271
#> 57   0.509153827  0.8430380660 -0.141691887
#> 58  -0.036709662 -1.7642728308 -1.100318726
#> 59   0.245372415 -0.3159263248  1.592596074
#> 60   0.346746780 -0.7571893662 -1.578883211
#> 61   0.037087020 -2.7243938659 -0.472737811
#> 62   0.334794431  0.0726124236 -0.457392904
#> 63  -0.004725372 -2.2122354017  1.606244388
#> 64   0.483198288 -0.2963621815  0.154773608
#> 65   0.036697389 -0.2885665991 -0.545304305
#> 66   0.158784483  0.2808074250  1.510130997
#> 67   0.570889027 -0.0002805368 -1.296066030
#> 68   0.098955264 -1.0674785836  0.521669416
#> 69   0.403689595 -1.8233319933  1.077326634
#> 70   0.102259705 -1.4469556048  0.145857431
#> 71   0.762317217  0.7244011459 -1.392741066
#> 72   0.092038633 -0.5246967113  0.668198398
#> 73   0.571545903 -1.1587172177  0.857459132
#> 74   0.397052306 -0.7062995609  0.609580816
#> 75   0.153854559 -0.2969375666  1.197899834
#> 76   0.188486413  0.0340595222  1.375125802
#> 77   0.331360284 -0.4703801738  1.960275563
#> 78   0.588667015  0.2246949601  0.909788869
#> 79   0.456014217 -0.1931242374 -0.197701859
#> 80  -0.170039909 -1.2331256084  0.583967232
#> 81   0.082178785 -1.6801866946  0.043736072
#> 82  -0.011070348 -1.7536550401  0.248189172
#> 83   0.086322127 -0.8664746406  0.244298557
#> 84   0.799750767 -0.6591740817 -0.343236779
#> 85   0.628719723 -0.0218422177 -1.789413776
#> 86   0.495547588  1.0736962331 -0.927614514
#> 87   0.351599317  0.3272421445  1.239910226
#> 88   0.237272250 -1.7370375945  1.588354194
#> 89   0.284825060 -0.1201836018 -0.821392159
#> 90   0.267890468 -1.2972829412 -0.476838806
#> 91   0.422769007 -1.2023682894 -0.548473561
#> 92   0.432628855 -0.0468783061  0.075988766
#> 93   0.136891560 -1.1159585160  0.323083400
#> 94  -0.064838428 -1.9894590527 -0.741976175
#> 95   0.336967657 -0.8416016020 -0.519269960
#> 96   0.262226281 -0.2099047329 -0.436032856
#> 97   0.306479145 -0.3588866367 -0.495933443
#> 98   0.211685255 -0.3184992475  0.704552087
#> 99  -0.201209210 -1.3792084899 -0.791557417
#> 100  0.257482876 -0.5813368861 -0.351380930
#> 101  1.547527424  1.4501859236 -2.113305128
#> 102  0.987980309 -0.4197802871 -1.351292320
#> 103  1.094916406  0.4941076389  0.914252785
#> 104  1.047278375 -0.0485111836 -0.334110222
#> 105  1.262091926  0.5299245677 -0.704475883
#> 106  1.298819620  0.4533941501  1.917435304
#> 107  0.864161938 -1.0816114149 -2.807578280
#> 108  1.106604849 -0.0353204702  1.902441663
#> 109  1.034329012 -0.9762896974  1.033492313
#> 110  1.334712395  2.2415978618 -0.261130140
#> 111  0.825106244  0.9225060666 -0.354487863
#> 112  0.914053921 -0.3821288704  0.062983249
#> 113  0.982531048  0.5158323696  0.305766507
#> 114  1.012152252 -0.8019932808 -1.513314266
#> 115  1.204525137  0.2511125679 -2.320807321
#> 116  1.083986139  1.1456470756 -1.181637201
#> 117  0.938878246  0.2225343727  0.080452682
#> 118  1.356860749  2.4253798352  1.066306652
#> 119  1.509331716 -0.3462733811  2.168993854
#> 120  0.710434544 -1.9124777392  0.419559711
#> 121  1.138540801  1.1454840258 -0.079803176
#> 122  0.988925004 -0.0913561210 -2.062110338
#> 123  1.277794005 -0.1082611058  2.526131963
#> 124  0.700371585 -0.4258276173  0.119413981
#> 125  1.108652351  1.1859190903 -0.341681071
#> 126  0.983811899  0.7023503156  1.419413261
#> 127  0.678717500 -0.1871245824 -0.206044735
#> 128  0.755842535  0.2605118890 -0.708939799
#> 129  1.149548455 -0.0127419300 -0.490475465
#> 130  0.798886798  0.0834794998  2.051656817
#> 131  1.022376663 -0.1464879076  2.154982620
#> 132  1.062748936  2.3135216381  2.001444434
#> 133  1.193014737  0.0742432285 -0.662044730
#> 134  0.668751858 -0.4778496565  0.456685427
#> 135  0.933603690 -1.1259146858  0.193825125
#> 136  1.107922583  0.7057293726  1.985389824
#> 137  1.304143157  1.6532350795 -1.921869201
#> 138  0.967007013  0.4477205946 -0.277889870
#> 139  0.734975032  0.2632478615 -0.922729837
#> 140  0.903046267  0.7760970854  0.473655537
#> 141  1.190841511  0.9884572541 -0.600167674
#> 142  0.840630279  0.9906178415  0.229168513
#> 143  0.987980309 -0.4197802871 -1.351292320
#> 144  1.267021851  1.1076695594 -0.392244720
#> 145  1.282517479  1.5338597244 -1.027958130
#> 146  0.949030408  0.7195722852 -0.185394390
#> 147  0.795193882 -0.8242933966  0.138298237
#> 148  0.876462259  0.4370551287 -0.164034342
#> 149  1.190026522  1.5825027065 -1.931206139
#> 150  0.913238932  0.2119165821 -1.268055216