Independent Component Analysis (ICA)

Performs an Independent Component Analysis using the FastICA algorithm. Contrary to PCA, that attempts to find uncorrelated sources (through least squares minimization), ICA attempts to find independent sources, i.e., the source space that maximizes the "non-gaussianity" of all sources. Contrary to PCA, ICA does not rank each source, which makes it a poor tool for dimensionality reduction. Requires the fastICA package to be installed.

ICA(x, n = "all", ...)

Arguments

x
A dataframe or a statistical model.

n
Number of components to extract. If n = "all" =, then n is set as the number of variables minus 1 (ncol(x)-1). If n = "auto" (default) or n = NULL, the number of components is selected through n_factors. In parameters_reduction, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

...
Arguments passed to or from other methods.

Examples

ICA(iris[, 1:4])

#>              ICA1         ICA2         ICA3
#> 1   -0.4149826692  0.172196077  1.395247259
#> 2    0.7872961012  0.221417739  1.336112687
#> 3    0.3146033943 -0.456338543  1.330881652
#> 4    0.5860612045 -0.661085190  1.202953322
#> 5   -0.6447262422 -0.183296111  1.367812948
#> 6   -1.5260993801  0.037626288  1.246735205
#> 7   -0.2284288468 -1.124237703  1.211335087
#> 8   -0.1567304197  0.000677522  1.316956517
#> 9    1.0629490214 -0.904438022  1.195284602
#> 10   0.6522056288  0.250166803  1.330168608
#> 11  -0.8995220439  0.662440234  1.430959712
#> 12  -0.1282217433 -0.526333220  1.211231464
#> 13   0.8840467569  0.145104433  1.351153041
#> 14   0.8826878641 -0.989667915  1.359847142
#> 15  -1.6819303501  1.423887223  1.693875671
#> 16  -2.7624408297  0.301752135  1.433188685
#> 17  -1.5789216228  0.170538657  1.445285556
#> 18  -0.5040817670  0.001618058  1.352163172
#> 19  -1.2225598249  1.057479143  1.373341069
#> 20  -1.2030615991 -0.357412621  1.304353848
#> 21  -0.1609255297  0.921786637  1.329856272
#> 22  -1.0547655661 -0.419389777  1.260660339
#> 23  -0.6669422536 -1.037949041  1.454188368
#> 24  -0.1678730188 -0.222020529  1.115863391
#> 25  -0.0886050613 -0.626017497  1.062318700
#> 26   0.8060556647  0.401852880  1.264881244
#> 27  -0.3217230547 -0.373706606  1.181150755
#> 28  -0.4094286663  0.385859309  1.373653404
#> 29  -0.1852390961  0.527688264  1.422681570
#> 30   0.3542200764 -0.556022820  1.181968889
#> 31   0.5839636495 -0.200530632  1.209403200
#> 32  -0.3655348468  0.647086785  1.342963273
#> 33  -1.7447003540 -0.095167847  1.420394019
#> 34  -2.1073548171  0.399555339  1.511688140
#> 35   0.5631065309  0.079588785  1.287084520
#> 36   0.2784431601  0.317563524  1.464650438
#> 37  -0.4587944612  1.192989468  1.557059778
#> 38  -0.5479755865 -0.259609417  1.382853302
#> 39   0.8123483298 -0.979810793  1.245531611
#> 40  -0.1643819775  0.247568847  1.345000250
#> 41  -0.5096357699 -0.212045174  1.373757027
#> 42   2.3773635906 -0.143291449  1.226225307
#> 43   0.3375580679 -1.197012518  1.246750454
#> 44  -0.7373163814 -0.823463505  1.095592002
#> 45  -1.2393384543 -0.660903008  1.062719409
#> 46   0.7058485612 -0.196051604  1.264984867
#> 47  -1.1007569406 -0.220062695  1.297800348
#> 48   0.3354605129 -0.736457960  1.253200332
#> 49  -0.8918704860  0.415548909  1.402915979
#> 50   0.0674591505  0.142506477  1.365984684
#> 51  -0.4815825473  2.045470575 -0.228799530
#> 52  -0.5511834195  0.460000792 -0.340870838
#> 53  -0.2992238351  1.670145910 -0.399811947
#> 54   1.7664071719 -0.277316872 -0.264393111
#> 55   0.4039511071  1.108067475 -0.364902378
#> 56   0.6301562049 -0.492678995 -0.453446478
#> 57  -0.8436149690 -0.132525598 -0.510664412
#> 58   1.7497797568 -1.122934983  0.044669291
#> 59   0.3371026140  1.587513973 -0.250081050
#> 60   0.7374766632 -1.589744222 -0.339533123
#> 61   2.7181198441 -0.508096393 -0.028999837
#> 62  -0.0777520504 -0.457569830 -0.333395580
#> 63   2.2328418071  1.577474669  0.004468393
#> 64   0.2994668661  0.149251239 -0.483021388
#> 65   0.2815625857 -0.549118353 -0.034142498
#> 66  -0.2608494249  1.513081740 -0.164627386
#> 67  -0.0151806948 -1.297928081 -0.566439542
#> 68   1.0743748289  0.507459615 -0.098209553
#> 69   1.8380710055  1.052226767 -0.403052517
#> 70   1.4489579872  0.126756857 -0.099324773
#> 71  -0.7406062418 -1.385874163 -0.759254527
#> 72   0.5335221701  0.661026765 -0.093083608
#> 73   1.1710562976  0.840403136 -0.571730871
#> 74   0.7150601928  0.599008138 -0.397462635
#> 75   0.3127890477  1.193415600 -0.157255752
#> 76  -0.0158027361  1.374791277 -0.193280540
#> 77   0.4965066528  1.952863283 -0.336962269
#> 78  -0.2115182235  0.910579994 -0.592314596
#> 79   0.1916082047 -0.201761920 -0.454874033
#> 80   1.2401881535  0.568537706  0.170962820
#> 81   1.6807991153  0.021694487 -0.078340339
#> 82   1.7566926525  0.225500597  0.014381336
#> 83   0.8697655118  0.232759763 -0.085102552
#> 84   0.6565327328 -0.354506766 -0.797002490
#> 85   0.0001224209 -1.791710731 -0.622527007
#> 86  -1.0844665478 -0.915344250 -0.494911013
#> 87  -0.3103318407  1.242819445 -0.356624237
#> 88   1.7580169518  1.564901358 -0.238593600
#> 89   0.1101952582 -0.823859676 -0.281721016
#> 90   1.2916169100 -0.494518597 -0.263174268
#> 91   1.1961431197 -0.565453810 -0.418031111
#> 92   0.0488661745  0.073878469 -0.432774379
#> 93   1.1203662034  0.308132534 -0.135349561
#> 94   1.9795233299 -0.767442796  0.072103603
#> 95   0.8355862117 -0.531285181 -0.333186868
#> 96   0.2048483589 -0.439618426 -0.260230784
#> 97   0.3531443920 -0.501595581 -0.303924293
#> 98   0.3280921634  0.699632951 -0.213343217
#> 99   1.3683657303 -0.808646937  0.207194854
#> 100  0.5773339622 -0.359766626 -0.254896126
#> 101 -1.4738345609 -2.099692960 -1.543709839
#> 102  0.4045385549 -1.360023470 -0.982342217
#> 103 -0.4796708002  0.916780391 -1.099214305
#> 104  0.0466174051 -0.338331014 -1.046008563
#> 105 -0.5360661116 -0.701917484 -1.260923201
#> 106 -0.4254896646  1.918640370 -1.306458757
#> 107  1.0471576691 -2.824319079 -0.851960952
#> 108  0.0625407514  1.897985589 -1.113034351
#> 109  0.9920028188  1.017181551 -1.035546493
#> 110 -2.2416784137 -0.236701716 -1.339125569
#> 111 -0.9250971025 -0.345366527 -0.826073068
#> 112  0.3850403292  0.054868294 -0.913354996
#> 113 -0.5095383694  0.309018786 -0.984795152
#> 114  0.7846757161 -1.527062996 -1.005051292
#> 115 -0.2783520653 -2.321514423 -1.197153232
#> 116 -1.1583317169 -1.170448090 -1.082644239
#> 117 -0.2192864022  0.080078865 -0.939674088
#> 118 -2.4081958071  1.092918685 -1.366261328
#> 119  0.3778577875  2.159094832 -1.515933647
#> 120  1.9194019246  0.392303657 -0.707327921
#> 121 -1.1437672634 -0.068903835 -1.140975927
#> 122  0.0669363205 -2.066528816 -0.981629173
#> 123  0.1439536979  2.520083345 -1.286187367
#> 124  0.4289687422  0.111467357 -0.699764290
#> 125 -1.1876610829 -0.330131311 -1.110285796
#> 126 -0.6816097656  1.424975953 -0.990337056
#> 127  0.1860196084 -0.210796738 -0.677561014
#> 128 -0.2679135349 -0.708117880 -0.754023492
#> 129  0.0090636846 -0.494572881 -1.147826513
#> 130 -0.0550324294  2.049789899 -0.806112548
#> 131  0.1767741052  2.149355942 -1.029408951
#> 132 -2.2849174091  2.027541647 -1.075092925
#> 133 -0.0800354133 -0.665150899 -1.190910600
#> 134  0.4852820262  0.448144364 -0.669177783
#> 135  1.1305023050  0.176000996 -0.931587943
#> 136 -0.6773672216  1.990516119 -1.116395259
#> 137 -1.6749528367 -1.904803435 -1.301465979
#> 138 -0.4490299753 -0.275413322 -0.967108399
#> 139 -0.2734675377 -0.921781113 -0.732429637
#> 140 -0.7677906188  0.480537340 -0.906504410
#> 141 -0.9933736753 -0.591435548 -1.191119313
#> 142 -0.9856054966  0.239065581 -0.843759821
#> 143  0.4045385549 -1.360023470 -0.982342217
#> 144 -1.1097045842 -0.382251344 -1.268294835
#> 145 -1.5440574744 -1.012443384 -1.282622145
#> 146 -0.7197016893 -0.179344299 -0.950094295
#> 147  0.8278654669  0.124862971 -0.793704808
#> 148 -0.4371012800 -0.161392894 -0.876929499
#> 149 -1.6046133024 -1.914660557 -1.187150449
#> 150 -0.2261992978 -1.268356715 -0.909386133