Independent Component Analysis (ICA)

Performs an Independent Component Analysis using the FastICA algorithm. Contrary to PCA, that attempts to find uncorrelated sources (through least squares minimization), ICA attempts to find independent sources, i.e., the source space that maximizes the "non-gaussianity" of all sources. Contrary to PCA, ICA does not rank each source, which makes it a poor tool for dimensionality reduction. Requires the fastICA package to be installed.

ICA(x, n = "all", ...)

Arguments

x
A dataframe or a statistical model.

n
Number of components to extract. If n = "all" =, then n is set as the number of variables minus 1 (ncol(x)-1). If n = "auto" (default) or n = NULL, the number of components is selected through n_factors. In parameters_reduction, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

...
Arguments passed to or from other methods.

Examples

ICA(iris[, 1:4])

#>             ICA1         ICA2          ICA3
#> 1    0.162066529  1.396790196  0.4138702779
#> 2    0.226991385  1.335038454 -0.7875311523
#> 3   -0.456783024  1.328569532 -0.3236022273
#> 4   -0.657573829  1.199307989 -0.5973734742
#> 5   -0.196263377  1.368664640  0.6390770225
#> 6    0.013685587  1.250391883  1.5235079817
#> 7   -1.131218176  1.207983070  0.2110573042
#> 8   -0.005838874  1.317310842  0.1536144669
#> 9   -0.894728428  1.189695849 -1.0773387889
#> 10   0.254015922  1.329506590 -0.6520676911
#> 11   0.645896081  1.435313696  0.9046024079
#> 12  -0.532074183  1.209705932  0.1185654004
#> 13   0.151881324  1.349591615 -0.8852926305
#> 14  -0.982837235  1.354381383 -0.8985803996
#> 15   1.396287242  1.702663118  1.6961366564
#> 16   0.261206488  1.440619364  2.7627070990
#> 17   0.145227196  1.449520640  1.5775697250
#> 18  -0.009498363  1.353325675  0.5008649407
#> 19   1.036930291  1.379804680  1.2328402049
#> 20  -0.377348623  1.305900898  1.1952701847
#> 21   0.915090134  1.333391904  0.1696508565
#> 22  -0.437258562  1.261650572  1.0462912876
#> 23  -1.051421581  1.452147958  0.6500715046
#> 24  -0.227974448  1.115478454  0.1623608390
#> 25  -0.630730389  1.060359364  0.0780190929
#> 26   0.407895534  1.264385435 -0.8037952087
#> 27  -0.381854060  1.180599610  0.3140883566
#> 28   0.375856080  1.375919367  0.4111216574
#> 29   0.520396434  1.424915752  0.1886635333
#> 30  -0.555439231  1.179222964 -0.3641485348
#> 31  -0.197109325  1.207348519 -0.5893552794
#> 32   0.637731155  1.346027240  0.3706710537
#> 33  -0.122508744  1.424098715  1.7399683542
#> 34   0.367181673  1.517936267  2.1087524587
#> 35   0.082451030  1.286042069 -0.5650730282
#> 36   0.316127238  1.465085802 -0.2777863455
#> 37   1.181651743  1.562217826  0.4704529750
#> 38  -0.271373438  1.383217802  0.5413155444
#> 39  -0.973497978  1.240263843 -0.8278497931
#> 40   0.240836079  1.346222202  0.1643812822
#> 41  -0.223287914  1.374196504  0.5036135612
#> 42  -0.116803249  1.220210051 -2.3819032345
#> 43  -1.196807883  1.241835452 -0.3559026732
#> 44  -0.836638796  1.094456371  0.7240512422
#> 45  -0.680455124  1.063307619  1.2282031042
#> 46  -0.191248460  1.262662572 -0.7113033049
#> 47  -0.238669134  1.299583230  1.0947600861
#> 48  -0.736343379  1.249875982 -0.3478844784
#> 49   0.399221127  1.406402336  0.8938355926
#> 50   0.138701481  1.366307227 -0.0688436572
#> 51   2.039858520 -0.220637885  0.5084456605
#> 52   0.454014714 -0.338006188  0.5578703031
#> 53   1.667502823 -0.393363950  0.3216690764
#> 54  -0.253608496 -0.269440679 -1.7692052206
#> 55   1.114424075 -0.362020235 -0.3887725571
#> 56  -0.482960364 -0.456599884 -0.6353809695
#> 57  -0.141650888 -0.509160644  0.8430408389
#> 58  -1.100420676  0.036745856 -1.7642084902
#> 59   1.592573859 -0.245394028 -0.3160215075
#> 60  -1.578932655 -0.346713873 -0.7571013282
#> 61  -0.472896717 -0.037050381 -2.7243667862
#> 62  -0.457393940 -0.334788029  0.0726354141
#> 63   1.606115898  0.004723882 -2.2123286924
#> 64   0.154748800 -0.483197539 -0.2963763571
#> 65  -0.545321649 -0.036685734 -0.2885353024
#> 66   1.510144820 -0.158811201  0.2807179621
#> 67  -1.296075006 -0.570868678 -0.0002113393
#> 68   0.521605826 -0.098952005 -1.0675099592
#> 69   1.077214334 -0.403686952 -1.8233989268
#> 70   0.145771738 -0.102246477 -1.4469651752
#> 71  -1.392710934 -0.762303122  0.7244739075
#> 72   0.668166460 -0.092043496 -0.5247365290
#> 73   0.857382820 -0.571546937 -1.1587731750
#> 74   0.609533536 -0.397054300 -0.7063392427
#> 75   1.197880160 -0.153870179 -0.2970088307
#> 76   1.375124820 -0.188508366  0.0339775869
#> 77   1.960243022 -0.331386012 -0.4704976457
#> 78   0.909792684 -0.588683707  0.2246357757
#> 79  -0.197720241 -0.456009042 -0.1931176380
#> 80   0.583898239  0.170043967 -1.2331577195
#> 81   0.043637140 -0.082161452 -1.6801901146
#> 82   0.248087434  0.011085259 -1.7536693416
#> 83   0.244246847 -0.086316670 -0.8664897619
#> 84  -0.343287641 -0.799738309 -0.6591627100
#> 85  -1.789424912 -0.628691398 -0.0217449699
#> 86  -0.927559895 -0.495544541  1.0737448245
#> 87   1.239923721 -0.351622291  0.3271663174
#> 88   1.588249520 -0.237278555 -1.7371324414
#> 89  -0.821403613 -0.284810877 -0.1201389216
#> 90  -0.476918401 -0.267869070 -1.2972581007
#> 91  -0.548550071 -0.422747502 -1.2023409469
#> 92   0.075979249 -0.432629545 -0.0468873614
#> 93   0.323016397 -0.136884663 -1.1159787577
#> 94  -0.742090770  0.064871412 -1.9894152348
#> 95  -0.519324158 -0.336950480 -0.8415750372
#> 96  -0.436049171 -0.262217185 -0.2098822049
#> 97  -0.495959110 -0.306467511 -0.3588611020
#> 98   0.704530254 -0.211692900 -0.3185424613
#> 99  -0.791634408  0.201236427 -1.3791603293
#> 100 -0.351418755 -0.257471126 -0.5813192261
#> 101 -2.113245143 -1.547509801  1.4502921384
#> 102 -1.351332222 -0.987954594 -0.4197123522
#> 103  0.914264309 -1.094936057  0.4940427657
#> 104 -0.334129481 -1.047272610 -0.0485029977
#> 105 -0.704464899 -1.262086550  0.5299519724
#> 106  1.917441259 -1.298854583  0.4532687915
#> 107 -2.807654698 -0.864106264 -1.0814575205
#> 108  1.902422234 -1.106634335 -0.0354428953
#> 109  1.033419337 -1.034334766 -0.9763608478
#> 110 -0.261020825 -1.334732336  2.2415987202
#> 111 -0.354447205 -0.825110572  0.9225178175
#> 112  0.062946692 -0.914050812 -0.3821423319
#> 113  0.305781058 -0.982541380  0.5158040631
#> 114 -1.513376760 -1.012119895 -0.8019161887
#> 115 -2.320811634 -1.204491397  0.2512345218
#> 116 -1.181587638 -1.083979876  1.1457041189
#> 117  0.080450874 -0.938881896  0.2225196287
#> 118  1.066426297 -1.356903499  2.4253033132
#> 119  2.168950032 -1.509362052 -0.3464156148
#> 120  0.419437416 -0.710420620 -1.9125097366
#> 121 -0.079754482 -1.138551833  1.1454764522
#> 122 -2.062131169 -0.988891652 -0.0912468884
#> 123  2.526105607 -1.277832500 -0.1084216117
#> 124  0.119378239 -0.700368893 -0.4258420668
#> 125 -0.341629557 -1.108659706  1.1859270559
#> 126  1.419438630 -0.983841713  0.7022572766
#> 127 -0.206066264 -0.678712259 -0.1871198863
#> 128 -0.708936525 -0.755834200  0.2605449824
#> 129 -0.490494251 -1.149540619 -0.0127257539
#> 130  2.051649123 -0.798919900  0.0833517028
#> 131  2.154958052 -1.022408922 -0.1466241055
#> 132  2.001562194 -1.062805167  2.3133939257
#> 133 -0.662059143 -1.193005140  0.0742689089
#> 134  0.456647158 -0.668753903 -0.4778833669
#> 135  0.193745037 -0.933594657 -1.1259359594
#> 136  1.985413440 -1.107961319  0.7056021116
#> 137 -1.921793595 -1.304130717  1.6533327788
#> 138 -0.277879031 -0.967007452  0.4477263733
#> 139 -0.922726075 -0.734963370  0.2632936029
#> 140  0.473686461 -0.903062026  0.7760598741
#> 141 -0.600128925 -1.190842690  0.9884793603
#> 142  0.229212884 -0.840644501  0.9905955073
#> 143 -1.351332222 -0.987954594 -0.4197123522
#> 144 -0.392200239 -1.267027572  1.1076787652
#> 145 -1.027889124 -1.282517792  1.5339057071
#> 146 -0.185367471 -0.949035215  0.7195728809
#> 147  0.138237851 -0.795187212 -0.8243099596
#> 148 -0.164022702 -0.876464371  0.4370552618
#> 149 -1.931132852 -1.190013177  1.5826021723
#> 150 -1.268057235 -0.913221299  0.2119804801