Independent Component Analysis (ICA)

Performs an Independent Component Analysis using the FastICA algorithm. Contrary to PCA, that attempts to find uncorrelated sources (through least squares minimization), ICA attempts to find independent sources, i.e., the source space that maximizes the "non-gaussianity" of all sources. Contrary to PCA, ICA does not rank each source, which makes it a poor tool for dimensionality reduction. Requires the fastICA package to be installed.

ICA(x, n = "all", ...)

Arguments

x
A dataframe or a statistical model.

n
Number of components to extract. If n = "all" =, then n is set as the number of variables minus 1 (ncol(x)-1). If n = "auto" (default) or n = NULL, the number of components is selected through n_factors. In parameters_reduction, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

...
Arguments passed to or from other methods.

Examples

ICA(iris[, 1:4])

#>              ICA1         ICA2          ICA3
#> 1    0.4149829946 -1.395246755  0.1721993778
#> 2   -0.7872960287 -1.336113037  0.2214158867
#> 3   -0.3146004575 -1.330882395 -0.4563384009
#> 4   -0.5860574978 -1.202954497 -0.6610863382
#> 5    0.6447280515 -1.367812664 -0.1832918676
#> 6    1.5261001629 -1.246734010  0.0376341274
#> 7    0.2284345200 -1.211336159 -1.1242353952
#> 8    0.1567314119 -1.316956398  0.0006796435
#> 9   -1.0629442905 -1.195286407 -0.9044411975
#> 10  -0.6522056825 -1.330168823  0.2501655153
#> 11   0.8995203215 -1.430958299  0.6624456253
#> 12   0.1282248860 -1.211231950 -0.5263313361
#> 13  -0.8840463501 -1.351153549  0.1451021872
#> 14  -0.8826826482 -1.359848904 -0.9896701449
#> 15   1.6819256039 -1.693872823  1.4238962162
#> 16   2.7624406355 -1.433186264  0.3017654129
#> 17   1.5789219931 -1.445284174  0.1705469397
#> 18   0.5040827818 -1.352162789  0.0016216890
#> 19   1.2225563871 -1.373338974  1.0574858381
#> 20   1.2030640973 -1.304353335 -0.3574060854
#> 21   0.1609226334 -1.329855130  0.9217887904
#> 22   1.0547682935 -1.260660007 -0.4193839170
#> 23   0.6669477451 -1.454189014 -1.0379446084
#> 24   0.1678748016 -1.115863510 -0.2220185835
#> 25   0.0886085133 -1.062319327 -0.6260159453
#> 26  -0.8060564097 -1.264881408  0.4018508689
#> 27   0.3217255288 -1.181150926 -0.3737039370
#> 28   0.4094280712 -1.373652667  0.3858625629
#> 29   0.1852379378 -1.422680846  0.5276906231
#> 30  -0.3542168302 -1.181969772 -0.5560230101
#> 31  -0.5839618870 -1.209403863 -0.2005317648
#> 32   0.3655331230 -1.342962281  0.6470898191
#> 33   1.7447018300 -1.420392807 -0.0951588909
#> 34   2.1073542683 -1.511686105  0.3995659311
#> 35  -0.5631058953 -1.287084858  0.0795878266
#> 36  -0.2784433974 -1.464650297  0.3175639669
#> 37   0.4587905888 -1.557058110  1.1929931339
#> 38   0.5479777301 -1.382853176 -0.2596055671
#> 39  -0.8123432420 -1.245533310 -0.9798128515
#> 40   0.1643819460 -1.344999852  0.2475710318
#> 41   0.5096377052 -1.373756877 -0.2120414962
#> 42  -2.3773620576 -1.226227262 -0.1433001523
#> 43  -0.3375520599 -1.246752034 -1.1970125657
#> 44   0.7373206942 -1.095592356 -0.8234591716
#> 45   1.2393420543 -1.062719205 -0.6608965864
#> 46  -0.7058467757 -1.264985617 -0.1960531903
#> 47   1.1007588525 -1.297799760 -0.2200565997
#> 48  -0.3354564492 -1.253201400 -0.7364579922
#> 49   0.8918697874 -1.402914845  0.4155542370
#> 50  -0.0674587216 -1.365984577  0.1425077037
#> 51   0.4815737177  0.228802159  2.0454723597
#> 52   0.5511812151  0.340871764  0.4600027476
#> 53   0.2992164647  0.399814022  1.6701467337
#> 54  -1.7664061980  0.264391469 -0.2773246400
#> 55  -0.4039560723  0.364903300  1.1080653612
#> 56  -0.6301544624  0.453445456 -0.4926821644
#> 57   0.8436151441  0.510664903 -0.1325225927
#> 58  -1.7497749707 -0.044671857 -1.1229423389
#> 59  -0.3371095216  0.250082553  1.5875122698
#> 60  -0.7374701917  0.339530805 -1.5897477189
#> 61  -2.7181177156  0.028997221 -0.5081079284
#> 62   0.0777537350  0.333395133 -0.4575698703
#> 63  -2.2328484798 -0.004468334  1.5774652246
#> 64  -0.2994678627  0.483021327  0.1492494367
#> 65  -0.2815602876  0.034141677 -0.5491195821
#> 66   0.2608428969  0.164629258  1.5130826613
#> 67   0.0151857598  0.566438116 -1.2979286443
#> 68  -1.0743770507  0.098209303  0.5074549595
#> 69  -1.8380757632  0.403052293  1.0522185421
#> 70  -1.4489585987  0.099323818  0.1267506146
#> 71   0.7406115333  0.759253552 -1.3858718692
#> 72  -0.5335250380  0.093083936  0.6610244040
#> 73  -1.1710602863  0.571730917  0.8403975467
#> 74  -0.7150630282  0.397462758  0.5990046713
#> 75  -0.3127942172  0.157256837  1.1934141025
#> 76   0.0157967718  0.193282074  1.3747911301
#> 77  -0.4965151722  0.336964056  1.9528608085
#> 78   0.2115139222  0.592315764  0.9105802337
#> 79  -0.1916076945  0.454873664 -0.2017632342
#> 80  -1.2401904305 -0.170963128  0.5685326468
#> 81  -1.6807992663  0.078339093  0.0216872865
#> 82  -1.7566935959 -0.014382414  0.2254931783
#> 83  -0.8697665611  0.085102152  0.2327559882
#> 84  -0.6565318347  0.797001601 -0.3545104271
#> 85  -0.0001153085  0.622525023 -1.7917114208
#> 86   1.0844700477  0.494910819 -0.9153402084
#> 87   0.3103263115  0.356625847  1.2428203633
#> 88  -1.7580237549  0.238594005  1.5648936538
#> 89  -0.1101919844  0.281720021 -0.8238604545
#> 90  -1.2916150160  0.263172744 -0.4945243542
#> 91  -1.1961410424  0.418029581 -0.5654593349
#> 92  -0.0488668141  0.432774424  0.0738777827
#> 93  -1.1203676097  0.135349056  0.3081276422
#> 94  -1.9795200275 -0.072105948 -0.7674510936
#> 95  -0.8355842149  0.333185648 -0.5312890863
#> 96  -0.2048466950  0.260230142 -0.4396195806
#> 97  -0.3531424988  0.303923470 -0.5015974122
#> 98  -0.3280952855  0.213343744  0.6996313260
#> 99  -1.3683621514 -0.207196784 -0.8086524991
#> 100 -0.5773326322  0.254895292 -0.3597693521
#> 101  1.4738422807  1.543708628 -2.0996884313
#> 102 -0.4045335413  0.982340405 -1.3600262699
#> 103  0.4796660897  1.099215683  0.9167812041
#> 104 -0.0466167636  1.046008153 -0.3383323694
#> 105  0.5360681295  1.260922829 -0.7019166111
#> 106  0.4254805575  1.306461202  1.9186407239
#> 107 -1.0471463599  0.851957034 -2.8243244541
#> 108 -0.0625496249  1.113036405  1.8979840917
#> 109 -0.9920079061  1.035546870  1.0171762059
#> 110  2.2416784036  1.339127001 -0.2366937113
#> 111  0.9250979400  0.826073385 -0.3453635263
#> 112 -0.3850412515  0.913354766  0.0548656535
#> 113  0.5095363174  0.984795879  0.3090198516
#> 114 -0.7846700127  1.005049008 -1.5270674296
#> 115  0.2783609858  1.197150872 -2.3215145707
#> 116  1.1583358524  1.082643818 -1.1704443869
#> 117  0.2192853533  0.939674342  0.0800787531
#> 118  2.4081901494  1.366264358  1.0929273636
#> 119 -0.3778680706  1.515935753  2.1590915539
#> 120 -1.9194041194  0.707326906  0.3922947502
#> 121  1.1437666929  1.140976714 -0.0689002579
#> 122 -0.0669283163  0.981626834 -2.0665301861
#> 123 -0.1439653351  1.286190049  2.5200813121
#> 124 -0.4289697426  0.699764090  0.1114647663
#> 125  1.1876616412  1.110286328 -0.3301275141
#> 126  0.6816029866  0.990339148  1.4249777414
#> 127 -0.1860192282  0.677560640 -0.2107982760
#> 128  0.2679159620  0.754022910 -0.7081175815
#> 129 -0.0090624588  1.147825959 -0.4945741903
#> 130  0.0550231453  0.806114859  2.0497892392
#> 131 -0.1767839794  1.029411197  2.1493540545
#> 132  2.2849080160  1.075096896  2.0275501268
#> 133  0.0800373285  1.190909924 -0.6651518790
#> 134 -0.4852844284  0.669177913  0.4481415693
#> 135 -1.1305037538  0.931587284  0.1759951804
#> 136  0.6773579539  1.116397975  1.9905177501
#> 137  1.6749599148  1.301465136 -1.9047977873
#> 138  0.4490304102  0.967108434 -0.2754124923
#> 139  0.2734708854  0.732428823 -0.9217807667
#> 140  0.7677879002  0.906505522  0.4805395859
#> 141  0.9933752784  1.191119408 -0.5914326630
#> 142  0.9856038473  0.843760830  0.2390688176
#> 143 -0.4045335413  0.982340405 -1.3600262699
#> 144  1.1097052436  1.268295250 -0.3822480523
#> 145  1.5440607900  1.282622191 -1.0124382690
#> 146  0.7197017304  0.950094641 -0.1793423049
#> 147 -0.8278665950  0.793704321  0.1248585887
#> 148  0.4371013004  0.876929651 -0.1613920152
#> 149  1.6046205085  1.187149542 -1.9146550807
#> 150  0.2262039786  0.909384899 -1.2683567642