Independent Component Analysis (ICA)

Performs an Independent Component Analysis using the FastICA algorithm. Contrary to PCA, that attempts to find uncorrelated sources (through least squares minimization), ICA attempts to find independent sources, i.e., the source space that maximizes the "non-gaussianity" of all sources. Contrary to PCA, ICA does not rank each source, which makes it a poor tool for dimensionality reduction. Requires the fastICA package to be installed.

ICA(x, n = "all", ...)

Arguments

x
A data frame or a statistical model.

n
Number of components to extract. If n="all", then n is set as the number of variables minus 1 (ncol(x)-1). If n="auto" (default) or n=NULL, the number of components is selected through n_factors. In reduce_parameters, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

...
Arguments passed to or from other methods.

Examples

ICA(iris[, 1:4])

#>              ICA1         ICA2         ICA3
#> 1   -0.4138599956 -1.396803098 -0.161981563
#> 2    0.7875332316 -1.335029432 -0.227037225
#> 3    0.3236776089 -1.328549866  0.456786812
#> 4    0.5974678544 -1.199277073  0.657544466
#> 5   -0.6390288167 -1.368671610  0.196371708
#> 6   -1.5234846603 -1.250422469 -0.013485710
#> 7   -0.2109118846 -1.207954433  1.131275877
#> 8   -0.1535877306 -1.317313716  0.005893585
#> 9    1.0774584427 -1.189648542  0.894647243
#> 10   0.6520667595 -1.329501037 -0.254047376
#> 11  -0.9046433156 -1.435350359 -0.645757297
#> 12  -0.1184842868 -1.209692845  0.532122005
#> 13   0.8853030634 -1.349578478 -0.151937241
#> 14   0.8987127801 -1.354335062  0.982780023
#> 15  -1.6962528421 -1.702737236 -1.396055698
#> 16  -2.7627065802 -1.440681688 -0.260868018
#> 17  -1.5775565824 -1.449556114 -0.145015736
#> 18  -0.5008370957 -1.353335323  0.009591408
#> 19  -1.2329242024 -1.379859193 -1.036757867
#> 20  -1.1952037735 -1.305913636  0.377514860
#> 21  -0.1697226962 -1.333421813 -0.915033229
#> 22  -1.0462193208 -1.261658620  0.437407517
#> 23  -0.6499298122 -1.452130335  1.051533509
#> 24  -0.1623142487 -1.115475057  0.228024242
#> 25  -0.0779303452 -1.060342611  0.630769524
#> 26   0.8037764610 -1.264381340 -0.407945169
#> 27  -0.3140239503 -1.180594755  0.381922035
#> 28  -0.4111347467 -1.375938417 -0.375772014
#> 29  -0.1886911745 -1.424934587 -0.520334835
#> 30   0.3642315504 -1.179199633  0.555434331
#> 31   0.5894003716 -1.207331121  0.197081059
#> 32  -0.3707128574 -1.346053086 -0.637652298
#> 33  -1.7399269633 -1.424129639  0.122736905
#> 34  -2.1087617917 -1.517988707 -0.366911185
#> 35   0.5650896594 -1.286033262 -0.082474404
#> 36   0.2777814334 -1.465089467 -0.316114565
#> 37  -0.4705489027 -1.562261429 -1.181555897
#> 38  -0.5412589850 -1.383220655  0.271471693
#> 39   0.8279789094 -1.240219194  0.973445052
#> 40  -0.1643804623 -1.346232446 -0.240779372
#> 41  -0.5035623445 -1.374200005  0.223381859
#> 42   2.3819399471 -1.220159461  0.116582861
#> 43   0.3560558038 -1.241793677  1.196805682
#> 44  -0.7239397032 -1.094446446  0.836748294
#> 45  -1.2281089515 -1.063312216  0.680617858
#> 46   0.7113488632 -1.262642928  0.191208703
#> 47  -1.0947086929 -1.299598000  0.238824395
#> 48   0.3479883211 -1.249847725  0.736342275
#> 49  -0.8938505839 -1.406431629 -0.399084340
#> 50   0.0688558416 -1.366309886 -0.138669237
#> 51  -0.5086690824  0.220568630 -2.039810308
#> 52  -0.5579257536  0.337981961 -0.453964609
#> 53  -0.3218559370  0.393309198 -1.667479681
#> 54   1.7692271039  0.269483095  0.253410689
#> 55   0.3886457064  0.361995608 -1.114476319
#> 56   0.6354237793  0.456626486  0.482878884
#> 57  -0.8430357138  0.509148048  0.141726648
#> 58   1.7643273776 -0.036678971  1.100232283
#> 59   0.3158456221  0.245354087 -1.592614905
#> 60   0.7572640071  0.346774683  1.578841285
#> 61   2.7244168181  0.037118087  0.472603079
#> 62  -0.0725929313  0.334799859  0.457392025
#> 63   2.2121562986 -0.004726638 -1.606353325
#> 64   0.2963501608  0.483198922 -0.154794643
#> 65   0.2885931334  0.036707273  0.545289597
#> 66  -0.2808832763  0.158761827 -1.510119272
#> 67   0.0003392052  0.570906283  1.296058414
#> 68   1.0674519785  0.098958027 -0.521723330
#> 69   1.8232752375  0.403691833 -1.077421846
#> 70   1.4469474863  0.102270921 -0.145930088
#> 71  -0.7243394540  0.762329170  1.392766610
#> 72   0.5246629500  0.092034510 -0.668225475
#> 73   1.1586697696  0.571545025 -0.857523832
#> 74   0.7062659138  0.397050614 -0.609620902
#> 75   0.2968771438  0.153841313 -1.197916511
#> 76  -0.0341289918  0.188467798 -1.375126631
#> 77   0.4702805725  0.331338466 -1.960303148
#> 78  -0.2247451406  0.588652861 -0.909785633
#> 79   0.1931298313  0.456018605  0.197686273
#> 80   1.2330983798 -0.170036470 -0.584025727
#> 81   1.6801837902  0.082193481 -0.043819952
#> 82   1.7536429098 -0.011057705 -0.248275431
#> 83   0.8664618173  0.086326754 -0.244342399
#> 84   0.6591837200  0.799761330  0.343193654
#> 85   0.0219246685  0.628743743  1.789404328
#> 86  -1.0736550326  0.495550173  0.927660820
#> 87  -0.3273064348  0.351579835 -1.239898781
#> 88   1.7369571727  0.237266901 -1.588442938
#> 89   0.1202214833  0.284837088  0.821382445
#> 90   1.2973039983  0.267908613  0.476771319
#> 91   1.2023914677  0.422787242  0.548408689
#> 92   0.0468706275  0.432628270 -0.075996834
#> 93   1.1159413506  0.136897407 -0.323140207
#> 94   1.9894961987 -0.064810459  0.741879011
#> 95   0.8416241221  0.336982223  0.519224006
#> 96   0.2099238322  0.262233994  0.436019023
#> 97   0.3589082849  0.306489010  0.495911679
#> 98   0.3184626071  0.211678772 -0.704570598
#> 99   1.3792493199 -0.201186131  0.791492137
#> 100  0.5813518572  0.257492840  0.351348858
#> 101 -1.4500958677  1.547542369  2.113355980
#> 102  0.4198378830  0.988002114  1.351258483
#> 103 -0.4941626428  1.094899742 -0.914243013
#> 104  0.0485181219  1.047283264  0.334093890
#> 105 -0.5299013335  1.262096485  0.704485192
#> 106 -0.4535004375  1.298789972 -1.917430251
#> 107  1.0817418905  0.864209148  2.807513480
#> 108  0.0352166690  1.106579844 -1.902458131
#> 109  0.9762293674  1.034324132 -1.033554184
#> 110 -2.2415971304  1.334695488  0.261222821
#> 111 -0.9224961025  0.825102574  0.354522333
#> 112  0.3821174541  0.914056558 -0.063014244
#> 113 -0.5158563700  0.982522287 -0.305754169
#> 114  0.8020586396  1.012179690  1.513261275
#> 115 -0.2510091702  1.204553748  2.320803657
#> 116 -1.1455987101  1.083991451  1.181679218
#> 117 -0.2225468747  0.938875152 -0.080454215
#> 118 -2.4254447109  1.356824499 -1.066205209
#> 119  0.3461527834  1.509305991 -2.169031005
#> 120  1.9124506034  0.710446350 -0.419663400
#> 121 -1.1454904463  1.138531447  0.079844461
#> 122  0.0914487324  0.988953285  2.062092670
#> 123  0.1081250171  1.277761361 -2.526154303
#> 124  0.4258153638  0.700373867 -0.119444286
#> 125 -1.1859123356  1.108646116  0.341724746
#> 126 -0.7024291993  0.983786617 -1.419391749
#> 127  0.1871285621  0.678721945  0.206026479
#> 128 -0.2604838312  0.755849603  0.708942573
#> 129  0.0127556428  1.149555100  0.490459534
#> 130 -0.0835878546  0.798858726 -2.051663336
#> 131  0.1463724290  1.022349308 -2.155003444
#> 132 -2.3136299156  1.062701255 -2.001344585
#> 133 -0.0742214573  1.193022875  0.662032506
#> 134  0.4778210723  0.668750124 -0.456717873
#> 135  1.1258966439  0.933611348 -0.193893029
#> 136 -0.7058372719  1.107889736 -1.985369797
#> 137 -1.6531522425  1.304153708  1.921933297
#> 138 -0.4477156959  0.967006640  0.277899057
#> 139 -0.2632090801  0.734984922  0.922733023
#> 140 -0.7761286349  0.903032904 -0.473629317
#> 141 -0.9884385108  1.190840512  0.600200525
#> 142 -0.9906367766  0.840618220 -0.229130893
#> 143  0.4198378830  0.988002114  1.351258483
#> 144 -1.1076617536  1.267017000  0.392282431
#> 145 -1.5338207359  1.282517215  1.028016633
#> 146 -0.7195717800  0.949026333  0.185417212
#> 147  0.8242793498  0.795199536 -0.138349437
#> 148 -0.4370550165  0.876460468  0.164044209
#> 149 -1.5824183718  1.190037840  1.931268268
#> 150 -0.2118624069  0.913253886  1.268053499