Independent Component Analysis (ICA)

Performs an Independent Component Analysis using the FastICA algorithm. Contrary to PCA, that attempts to find uncorrelated sources (through least squares minimization), ICA attempts to find independent sources, i.e., the source space that maximizes the "non-gaussianity" of all sources. Contrary to PCA, ICA does not rank each source, which makes it a poor tool for dimensionality reduction. Requires the fastICA package to be installed.

ICA(x, n = "all", ...)

Arguments

x
A dataframe or a statistical model.

n
Number of components to extract. If n="all", then n is set as the number of variables minus 1 (ncol(x)-1). If n="auto" (default) or n=NULL, the number of components is selected through n_factors. In reduce_parameters, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

...
Arguments passed to or from other methods.

Examples

ICA(iris[, 1:4])

#>             ICA1         ICA2         ICA3
#> 1    1.396790388  0.162065264 -0.413870125
#> 2    1.335038320  0.226992068  0.787531183
#> 3    1.328569239 -0.456783080  0.323603349
#> 4    1.199307529 -0.657573392  0.597374879
#> 5    1.368664744 -0.196264989 -0.639076305
#> 6    1.250392338  0.013682613 -1.523507635
#> 7    1.207982644 -1.131219035 -0.211055140
#> 8    1.317310885 -0.005839688 -0.153614069
#> 9    1.189695145 -0.894727220  1.077340569
#> 10   1.329506508  0.254016390  0.652067677
#> 11   1.435314241  0.645894016 -0.904603017
#> 12   1.209705738 -0.532074894 -0.118564194
#> 13   1.349591420  0.151882156  0.885292786
#> 14   1.354380694 -0.982836384  0.898582369
#> 15   1.702664221  1.396283797 -1.696138386
#> 16   1.440620291  0.261201452 -2.762707092
#> 17   1.449521168  0.145224050 -1.577569530
#> 18   1.353325818 -0.009499748 -0.500864527
#> 19   1.379805491  1.036927726 -1.232841455
#> 20   1.305901088 -0.377351097 -1.195269197
#> 21   1.333392349  0.915089287 -0.169651926
#> 22   1.261650692 -0.437260779 -1.046290217
#> 23   1.452147696 -1.051423246 -0.650069396
#> 24   1.115478404 -0.227975189 -0.162360146
#> 25   1.060359115 -0.630730972 -0.078017772
#> 26   1.264385374  0.407896273  0.803794930
#> 27   1.180599537 -0.381855071 -0.314087398
#> 28   1.375919650  0.375854829 -0.411121852
#> 29   1.424916032  0.520395518 -0.188663945
#> 30   1.179222616 -0.555439158  0.364149770
#> 31   1.207348260 -0.197108904  0.589355950
#> 32   1.346027624  0.637729982 -0.370671676
#> 33   1.424099175 -0.122512139 -1.739967739
#> 34   1.517937048  0.367177648 -2.108752598
#> 35   1.286041938  0.082451378  0.565073276
#> 36   1.465085856  0.316127049  0.277786272
#> 37   1.562218475  1.181650317 -0.470454403
#> 38   1.383217844 -0.271374900 -0.541314703
#> 39   1.240263178 -0.973497191  0.827851714
#> 40   1.346222355  0.240835235 -0.164381270
#> 41   1.374196556 -0.223289312 -0.503612799
#> 42   1.220209298 -0.116799969  2.381903781
#> 43   1.241834830 -1.196807851  0.355904952
#> 44   1.094456223 -0.836640425 -0.724049583
#> 45   1.063307688 -0.680457545 -1.228201703
#> 46   1.262662280 -0.191247868  0.711303983
#> 47   1.299583450 -0.238671444 -1.094759322
#> 48   1.249875562 -0.736343363  0.347886024
#> 49   1.406402772  0.399219092 -0.893835816
#> 50   1.366307267  0.138701001  0.068843838
#> 51  -0.220636854  2.039857803 -0.508448985
#> 52  -0.338005828  0.454013969 -0.557871128
#> 53  -0.393363135  1.667502479 -0.321671857
#> 54  -0.269441310 -0.253605553  1.769205546
#> 55  -0.362019869  1.114424853  0.388770670
#> 56  -0.456600279 -0.482959152  0.635381607
#> 57  -0.509160457 -0.141652015 -0.843040763
#> 58   0.036744861 -1.100417873  1.764210259
#> 59  -0.245393433  1.592574470  0.316018890
#> 60  -0.346714778 -1.578931296  0.757103749
#> 61  -0.037051388 -0.472892348  2.724367531
#> 62  -0.334788205 -0.457393911 -0.072634782
#> 63   0.004723923  1.606119431  2.212326127
#> 64  -0.483197560  0.154749482  0.296375967
#> 65  -0.036686055 -0.545321172  0.288536163
#> 66  -0.158810467  1.510144440 -0.280720422
#> 67  -0.570869237 -1.296074759  0.000213242
#> 68  -0.098952095  0.521607575  1.067509097
#> 69  -0.403687025  1.077217422  1.823397087
#> 70  -0.102246841  0.145774094  1.446964912
#> 71  -0.762303509 -1.392711762 -0.724471907
#> 72  -0.092043362  0.668167338  0.524735434
#> 73  -0.571546908  0.857384918  1.158771637
#> 74  -0.397054246  0.609534836  0.706338152
#> 75  -0.153869750  1.197880701  0.297006871
#> 76  -0.188507762  1.375124847 -0.033979840
#> 77  -0.331385305  1.960243917  0.470494416
#> 78  -0.588683248  0.909792579 -0.224637403
#> 79  -0.456009184 -0.197719735  0.193117819
#> 80   0.170043855  0.583900136  1.233156837
#> 81  -0.082161929  0.043639860  1.680190021
#> 82   0.011084849  0.248090231  1.753668948
#> 83  -0.086316820  0.244248269  0.866489346
#> 84  -0.799738651 -0.343286242  0.659163023
#> 85  -0.628692177 -1.789424606  0.021747644
#> 86  -0.495544625 -0.927561397 -1.073743488
#> 87  -0.351621659  1.239923350 -0.327168402
#> 88  -0.237278382  1.588252398  1.737129834
#> 89  -0.284811267 -0.821403298  0.120140150
#> 90  -0.267869658 -0.476916213  1.297258784
#> 91  -0.422748093 -0.548547967  1.202341699
#> 92  -0.432629526  0.075979511  0.046887112
#> 93  -0.136884853  0.323018240  1.115978201
#> 94   0.064870505 -0.742087620  1.989416440
#> 95  -0.336950952 -0.519322668  0.841575768
#> 96  -0.262217435 -0.436048722  0.209882824
#> 97  -0.306467831 -0.495958404  0.358861804
#> 98  -0.211692689  0.704530854  0.318541273
#> 99   0.201235679 -0.791632291  1.379161654
#> 100 -0.257471449 -0.351417715  0.581319712
#> 101 -1.547510285 -2.113246793 -1.450289218
#> 102 -0.987955301 -1.351331125  0.419714220
#> 103 -1.094935517  0.914263992 -0.494044549
#> 104 -1.047272768 -0.334128952  0.048503223
#> 105 -1.262086698 -0.704465202 -0.529951219
#> 106 -1.298853622  1.917441095 -0.453272238
#> 107 -0.864107794 -2.807652597  1.081461752
#> 108 -1.106633524  1.902422768  0.035439529
#> 109 -1.034334608  1.033421343  0.976358892
#> 110 -1.334731788 -0.261023831 -2.241598697
#> 111 -0.825110453 -0.354448323 -0.922517494
#> 112 -0.914050897  0.062947697  0.382141962
#> 113 -0.982541096  0.305780658 -0.515804841
#> 114 -1.012120784 -1.513375042  0.801918308
#> 115 -1.204492324 -2.320811516 -0.251231169
#> 116 -1.083980048 -1.181589001 -1.145702551
#> 117 -0.938881795  0.080450924 -0.222520034
#> 118 -1.356902323  1.066423007 -2.425305417
#> 119 -1.509361218  2.168951237  0.346411704
#> 120 -0.710421003  0.419440779  1.912508857
#> 121 -1.138551530 -0.079755821 -1.145476660
#> 122 -0.988892569 -2.062130596  0.091249892
#> 123 -1.277831441  2.526106331  0.108417199
#> 124 -0.700368967  0.119379222  0.425841669
#> 125 -1.108659504 -0.341630973 -1.185926837
#> 126 -0.983840894  1.419437932 -0.702259835
#> 127 -0.678712403 -0.206065672  0.187120015
#> 128 -0.755834429 -0.708936615 -0.260544072
#> 129 -1.149540834 -0.490493735  0.012726199
#> 130 -0.798918990  2.051649335 -0.083355217
#> 131 -1.022408035  2.154958728  0.146620361
#> 132 -1.062803621  2.001558956 -2.313397437
#> 133 -1.193005404 -0.662058747 -0.074268203
#> 134 -0.668753846  0.456648210  0.477882440
#> 135 -0.933594906  0.193747239  1.125935375
#> 136 -1.107960254  1.985412790 -0.705605611
#> 137 -1.304131059 -1.921795674 -1.653330093
#> 138 -0.967007439 -0.277879329 -0.447726214
#> 139 -0.734963691 -0.922726179 -0.263292345
#> 140 -0.903061593  0.473685611 -0.776060897
#> 141 -1.190842657 -0.600129990 -0.988478752
#> 142 -0.840644110  0.229211664 -0.990596121
#> 143 -0.987955301 -1.351331125  0.419714220
#> 144 -1.267027415 -0.392201462 -1.107678512
#> 145 -1.282517783 -1.027891022 -1.533904443
#> 146 -0.949035083 -0.185368212 -0.719572864
#> 147 -0.795187396  0.138239511  0.824309504
#> 148 -0.876464312 -0.164023022 -0.437055258
#> 149 -1.190013544 -1.931134867 -1.582599437
#> 150 -0.913221784 -1.268057180 -0.211978723