Independent Component Analysis (ICA)

Performs an Independent Component Analysis using the FastICA algorithm. Contrary to PCA, that attempts to find uncorrelated sources (through least squares minimization), ICA attempts to find independent sources, i.e., the source space that maximizes the "non-gaussianity" of all sources. Contrary to PCA, ICA does not rank each source, which makes it a poor tool for dimensionality reduction. Requires the fastICA package to be installed.

ICA(x, n = "all", ...)

Arguments

x
A dataframe or a statistical model.

n
Number of components to extract. If n = "all" =, then n is set as the number of variables minus 1 (ncol(x)-1). If n = "auto" (default) or n = NULL, the number of components is selected through n_factors. In parameters_reduction, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

...
Arguments passed to or from other methods.

Examples

ICA(iris[, 1:4])

#>             ICA1          ICA2         ICA3
#> 1   -0.161988658  0.4138608548 -1.396802021
#> 2   -0.227033398 -0.7875330579 -1.335030186
#> 3    0.456786495 -0.3236713144 -1.328551509
#> 4    0.657546918 -0.5974599737 -1.199279655
#> 5    0.196362663  0.6390328425 -1.368671028
#> 6   -0.013502400  1.5234866087 -1.250419915
#> 7    1.131271059  0.2109240274 -1.207956824
#> 8    0.005889016  0.1535899635 -1.317313476
#> 9    0.894654022 -1.0774484519 -1.189652492
#> 10  -0.254044750 -0.6520668371 -1.329501501
#> 11  -0.645768886  0.9046399007 -1.435347298
#> 12   0.532118012  0.1184910601 -1.209693938
#> 13  -0.151932572 -0.8853021922 -1.349579575
#> 14   0.982784801 -0.8987017265 -1.354338930
#> 15  -1.396075033  1.6962431421 -1.702731047
#> 16  -0.260896280  2.7627066252 -1.440676483
#> 17  -0.145033393  1.5775576809 -1.449553152
#> 18   0.009583639  0.5008394213 -1.353334518
#> 19  -1.036772265  1.2329171897 -1.379854641
#> 20   0.377500980  1.1952093197 -1.305912572
#> 21  -0.915037981  0.1697166982 -1.333419315
#> 22   0.437395079  1.0462253308 -1.261657948
#> 23   1.051524163  0.6499416440 -1.452131807
#> 24   0.228020084  0.1623181393 -1.115475341
#> 25   0.630766257  0.0779377559 -1.060344010
#> 26  -0.407941025 -0.8037780264 -1.264381682
#> 27   0.381916359  0.3140293286 -1.180595160
#> 28  -0.375779033  0.4111336544 -1.375936826
#> 29  -0.520339979  0.1886888670 -1.424933014
#> 30   0.555434740 -0.3642246186 -1.179201581
#> 31   0.197083419 -0.5893966063 -1.207332574
#> 32  -0.637658883  0.3707093675 -1.346050927
#> 33   0.122717854  1.7399304206 -1.424127057
#> 34  -0.366933771  2.1087610138 -1.517984328
#> 35  -0.082472452 -0.5650882705 -1.286033998
#> 36  -0.316115624 -0.2777818432 -1.465089161
#> 37  -1.181563901  0.4705408937 -1.562257788
#> 38   0.271463489  0.5412637083 -1.383220417
#> 39   0.973449471 -0.8279681284 -1.240222923
#> 40  -0.240784107  0.1643805312 -1.346231591
#> 41   0.223374014  0.5035666217 -1.374199713
#> 42   0.116601263 -2.3819368824 -1.220163685
#> 43   1.196805866 -0.3560430175 -1.241797165
#> 44   0.836739151  0.7239490172 -1.094447275
#> 45   0.680604270  1.2281168140 -1.063311832
#> 46   0.191212023 -0.7113450591 -1.262644568
#> 47   0.238811431  1.0947129850 -1.299596766
#> 48   0.736342367 -0.3479796501 -1.249850085
#> 49  -0.399095762  0.8938493330 -1.406429183
#> 50  -0.138671930 -0.0688548238 -1.366309664
#> 51  -2.039814334  0.5086504272  0.220574413
#> 52  -0.453968793  0.5579211237  0.337983984
#> 53  -1.667481614  0.3218403345  0.393313771
#> 54   0.253427206 -1.7692252775  0.269479553
#> 55  -1.114471957 -0.3886562985  0.361997664
#> 56   0.482885688 -0.6354202052  0.456624264
#> 57   0.141720322  0.8430361419  0.509149099
#> 58   1.100248014 -1.7643174515 -0.036684556
#> 59  -1.592611478 -0.3158603084  0.245357422
#> 60   1.578848915 -0.7572504242  0.346769605
#> 61   0.472627598 -2.7244126418  0.037112433
#> 62   0.457392185  0.0725964785  0.334798871
#> 63  -1.606333501 -2.2121706942 -0.004726408
#> 64  -0.154790815 -0.2963523485  0.483198807
#> 65   0.545292274 -0.2885883048  0.036705474
#> 66  -1.510121406  0.2808694729  0.158765950
#> 67   1.296059801 -0.0003285288  0.570903142
#> 68  -0.521713518 -1.0674568203  0.098957524
#> 69  -1.077404520 -1.8232855663  0.403691426
#> 70  -0.145916866 -1.4469489640  0.102268880
#> 71   1.392761962  0.7243506807  0.762326994
#> 72  -0.668220548 -0.5246690940  0.092035260
#> 73  -0.857512058 -1.1586784045  0.571545185
#> 74  -0.609613607 -0.7062720371  0.397050922
#> 75  -1.197913476 -0.2968881396  0.153843723
#> 76  -1.375126480  0.0341163497  0.188471186
#> 77  -1.960298129 -0.4702986981  0.331342436
#> 78  -0.909786222  0.2247360087  0.588655437
#> 79   0.197689109 -0.1931288135  0.456017806
#> 80  -0.584015082 -1.2331033351 -0.170037096
#> 81  -0.043804688 -1.6801843191  0.082190807
#> 82  -0.248259733 -1.7536451176 -0.011060006
#> 83  -0.244334421 -0.8664641511  0.086325912
#> 84   0.343201502 -0.6591819662  0.799759407
#> 85   1.789406048 -0.0219096642  0.628739371
#> 86   0.927652393  1.0736625304  0.495549703
#> 87  -1.239900864  0.3272947352  0.351583381
#> 88  -1.588426789 -1.7369718082  0.237267875
#> 89   0.821384213 -0.1202145897  0.284834899
#> 90   0.476783601 -1.2973001666  0.267905311
#> 91   0.548420494 -1.2023872500  0.422783923
#> 92  -0.075995366 -0.0468720249  0.432628376
#> 93  -0.323129870 -1.1159444746  0.136896343
#> 94   0.741896693 -1.9894894393 -0.064815549
#> 95   0.519232369 -0.8416200241  0.336979572
#> 96   0.436021540 -0.2099203566  0.262232591
#> 97   0.495915640 -0.3589043455  0.306487215
#> 98  -0.704567229 -0.3184692750  0.211679952
#> 99   0.791504017 -1.3792418899 -0.201190331
#> 100  0.351354694 -0.5813491329  0.257491027
#> 101  2.113346726  1.4501122561  1.547539649
#> 102  1.351264641 -0.4198274019  0.987998146
#> 103 -0.914244792  0.4941526332  1.094902775
#> 104  0.334096862 -0.0485168594  1.047282374
#> 105  0.704483498  0.5299055616  1.262095656
#> 106 -1.917431171  0.4534810952  1.298795368
#> 107  2.807525273 -1.0817181469  0.864200556
#> 108 -1.902455135 -0.0352355589  1.106584395
#> 109 -1.033542925 -0.9762403466  1.034325020
#> 110  0.261205955  2.2415972638  1.334698564
#> 111  0.354516060  0.9224979159  0.825103242
#> 112 -0.063008604 -0.3821195319  0.914056078
#> 113 -0.305756414  0.5158520023  0.982523881
#> 114  1.513270918 -0.8020467459  1.012174697
#> 115  2.320804324  0.2510279864  1.204548541
#> 116  1.181671572  1.1456075117  1.083990484
#> 117 -0.080453936  0.2225445995  0.938875715
#> 118 -1.066223670  2.4254329051  1.356831096
#> 119 -2.169024244 -0.3461747300  1.509310673
#> 120 -0.419644531 -1.9124555420  0.710444201
#> 121  0.079836948  1.1454892780  1.138533149
#> 122  2.062095886 -0.0914318792  0.988948138
#> 123 -2.526150238 -0.1081497827  1.277767302
#> 124 -0.119438771 -0.4258175939  0.700373452
#> 125  0.341716798  1.1859135649  1.108647250
#> 126 -1.419395664  0.7024148440  0.983791218
#> 127  0.206029802 -0.1871278380  0.678721136
#> 128  0.708942068  0.2604889371  0.755848317
#> 129  0.490462433 -0.0127531475  1.149553891
#> 130 -2.051662150  0.0835681361  0.798863835
#> 131 -2.154999655 -0.1463934439  1.022354286
#> 132 -2.001362756  2.3136102116  1.062709932
#> 133  0.662034731  0.0742254191  1.193021394
#> 134 -0.456711969 -0.4778262742  0.668750440
#> 135 -0.193880672 -1.1258999275  0.933609954
#> 136 -1.985373441  0.7058176364  1.107895714
#> 137  1.921921633  1.6531673173  1.304151787
#> 138  0.277897385  0.4477165873  0.967006708
#> 139  0.922732443  0.2632161375  0.734983122
#> 140 -0.473634089  0.7761228936  0.903035336
#> 141  0.600194547  0.9884419217  1.190840693
#> 142 -0.229137739  0.9906333309  0.840620415
#> 143  1.351264641 -0.4198274019  0.987998146
#> 144  0.392275568  1.1076631741  1.267017882
#> 145  1.028005987  1.5338278311  1.282517263
#> 146  0.185413059  0.7195718720  0.949027074
#> 147 -0.138340120 -0.8242819063  0.795198507
#> 148  0.164042414  0.4370550369  0.876460794
#> 149  1.931256962  1.5824337192  1.190035780
#> 150  1.268053812  0.2118722656  0.913251164