cluster_discrimination {parameters}	R Documentation

Compute a linear discriminant analysis on classified cluster groups

Description

Computes linear discriminant analysis (LDA) on classified cluster groups, and determines the goodness of classification for each cluster group. See MASS::lda() for details.

Usage


cluster_discrimination(x, cluster_groups = NULL, ...)

Arguments

x
A data frame

cluster_groups
Group classification of the cluster analysis, which can be retrieved from the cluster_analysis() function.

...
Other arguments to be passed to or from.

See Also

n_clusters() to determine the number of clusters to extract, cluster_analysis() to compute a cluster analysis and performance::check_clusterstructure() to check suitability of data for clustering.

Examples



# Retrieve group classification from hierarchical cluster analysis
clustering <- cluster_analysis(iris[, 1:4], n = 3)

# Goodness of group classification
cluster_discrimination(clustering)


[Package parameters version 0.21.6 Index]