Compare model parameters of multiple models

Compute and extract model parameters of multiple regression models. See model_parameters for further details.

compare_parameters(
  ...,
  ci = 0.95,
  effects = "fixed",
  component = "conditional",
  standardize = NULL,
  exponentiate = FALSE,
  df_method = "wald",
  p_adjust = NULL,
  style = NULL,
  column_names = NULL,
  parameters = NULL,
  verbose = TRUE
)

compare_models(
  ...,
  ci = 0.95,
  effects = "fixed",
  component = "conditional",
  standardize = NULL,
  exponentiate = FALSE,
  df_method = "wald",
  p_adjust = NULL,
  style = NULL,
  column_names = NULL,
  parameters = NULL,
  verbose = TRUE
)

Arguments

...
One or more regression model objects, or objects returned by model_parameters(). Regression models may be of different model types.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

effects
Should parameters for fixed effects ("fixed"), random effects ("random"), or both ("all") be returned? Only applies to mixed models. May be abbreviated.

component
Model component for which parameters should be shown. See documentation for related model class in model_parameters.

standardize
The method used for standardizing the parameters. Can be "refit", "posthoc", "smart", "basic", "pseudo" or NULL (default) for no standardization. See 'Details' in standardize_parameters. Important: Categorical predictors (i.e. factors) are never standardized by default, which may be a different behaviour compared to other R packages or other software packages (like SPSS). If standardizing categorical predictors is desired, either use standardize="basic" to mimic behaviour of SPSS or packages such as lm.beta, or standardize the data with effectsize::standardize(force=TRUE) before fitting the model. Robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for logistic regression, or more generally speaking, for models with log or logit links. Note: Delta-method standard errors are also computed (by multiplying the standard errors by the transformed coefficients). This is to mimic behaviour of other software packages, such as Stata, but these standard errors poorly estimate uncertainty for the transformed coefficient. The transformed confidence interval more clearly captures this uncertainty. For compare_parameters(), exponentiate = "nongaussian" will only exponentiate coefficients from non-Gaussian families.

df_method
Method for computing degrees of freedom for p values, standard errors and confidence intervals (CI). See documentation for related model class in model_parameters.

p_adjust
Character vector, if not NULL, indicates the method to adjust p-values. See p.adjust for details. Further possible adjustment methods are "tukey", "scheffe", "sidak" and "none" to explicitly disable adjustment for emmGrid objects (from emmeans).

style
String, indicating which style of output is requested. Following templates are possible:

"ci": Estimate and confidence intervals, no asterisks for p-values.

"se": Estimate and standard errors, no asterisks for p-values.

"ci_p": Estimate, confidence intervals and asterisks for p-values.

"se_p": Estimate, standard errors and asterisks for p-values.

"ci_p2": Estimate, confidence intervals and numeric p-values, in two columns.

"se_p2": Estimate, standard errors and numeric p-values, in two columns.

column_names
Character vector with strings that should be used as column headers. Must be of same length as number of models in ....

parameters
Character vector of length 1 with a regular expression pattern that describes the parameters that should be returned from the data frame, or a named list of regular expressions. All non-matching parameters will be removed from the output. If parameters is a character vector, every parameter in the "Parameters" column that matches the regular expression in parameters will be selected from the returned data frame. Furthermore, if parameters has more than one element, these will be merged into a regular expression pattern like this: "(one|two|three)". If parameters is a named list of regular expression patterns, the names of the list-element should equal the column name where selection should be applied. This is useful for model objects where model_parameters() returns multiple columns with parameter components, like in model_parameters.lavaan.

verbose
Toggle warnings and messages.

Value

A data frame of indices related to the model's parameters.

Note

This function is in an early stage and does not yet cope with more complex models, and probably does not yet properly render all model components.

Examples

data(iris)
lm1 <- lm(Sepal.Length ~ Species, data = iris)
lm2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
lm3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)
compare_parameters(lm1, lm2, lm3)
#> Parameter                           |               lm1 |                  lm2 |                  lm3
#> -----------------------------------------------------------------------------------------------------
#> (Intercept)                         | 5.01 (4.86, 5.15) |  3.68 ( 3.47,  3.89) |  4.21 ( 3.41,  5.02)
#> Species (versicolor)                | 0.93 (0.73, 1.13) | -1.60 (-1.98, -1.22) | -1.81 (-2.99, -0.62)
#> Species (virginica)                 | 1.58 (1.38, 1.79) | -2.12 (-2.66, -1.58) | -3.15 (-4.41, -1.90)
#> Petal.Length                        |                   |  0.90 ( 0.78,  1.03) |  0.54 ( 0.00,  1.09)
#> Species (versicolor) * Petal.Length |                   |                      |  0.29 (-0.30,  0.87)
#> Species (virginica) * Petal.Length  |                   |                      |  0.45 (-0.12,  1.03)
#> -----------------------------------------------------------------------------------------------------
#> Observations                        |               150 |                  150 |                  150

data(mtcars)
m1 <- lm(mpg ~ wt, data = mtcars)
m2 <- glm(vs ~ wt + cyl, data = mtcars, family = "binomial")
compare_parameters(m1, m2)
#> Parameter    |                   m1 |                   m2
#> ----------------------------------------------------------
#> (Intercept)  | 37.29 (33.45, 41.12) | 10.62 ( 2.45, 18.79)
#> wt           | -5.34 (-6.49, -4.20) |  2.10 (-0.93,  5.13)
#> cyl          |                      | -2.93 (-5.63, -0.23)
#> ----------------------------------------------------------
#> Observations |                   32 |                   32

if (FALSE) {
# exponentiate coefficients, but not for lm
compare_parameters(m1, m2, exponentiate = "nongaussian")

# change column names
compare_parameters(m1, m2, column_names = c("linear model", "logistic reg."))
}