Degrees of Freedom (DoF)

Estimate or extract degrees of freedom of models.

degrees_of_freedom(model, method = "analytical")

dof(model, method = "analytical")

Arguments

model
A statistical model.

method
Can be "analytical" (default, DoFs are estimated based on the model type), "fit", in which case they are directly taken from the model if available (for Bayesian models, the goal (looking for help to make it happen) would be to refit the model as a frequentist one before extracting the DoFs), "ml1" (see dof_ml1), "betwithin" (see dof_betwithin), "satterthwaite" (see dof_satterthwaite), "kenward" (see dof_kenward) or "any", which tries to extract DoF by any of those methods, whichever succeeds.

Details

Methods for calculating degrees of freedom:

"analytical" For models of class lmerMod, Kenward-Roger approximated degrees of freedoms are calculated, for other models, n-k (number of observations minus number of parameters).

"fit" Tries to extract residual degrees of freedom, and returns Inf if residual df's could not be extracted.

"any" First tries to extract residual degrees of freedom, and if these are not available, extracts analytical degrees of freedom.

"nokr" Same as "analytical", but does not Kenward-Roger approximation for models of class lmerMod. Instead, always uses n-k to calculate df for any model.

"wald" Returns Inf.

"kenward" Calls dof_kenward.

"satterthwaite" Calls dof_satterthwaite.

"ml1" Calls dof_ml1.

"betwithin" Calls dof_betwithin.

Examples

model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
dof(model)

#> [1] 144

model <- glm(vs ~ mpg * cyl, data = mtcars, family = "binomial")
dof(model)

#> [1] 28

if (require("lme4")) {
  model <- lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
  dof(model)
}

#> Loading required package: lme4

#> Loading required package: Matrix

#> [1]   2.53354 144.66692

# \donttest{
if (require("rstanarm")) {
  model <- stan_glm(
    Sepal.Length ~ Petal.Length * Species,
    data = iris,
    chains = 2,
    refresh = 0
  )
  dof(model)
}

#> Loading required package: rstanarm

#> Loading required package: Rcpp

#> rstanarm (Version 2.19.3, packaged: 2020-02-11 05:16:41 UTC)

#> - Do not expect the default priors to remain the same in future rstanarm versions.

#> Thus, R scripts should specify priors explicitly, even if they are just the defaults.

#> - For execution on a local, multicore CPU with excess RAM we recommend calling

#> options(mc.cores = parallel::detectCores())

#> - bayesplot theme set to bayesplot::theme_default()

#>    * Does _not_ affect other ggplot2 plots

#>    * See ?bayesplot_theme_set for details on theme setting

#> [1] 144

# }