Print tables in different output formats

Prints tables (i.e. data frame) in different output formats. print_md() is a alias for display(format = "markdown").

# S3 method for parameters_model
display(
  object,
  format = "markdown",
  pretty_names = TRUE,
  split_components = TRUE,
  select = NULL,
  caption = NULL,
  subtitle = NULL,
  footer = NULL,
  align = NULL,
  digits = 2,
  ci_digits = 2,
  p_digits = 3,
  ci_brackets = c("(", ")"),
  ...
)

# S3 method for parameters_stan
display(
  object,
  split_components = TRUE,
  select = NULL,
  format = "markdown",
  ...
)

# S3 method for parameters_sem
display(
  object,
  format = "markdown",
  digits = 2,
  ci_digits = 2,
  p_digits = 3,
  ci_brackets = c("(", ")"),
  ...
)

# S3 method for parameters_efa_summary
display(object, format = "markdown", digits = 3, ...)

# S3 method for parameters_efa
display(
  object,
  format = "markdown",
  digits = 2,
  sort = FALSE,
  threshold = NULL,
  labels = NULL,
  ...
)

# S3 method for equivalence_test_lm
display(object, format = "markdown", digits = 2, ...)

# S3 method for parameters_model
format(
  x,
  pretty_names = TRUE,
  split_components = TRUE,
  select = NULL,
  digits = 2,
  ci_digits = 2,
  p_digits = 3,
  ci_width = NULL,
  ci_brackets = NULL,
  zap_small = FALSE,
  format = NULL,
  ...
)

# S3 method for parameters_model
print_html(
  x,
  pretty_names = TRUE,
  split_components = TRUE,
  select = NULL,
  caption = NULL,
  subtitle = NULL,
  footer = NULL,
  align = NULL,
  digits = 2,
  ci_digits = 2,
  p_digits = 3,
  ci_brackets = c("(", ")"),
  show_sigma = FALSE,
  show_formula = FALSE,
  verbose = TRUE,
  ...
)

# S3 method for parameters_model
print_md(
  x,
  pretty_names = TRUE,
  split_components = TRUE,
  select = NULL,
  caption = NULL,
  subtitle = NULL,
  footer = NULL,
  digits = 2,
  ci_digits = 2,
  p_digits = 3,
  ci_brackets = c("(", ")"),
  ...
)

Arguments

object
An object returned by model_parameters(), simulate_parameters(), equivalence_test() or principal_components().

format
String, indicating the output format. Can be "markdown" or "html".

pretty_names
Return "pretty" (i.e. more human readable) parameter names.

split_components
Logical, if TRUE (default), For models with multiple components (zero-inflation, smooth terms, ...), each component is printed in a separate table. If FALSE, model parameters are printed in a single table and a Component column is added to the output.

select
Character vector (or numeric index) of column names that should be printed. If NULL (default), all columns are printed. The shortcut select = "minimal" prints coefficient, confidence intervals and p-values, while select = "short" prints coefficient, standard errors and p-values.

caption
Table title (same as caption) and subtitle, as strings. If NULL, no title or subtitle is printed, unless it is stored as attributes (table_title, or its alias table_caption, and table_subtitle).

subtitle
Table title (same as caption) and subtitle, as strings. If NULL, no title or subtitle is printed, unless it is stored as attributes (table_title, or its alias table_caption, and table_subtitle).

footer
Table footer, as string. For markdown-formatted tables, table footers, due to the limitation in markdown rendering, are actually just a new text line under the table.

align
Only applies to HTML tables. May be one of "left", "right" or "center".

digits
Number of decimal places for numeric values (except confidence intervals and p-values).

ci_digits
Number of decimal places for confidence intervals.

p_digits
Number of decimal places for p-values. May also be "scientific" for scientific notation of p-values.

ci_brackets
Logical, if TRUE (default), CI-values are encompassed in square brackets (else in parentheses).

...
Arguments passed to or from other methods.

sort
Sort the loadings.

threshold
A value between 0 and 1 indicates which (absolute) values from the loadings should be removed. An integer higher than 1 indicates the n strongest loadings to retain. Can also be "max", in which case it will only display the maximum loading per variable (the most simple structure).

labels
A character vector containing labels to be added to the loadings data. Usually, the question related to the item.

x
An object returned by model_parameters().

ci_width
Minimum width of the returned string for confidence intervals. If not NULL and width is larger than the string's length, leading whitespaces are added to the string. If width="auto", width will be set to the length of the longest string.

zap_small
Logical, if TRUE, small values are rounded after digits decimal places. If FALSE, values with more decimal places than digits are printed in scientific notation.

show_sigma
Logical, if TRUE, adds information about the residual standard deviation.

show_formula
Logical, if TRUE, adds the model formula to the output.

verbose
Toggle messages and warnings.

Value

If format = "markdown", the return value will be a character vector in markdown-table format. If format = "html", an object of class gt_tbl.

Details

display() is useful when the table-output from functions, which is usually printed as formatted text-table to console, should be formatted for pretty table-rendering in markdown documents, or if knitted from rmarkdown to PDF or Word files. See vignette for examples.

Examples

model <- lm(mpg ~ wt + cyl, data = mtcars)
mp <- model_parameters(model)
display(mp)

#> 
#> 
#> |Parameter   | Coefficient |   SE |         95% CI | t(29) |      p |
#> |:-----------|:-----------:|:----:|:--------------:|:-----:|:------:|
#> |(Intercept) |       39.69 | 1.71 | (36.18, 43.19) | 23.14 | < .001 |
#> |wt          |       -3.19 | 0.76 | (-4.74, -1.64) | -4.22 | < .001 |
#> |cyl         |       -1.51 | 0.41 | (-2.36, -0.66) | -3.64 | 0.001  |