Factor Analysis (FA)

This function performs a Factor Analysis (FA).

factor_analysis(
  x,
  n = "auto",
  rotation = "none",
  sort = FALSE,
  threshold = NULL,
  standardize = TRUE,
  ...
)

Arguments

x
A data frame or a statistical model.

n
Number of components to extract. If n="all", then n is set as the number of variables minus 1 (ncol(x)-1). If n="auto" (default) or n=NULL, the number of components is selected through n_factors. In reduce_parameters, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

rotation
If not "none", the PCA / FA will be computed using the psych package. Possible options include "varimax", "quartimax", "promax", "oblimin", "simplimax", and "cluster". See fa for details.

sort
Sort the loadings.

threshold
A value between 0 and 1 indicates which (absolute) values from the loadings should be removed. An integer higher than 1 indicates the n strongest loadings to retain. Can also be "max", in which case it will only display the maximum loading per variable (the most simple structure).

standardize
A logical value indicating whether the variables should be standardized (centered and scaled) to have unit variance before the analysis takes place (in general, such scaling is advisable).

...
Arguments passed to or from other methods.

Value

A data frame of loadings.

Details

Complexity

Complexity represents the number of latent components needed to account for the observed variables. Whereas a perfect simple structure solution has a complexity of 1 in that each item would only load on one factor, a solution with evenly distributed items has a complexity greater than 1 (Hofman, 1978; Pettersson and Turkheimer, 2010) .

FA or PCA?

There is a simplified rule of thumb that may help do decide whether to run a principal component analysis or a factor analysis:

Run factor analysis if you assume or wish to test a theoretical model of latent factors causing observed variables.

Run principal component analysis If you want to simply reduce your correlated observed variables to a smaller set of important independent composite variables.

(Source: CrossValidated)

Note

There is a summary()-method that prints the Eigenvalues and (explained) variance for each extracted component.

References

Hofmann, R. (1978). Complexity and simplicity as objective indices descriptive of factor solutions. Multivariate Behavioral Research, 13:2, 247-250, doi: 10.1207/s15327906mbr1302_9

Pettersson, E., & Turkheimer, E. (2010). Item selection, evaluation, and simple structure in personality data. Journal of research in personality, 44(4), 407-420, doi: 10.1016/j.jrp.2010.03.002

Examples

library(parameters)

factor_analysis(mtcars[, 1:7], n = "all", threshold = 0.2)

#> # Loadings from Factor Analysis (no rotation)
#> 
#> Variable |   MR1 |   MR2 |  MR3 |  MR4 | MR5 | MR6 | Complexity | Uniqueness
#> ----------------------------------------------------------------------------
#> mpg      | -0.92 |       |      | 0.22 |     |     |       1.16 |       0.08
#> cyl      |  0.96 |       |      |      |     |     |       1.14 |       0.01
#> disp     |  0.95 |       |      | 0.20 |     |     |       1.13 |       0.03
#> hp       |  0.86 | -0.35 |      |      |     |     |       1.38 |       0.12
#> drat     | -0.72 | -0.40 | 0.29 |      |     |     |       1.96 |       0.24
#> wt       |  0.89 |  0.38 | 0.24 |      |     |     |       1.53 |       0.00
#> qsec     | -0.53 |  0.76 |      |      |     |     |       1.81 |       0.14
#> 
#> The 6 latent factors accounted for 90.99% of the total variance of the original data (MR1 = 71.60%, MR2 = 14.55%, MR3 = 2.91%, MR4 = 1.35%, MR5 = 0.45%, MR6 = 0.12%).

factor_analysis(mtcars[, 1:7], n = 2, rotation = "oblimin", threshold = "max", sort = TRUE)

#> # Rotated loadings from Factor Analysis (oblimin-rotation)
#> 
#> Variable |   MR1 |  MR2 | Complexity | Uniqueness
#> -------------------------------------------------
#> wt       |  1.00 |      |       1.07 |       0.10
#> disp     |  0.92 |      |       1.02 |       0.09
#> mpg      | -0.88 |      |       1.02 |       0.15
#> drat     | -0.84 |      |       1.14 |       0.39
#> cyl      |  0.82 |      |       1.23 |       0.08
#> hp       |  0.60 |      |       1.94 |       0.18
#> qsec     |       | 1.00 |       1.00 |       0.00
#> 
#> The 2 latent factors (oblimin rotation) accounted for 85.83% of the total variance of the original data (MR1 = 63.85%, MR2 = 21.99%).

factor_analysis(mtcars[, 1:7], n = 2, threshold = 2, sort = TRUE)

#> # Loadings from Factor Analysis (no rotation)
#> 
#> Variable |  MR1 |  MR2 | Complexity | Uniqueness
#> ------------------------------------------------
#> cyl      | 0.96 |      |       1.01 |       0.08
#> disp     | 0.95 |      |       1.03 |       0.09
#> mpg      |      |      |       1.03 |       0.15
#> wt       |      | 0.36 |       1.33 |       0.10
#> hp       |      |      |       1.23 |       0.18
#> drat     |      |      |       1.49 |       0.39
#> qsec     |      | 0.83 |       1.74 |       0.00
#> 
#> The 2 latent factors accounted for 85.83% of the total variance of the original data (MR1 = 70.67%, MR2 = 15.16%).

efa <- factor_analysis(mtcars[, 1:5], n = 2)
summary(efa)

#> # (Explained) Variance of Components
#> 
#> Parameter                       |   MR1 |   MR2
#> -----------------------------------------------
#> Eigenvalues                     | 3.908 | 0.398
#> Variance Explained              | 0.782 | 0.080
#> Variance Explained (Cumulative) | 0.782 | 0.861
#> Variance Explained (Proportion) | 0.908 | 0.092

predict(efa)

#>            MR1         MR2
#> 1  -0.41953471 -0.35544083
#> 2  -0.41953471 -0.35544083
#> 3  -0.93451791  0.05188420
#> 4  -0.02247401 -1.06704187
#> 5   0.80423811 -0.57723722
#> 6   0.05132529 -1.40382795
#> 7   1.17796287  0.82565287
#> 8  -0.97981546 -0.77961043
#> 9  -0.87835186  0.02703875
#> 10 -0.31368972 -0.16476124
#> 11 -0.28032337 -0.23570818
#> 12  0.73053450 -0.34851886
#> 13  0.70908470 -0.30291012
#> 14  0.75913424 -0.40933052
#> 15  1.40523363 -0.78714835
#> 16  1.40381792 -0.48501471
#> 17  1.26324243  0.25588537
#> 18 -1.39062136  0.13411535
#> 19 -1.62307787  0.15402661
#> 20 -1.48149201  0.28280884
#> 21 -0.82463188 -0.03789222
#> 22  0.79425443 -1.37121293
#> 23  0.67357950 -1.14361360
#> 24  1.04878537  1.07073183
#> 25  0.89062471 -0.71650907
#> 26 -1.26846941 -0.12530276
#> 27 -1.14242028  0.42262235
#> 28 -1.03993655  0.89398689
#> 29  0.94311149  1.87551864
#> 30 -0.08324543  0.96674422
#> 31  1.32372015  3.27012728
#> 32 -0.87651282  0.43537850

# \donttest{
# Automated number of components
factor_analysis(mtcars[, 1:4], n = "auto")

#> # Loadings from Factor Analysis (no rotation)
#> 
#> Variable |   MR1 | Complexity | Uniqueness
#> ------------------------------------------
#> mpg      | -0.90 |       1.00 |       0.19
#> cyl      |  0.96 |       1.00 |       0.08
#> disp     |  0.93 |       1.00 |       0.13
#> hp       |  0.86 |       1.00 |       0.26
#> 
#> The unique latent factor accounted for 83.55% of the total variance of the original data.

# }