factor_scores {parameters}	R Documentation

Extract factor scores from Factor Analysis (EFA) or Omega

Description

factor_scores() extracts the factor scores from objects returned by psych::fa(), factor_analysis(), or psych::omega()

Usage

factor_scores(x, ...)

Arguments

x
An object returned by psych::fa(), factor_analysis(), or psych::omega().

...
Currently unused.

Value

A data frame with the factor scores. It simply extracts the $scores element from the object and converts it into a data frame.

See Also

factor_analysis()

Examples

data(mtcars)
out <- factor_analysis(mtcars[, 1:7], n = 2)
head(factor_scores(out))

[Package parameters version 0.28.3 Index]