Model Parameters

Compute and extract model parameters. See the documentation for your object's class:

Correlations, t-tests, ... (htest, pairwise.htest)

ANOVAs (aov, anova, afex, ...)

Regression models (lm, glm, survey, ...)

Additive models (gam, gamm, ...)

Zero-inflated models (hurdle, zeroinfl, zerocount)

Multinomial, ordinal and cumulative link models (bracl, multinom, mlm, ...)

Other special models (model.avg, betareg, glmx, ...)

Mixed models (lme4, nlme, glmmTMB, afex, ...)

Bayesian tests (BayesFactor)

Bayesian models (rstanarm, brms, MCMCglmm, ...)

PCA and FA (psych)

CFA and SEM (lavaan, blavaan)

Cluster models (k-means, ...)

Meta-Analysis via linear (mixed) models (rma, metaplus, metaBMA, ...)

Hypothesis testing (glht, PMCMRplus)

Robust statistical tests (WRS2)

Multiply imputed repeated analyses (mira)

model_parameters(model, ...)

parameters(model, ...)

Arguments

model
Statistical Model.

...
Arguments passed to or from other methods. Non-documented arguments are digits, p_digits and ci_digits to set the number of digits for the output. See 'Examples' in model_parameters.default.

Value

A data frame of indices related to the model's parameters.

Details

Standardization is based on standardize_parameters(). In case of standardize = "refit", the data used to fit the model will be standardized and the model is completely refitted. In such cases, standard errors and confidence intervals refer to the standardized coefficient.

Note

The print() method has several arguments to tweak the output. There is also a plot()-method implemented in the see-package, and a dedicated method for use inside rmarkdown files, print_md().

Labeling the Degrees of Freedom

Throughout the parameters package, we decided to label the residual degrees of freedom df_error. The reason for this is that these degrees of freedom not always refer to the residuals. For certain models, they refer to the estimate error - in a linear model these are the same, but in - for instance - any mixed effects model, this isn't strictly true. Hence, we think that df_error is the most generic label for these degrees of freedom.

Interpretation of Interaction Terms

Note that the interpretation of interaction terms depends on many characteristics of the model. The number of parameters, and overall performance of the model, can differ or not between a * b a : b, and a / b, suggesting that sometimes interaction terms give different parameterizations of the same model, but other times it gives completely different models (depending on a or b being factors of covariates, included as main effects or not, etc.). Their interpretation depends of the full context of the model, which should not be inferred from the parameters table alone - rather, we recommend to use packages that calculate estimated marginal means or marginal effects, such as modelbased, emmeans or ggeffects. To raise awareness for this issue, you may use print(...,show_formula=TRUE) to add the model-specification to the output of the print() method for model_parameters().

See also