Model Parameters

Compute and extract model parameters. See the documentation for your object's class:

Correlations and t-tests

ANOVAs

Regression models (lm, glm, survey, ...)

Additive models (gam, gamm, ...)

Zero-inflated models (hurdle, zeroinfl, zerocount)

Mixed models (lme4, nlme, glmmTMB, ...)

Bayesian tests (BayesFactor)

Bayesian models (rstanarm, brms, MCMCglmm)

PCA and FA (psych)

CFA and SEM (lavaan)

Cluster models (k-means, ...)

Meta-Analysis via linear (mixed) models (rma)

model_parameters(model, ...)

parameters(model, ...)

Arguments

model
Statistical Model.

...
Arguments passed to or from other methods.

Value

A data frame of indices related to the model's parameters.

Details

Standardization is based on standardize_parameters(). In case of standardize = "refit", the data used to fit the model will be standardized and the model is completely refitted. In such cases, standard errors and confidence intervals refer to the standardized coefficient. For other methods of standardizing, unstandardized and standardized coefficients are shown, where standard errors and confidence intervals relate to the unstandardized coefficients.

Note

The print() method has several arguments to tweak the output.

See also