Model Parameters

Compute and extract model parameters. See the documentation for your object's class:

Correlations and t-tests

ANOVAs

Regression models (lm, glm, survey, ...)

Additive models (gam, gamm, ...)

Zero-inflated models (hurdle, zeroinfl, zerocount)

Multinomial, ordinal and cumulative link models (bracl, multinom, mlm, ...)

Mixed models (lme4, nlme, glmmTMB, ...)

Bayesian tests (BayesFactor)

Bayesian models (rstanarm, brms, MCMCglmm, ...)

PCA and FA (psych)

CFA and SEM (lavaan, blavaan)

Cluster models (k-means, ...)

Meta-Analysis via linear (mixed) models (rma)

model_parameters(model, ...)

parameters(model, ...)

Arguments

model
Statistical Model.

...
Arguments passed to or from other methods. Non-documented arguments are digits, p_digits and ci_digits to set the number of digits for the output. See 'Examples' in model_parameters.default.

Value

A data frame of indices related to the model's parameters.

Details

Standardization is based on standardize_parameters. In case of standardize = "refit", the data used to fit the model will be standardized and the model is completely refitted. In such cases, standard errors and confidence intervals refer to the standardized coefficient.

Note

The print() method has several arguments to tweak the output. There is also a plot()-method implemented in the see-package.

See also