Parameters from BayesFactor objects

Parameters from BayesFactor objects.

# S3 method for BFBayesFactor
model_parameters(
  model,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.95,
  ci_method = "hdi",
  test = c("pd", "rope"),
  rope_range = "default",
  rope_ci = 0.95,
  priors = TRUE,
  verbose = TRUE,
  ...
)

Arguments

model

Object of class BFBayesFactor.

centrality

The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

dispersion

Logical, if TRUE, computes indices of dispersion related to the estimate(s) (SD and MAD for mean and median, respectively).

ci

Value or vector of probability of the CI (between 0 and 1) to be estimated. Default to .95 (95%).

ci_method

The type of index used for Credible Interval. Can be "HDI" (default, see hdi()), "ETI" (see eti()), "BCI" (see bci()) or "SI" (see si()).

test

The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests. For each "test", the corresponding bayestestR function is called (e.g. rope() or p_direction()) and its results included in the summary output.

rope_range

ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.1, 0.1)) or "default". If "default", the bounds are set to x +- 0.1*SD(response).

rope_ci

The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use for the percentage in ROPE.

priors

Add the prior used for each parameter.

verbose

Toggle off warnings.

...

Additional arguments to be passed to or from methods.

Value

A data frame of indices related to the model's parameters.

Details

The meaning of the extracted parameters:

For BayesFactor::ttestBF(): Difference is the raw difference between the means.

For BayesFactor::correlationBF(): rho is the linear correlation estimate (equivalent to Pearson's r).

For BayesFactor::lmBF() / BayesFactor::generalTestBF() / BayesFactor::regressionBF() / BayesFactor::anovaBF(): in addition to parameters of the fixed and random effects, there are: mu is the (mean-centered) intercept; sig2 is the model's sigma; g / g_* are the g parameters; See the Bayes Factors for ANOVAs paper (doi: 10.1016/j.jmp.2012.08.001 ).

Examples

# \donttest{
if (require("BayesFactor")) {
  model <- ttestBF(x = rnorm(100, 1, 1))
  model_parameters(model)
}
#> Loading required package: BayesFactor
#> Loading required package: coda
#> Loading required package: Matrix
#> ************
#> Welcome to BayesFactor 0.9.12-4.2. If you have questions, please contact Richard Morey (richarddmorey@gmail.com).
#> 
#> Type BFManual() to open the manual.
#> ************
#> Bayesian t-test
#> 
#> Parameter  | Median |       95% CI |   pd | % in ROPE |              Prior |     BF
#> -----------------------------------------------------------------------------------
#> Difference |   0.88 | [0.69, 1.08] | 100% |        0% | Cauchy (0 +- 0.71) | > 1000
#> Cohen's D  |   0.88 | [0.64, 1.09] | 100% |        0% |                    |       
#> 
#> Uncertainty intervals (highest-density) and p values (two-tailed) computed
#>   using a MCMC distribution approximation.
# }