Format PCA/FA from the psych package

Format PCA/FA objects from the psych package (Revelle, 2016).

# S3 method for befa
model_parameters(
  model,
  sort = FALSE,
  centrality = "median",
  dispersion = FALSE,
  ci = 0.89,
  ci_method = "hdi",
  test = NULL,
  ...
)

Arguments

model
Bayesian EFA created by the BayesFM::befa.

sort
Sort the loadings.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

dispersion
Logical, if TRUE, computes indices of dispersion related to the estimate(s) (SD and MAD for mean and median, respectively).

ci
Value or vector of probability of the CI (between 0 and 1) to be estimated. Default to .89 (89%) for Bayesian models and .95 (95%) for frequentist models.

ci_method
The type of index used for Credible Interval. Can be "HDI" (default, see hdi) or "ETI" (see eti).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests. For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

...
Arguments passed to or from other methods.

Value

A data.frame of loadings.

Examples

library(parameters)
# \donttest{
library(BayesFM)
efa <- BayesFM::befa(mtcars, iter = 1000)

#> starting MCMC sampling...
#>     5%
#>    10%
#>    15%
#>    20%
#>    25%
#>    30%
#>    35%
#>    40%
#>    45%
#> done with burn-in period
#>    50%
#>    55%
#>    60%
#>    65%
#>    70%
#>    75%
#>    80%
#>    85%
#>    90%
#>    95%
#>   100%
#> done with sampling!

results <- model_parameters(efa, sort = TRUE)

results

#> Error in if (rotation_name == "none") {    insight::print_color("# Loadings from Principal Component Analysis (no rotation)\n\n",         "blue")} else {    insight::print_color(sprintf("# Rotated loadings from Principal Component Analysis (%s-rotation)\n\n",         rotation_name), "blue")}: argument is of length zero

attributes(results)$loadings_long

#>    Component Variable      Median      CI_low     CI_high
#> 1         F1       am   0.4695616   0.3311917   0.6395318
#> 2         F2     carb   2.9320088   2.2810250   3.6029693
#> 3         F2      cyl   6.2701942   5.0651192   7.4969249
#> 4         F2     disp 249.9062459 196.8312214 298.9218372
#> 5         F3     drat   3.5561627   2.9004151   4.2487609
#> 6         F3     gear   3.6446566   2.9236370   4.2825975
#> 7         F2       hp 154.9943096 128.7823320 190.3206426
#> 8         F1      mpg  20.4179358  16.3929602  24.0807624
#> 9         F3     qsec  17.4950889  14.3109675  20.9080144
#> 10        F1       vs   0.5029808   0.3347645   0.6469265
#> 11        F2       wt   3.2577404   2.6000331   3.8877100

efa_to_cfa(results)

#> # Latent variables
#> F1 =~ am + mpg + vs
#> F2 =~ carb + cyl + disp + hp + wt
#> F3 =~ drat + gear + qsec

# }