Format PCA/FA from the psych package

Format PCA/FA objects from the psych package (Revelle, 2016).

# S3 method for befa
model_parameters(model, sort = FALSE,
  centrality = "median", dispersion = FALSE, ci = 0.89,
  ci_method = "hdi", test = NULL, ...)

Arguments

model
Bayesain EFA created by the BayesFM::befa.

sort
Sort the loadings.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

dispersion
Logical, if TRUE, computes indices of dispersion related to the estimate(s) (SD and MAD for mean and median, respectively).

ci
Value or vector of probability of the CI (between 0 and 1) to be estimated. Default to .89 (89%) for Bayesian models and .95 (95%) for frequentist models.

ci_method
The type of index used for Credible Interval. Can be "HDI" (default, see hdi) or "ETI" (see eti).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests. For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

...
Arguments passed to or from other methods.

Value

A data.frame of loadings.

Examples

library(parameters)

library(BayesFM)
efa <- BayesFM::befa(mtcars, iter = 1000)

#> starting MCMC sampling...
#>     5%
#>    10%
#>    15%
#>    20%
#>    25%
#>    30%
#>    35%
#>    40%
#>    45%
#> done with burn-in period
#>    50%
#>    55%
#>    60%
#>    65%
#>    70%
#>    75%
#>    80%
#>    85%
#>    90%
#>    95%
#>   100%
#> done with sampling!

results <- model_parameters(efa, sort = TRUE)

results

#> Error in if (.rotation == "none") {    insight::print_color("# Loadings from Principal Component Analysis (no rotation)\n\n",         "blue")} else {    insight::print_color(sprintf("# Rotated loadings from Principal Component Analysis (%s-rotation)\n\n",         .rotation), "blue")}: argument is of length zero

attributes(results)$loadings_long

#>    Component Variable      Median      CI_low     CI_high
#> 1         F1       am   0.4661396   0.3121343   0.6233666
#> 2         F2     carb   2.8962850   2.2688717   3.6219137
#> 3         F2      cyl   6.2192914   4.9456329   7.4280836
#> 4         F2     disp 245.6563182 199.1204195 302.9297170
#> 5         F3     drat   3.5478683   2.8443476   4.1295327
#> 6         F3     gear   3.6400011   2.9764192   4.3040214
#> 7         F2       hp 153.3919473 120.2079746 183.9712342
#> 8         F1      mpg  20.3826710  16.6750021  24.1551478
#> 9         F3     qsec  17.4530735  14.0305072  20.4157769
#> 10        F1       vs   0.5012092   0.3248375   0.6581825
#> 11        F2       wt   3.2111347   2.6503087   3.9499441

efa_to_cfa(results)

#> # Latent variables
#> F1 =~ am + mpg + vs
#> F2 =~ carb + cyl + disp + hp + wt
#> F3 =~ drat + gear + qsec