Parameters from Generalized Additive (Mixed) Models

Extract and compute indices and measures to describe parameters of generalized additive models (GAM(M)s).

# S3 method for cgam
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("conditional", "smooth_terms", "all"),
  standardize = NULL,
  exponentiate = FALSE,
  verbose = TRUE,
  ...
)

# S3 method for gam
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for rqss
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("conditional", "smooth_terms", "all"),
  standardize = NULL,
  exponentiate = FALSE,
  verbose = TRUE,
  ...
)

Arguments

model
A gam/gamm model.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

bootstrap
Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also bootstrap_parameters()).

iterations
The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

component
Model component for which parameters should be shown. May be one of "conditional", "precision" (betareg), "scale" (ordinal), "extra" (glmx), "marginal" (mfx), "conditional" or "full" (for MuMIn::model.avg()) or "all".

standardize
The method used for standardizing the parameters. Can be "refit", "posthoc", "smart", "basic", "pseudo" or NULL (default) for no standardization. See 'Details' in standardize_parameters. Note that robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for, say, logistic regressions, or more generally speaking: for models with log or logit link. Note: standard errors are also transformed (by multiplying the standard errors with the exponentiated coefficients), to mimic behaviour of other software packages, such as Stata.

verbose
Toggle warnings and messages.

...
Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like ci_method are passed down to describe_posterior.

robust
Logical, if TRUE, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors. Additional arguments like vcov_estimation or vcov_type are passed down to other methods, see standard_error_robust() for details.

p_adjust
Character vector, if not NULL, indicates the method to adjust p-values. See p.adjust for details.

Value

A data frame of indices related to the model's parameters.

Details

The reporting of degrees of freedom for the spline terms slightly differs from the output of summary(model), for example in the case of mgcv::gam(). The estimated degrees of freedom, column edf in the summary-output, is named df in the returned data frame, while the column df_error in the returned data frame refers to the residual degrees of freedom that are returned by df.residual(). Hence, the values in the the column df_error differ from the column Ref.df from the summary, which is intentional, as these reference degrees of freedom “is not very interpretable” (web).

See also

Examples

library(parameters)
if (require("mgcv")) {
  dat <- gamSim(1, n = 400, dist = "normal", scale = 2)
  model <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat)
  model_parameters(model)
}

#> Loading required package: mgcv

#> Loading required package: nlme

#> 
#> Attaching package: ‘nlme’

#> The following object is masked from ‘package:lme4’:
#> 
#>     lmList

#> This is mgcv 1.8-33. For overview type 'help("mgcv-package")'.

#> 
#> Attaching package: ‘mgcv’

#> The following object is masked from ‘package:mclust’:
#> 
#>     mvn

#> Gu & Wahba 4 term additive model

#> # Fixed Effects 
#> 
#> Parameter   | Coefficient |   SE |       95% CI | t(381.66) |      p
#> --------------------------------------------------------------------
#> (Intercept) |        7.77 | 0.10 | [7.58, 7.96] |     80.10 | < .001
#> 
#> # Smooth Terms 
#> 
#> Parameter        |     F |   df | df (error) |      p
#> -----------------------------------------------------
#> Smooth term (x0) | 16.10 | 3.00 |     381.66 | < .001
#> Smooth term (x1) | 75.24 | 3.90 |     381.66 | < .001
#> Smooth term (x2) | 85.76 | 7.70 |     381.66 | < .001
#> Smooth term (x3) |  0.74 | 2.74 |     381.66 | 0.465