Parameters from Generalized Additive (Mixed) Models

Extract and compute indices and measures to describe parameters of generalized additive models (GAM(M)s).

# S3 method for cgam
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  parameters = NULL,
  verbose = TRUE,
  ...
)

# S3 method for gam
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  parameters = NULL,
  verbose = TRUE,
  ...
)

# S3 method for rqss
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  parameters = NULL,
  verbose = TRUE,
  ...
)

Arguments

model
A gam/gamm model.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

bootstrap
Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also bootstrap_parameters()).

iterations
The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

standardize
The method used for standardizing the parameters. Can be "refit", "posthoc", "smart", "basic", "pseudo" or NULL (default) for no standardization. See 'Details' in standardize_parameters. Important: Categorical predictors (i.e. factors) are never standardized by default, which may be a different behaviour compared to other R packages or other software packages (like SPSS). If standardizing categorical predictors is desired, standardize the data with effectsize::standardize(force = TRUE) before fitting the model. Robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for logistic regression, or more generally speaking, for models with log or logit links. Note: Delta-method standard errors are also computed (by multiplying the standard errors by the transformed coefficients). This is to mimic behaviour of other software packages, such as Stata, but these standard errors poorly estimate uncertainty for the transformed coefficient. The transformed confidence interval more clearly captures this uncertainty. For compare_parameters(), exponentiate = "nongaussian" will only exponentiate coefficients from non-Gaussian families.

robust
Logical, if TRUE, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors. Additional arguments like vcov_estimation or vcov_type are passed down to other methods, see standard_error_robust() for details and this vignette for working examples.

p_adjust
Character vector, if not NULL, indicates the method to adjust p-values. See p.adjust for details. Further possible adjustment methods are "tukey", "scheffe", "sidak" and "none" to explicitly disable adjustment for emmGrid objects (from emmeans).

parameters
Character vector of length 1 with a regular expression pattern that describes the parameters that should be returned from the data frame, or a named list of regular expressions. All non-matching parameters will be removed from the output. If parameters is a character vector, every parameter in the "Parameters" column that matches the regular expression in parameters will be selected from the returned data frame. Furthermore, if parameters has more than one element, these will be merged into a regular expression pattern like this: "(one|two|three)". If parameters is a named list of regular expression patterns, the names of the list-element should equal the column name where selection should be applied. This is useful for model objects where model_parameters() returns multiple columns with parameter components, like in model_parameters.lavaan.

verbose
Toggle warnings and messages.

...
Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like ci_method are passed down to describe_posterior.

Value

A data frame of indices related to the model's parameters.

Details

The reporting of degrees of freedom for the spline terms slightly differs from the output of summary(model), for example in the case of mgcv::gam(). The estimated degrees of freedom, column edf in the summary-output, is named df in the returned data frame, while the column df_error in the returned data frame refers to the residual degrees of freedom that are returned by df.residual(). Hence, the values in the the column df_error differ from the column Ref.df from the summary, which is intentional, as these reference degrees of freedom “is not very interpretable” (web).

See also

Examples

library(parameters)
if (require("mgcv")) {
  dat <- gamSim(1, n = 400, dist = "normal", scale = 2)
  model <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat)
  model_parameters(model)
}
#> Loading required package: mgcv
#> Loading required package: nlme
#> This is mgcv 1.8-35. For overview type 'help("mgcv-package")'.
#> 
#> Attaching package: ‘mgcv’
#> The following object is masked from ‘package:mclust’:
#> 
#>     mvn
#> Gu & Wahba 4 term additive model
#> # Fixed Effects 
#> 
#> Parameter   | Coefficient |   SE |       95% CI | t(382.25) |      p
#> --------------------------------------------------------------------
#> (Intercept) |        7.72 | 0.10 | [7.53, 7.91] |     78.98 | < .001
#> 
#> # Smooth Terms 
#> 
#> Parameter        |     F |   df |      p
#> ----------------------------------------
#> Smooth term (x0) | 11.76 | 3.07 | < .001
#> Smooth term (x1) | 91.93 | 3.09 | < .001
#> Smooth term (x2) | 89.17 | 8.69 | < .001
#> Smooth term (x3) |  0.86 | 1.90 | 0.362