Parameters from (General) Linear Models

Extract and compute indices and measures to describe parameters of (general) linear models (GLMs).

# S3 method for default
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  summary = FALSE,
  parameters = NULL,
  verbose = TRUE,
  ...
)

# S3 method for glm
model_parameters(
  model,
  ci = 0.95,
  df_method = "profile",
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  summary = FALSE,
  verbose = TRUE,
  ...
)

# S3 method for logitor
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = TRUE,
  robust = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for poissonmfx
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "marginal"),
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for betamfx
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "precision", "marginal"),
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

Arguments

model
Model object.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

bootstrap
Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also bootstrap_parameters()).

iterations
The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

standardize
The method used for standardizing the parameters. Can be "refit", "posthoc", "smart", "basic", "pseudo" or NULL (default) for no standardization. See 'Details' in standardize_parameters. Important: Categorical predictors (i.e. factors) are never standardized by default, which may be a different behaviour compared to other R packages or other software packages (like SPSS). If standardizing categorical predictors is desired, either use standardize="basic" to mimic behaviour of SPSS or packages such as lm.beta, or standardize the data with effectsize::standardize(force=TRUE) before fitting the model. Robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for logistic regression, or more generally speaking, for models with log or logit links. Note: Delta-method standard errors are also computed (by multiplying the standard errors by the transformed coefficients). This is to mimic behaviour of other software packages, such as Stata, but these standard errors poorly estimate uncertainty for the transformed coefficient. The transformed confidence interval more clearly captures this uncertainty. For compare_parameters(), exponentiate = "nongaussian" will only exponentiate coefficients from non-Gaussian families.

robust
Logical, if TRUE, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors. Additional arguments like vcov_estimation or vcov_type are passed down to other methods, see standard_error_robust() for details and this vignette for working examples.

p_adjust
Character vector, if not NULL, indicates the method to adjust p-values. See p.adjust for details. Further possible adjustment methods are "tukey", "scheffe", "sidak" and "none" to explicitly disable adjustment for emmGrid objects (from emmeans).

summary
Logical, if TRUE, prints summary information about the model (model formula, number of observations, residual standard deviation and more).

parameters
Character vector of length 1 with a regular expression pattern that describes the parameters that should be returned from the data frame, or a named list of regular expressions. All non-matching parameters will be removed from the output. If parameters is a character vector, every parameter in the "Parameters" column that matches the regular expression in parameters will be selected from the returned data frame. Furthermore, if parameters has more than one element, these will be merged into a regular expression pattern like this: "(one|two|three)". If parameters is a named list of regular expression patterns, the names of the list-element should equal the column name where selection should be applied. This is useful for model objects where model_parameters() returns multiple columns with parameter components, like in model_parameters.lavaan.

verbose
Toggle warnings and messages.

...
Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like type or parallel are passed down to bootstrap_model(), and arguments like ci_method are passed down to describe_posterior.

df_method
Method for computing degrees of freedom for confidence intervals (CI). Only applies to models of class glm or polr. May be "profile" or "wald".

component
Model component for which parameters should be shown. May be one of "conditional", "precision" (betareg), "scale" (ordinal), "extra" (glmx), "marginal" (mfx), "conditional" or "full" (for MuMIn::model.avg()) or "all".

Value

A data frame of indices related to the model's parameters.

See also

Examples

library(parameters)
model <- lm(mpg ~ wt + cyl, data = mtcars)

model_parameters(model)
#> Parameter   | Coefficient |   SE |         95% CI | t(29) |      p
#> ------------------------------------------------------------------
#> (Intercept) |       39.69 | 1.71 | [36.18, 43.19] | 23.14 | < .001
#> wt          |       -3.19 | 0.76 | [-4.74, -1.64] | -4.22 | < .001
#> cyl         |       -1.51 | 0.41 | [-2.36, -0.66] | -3.64 | 0.001 

# bootstrapped parameters
model_parameters(model, bootstrap = TRUE)
#> Parameter   | Coefficient |         95% CI |      p
#> ---------------------------------------------------
#> (Intercept) |       39.91 | [35.34, 43.89] | < .001
#> wt          |       -3.23 | [-4.85, -2.00] | 0.002 
#> cyl         |       -1.48 | [-2.26, -0.75] | 0.002 

# standardized parameters
model_parameters(model, standardize = "refit")
#> Parameter   | Coefficient |   SE |         95% CI |    t(29) |      p
#> ---------------------------------------------------------------------
#> (Intercept) |    4.90e-17 | 0.08 | [-0.15,  0.15] | 6.50e-16 | > .999
#> wt          |       -0.52 | 0.12 | [-0.77, -0.27] |    -4.22 | < .001
#> cyl         |       -0.45 | 0.12 | [-0.70, -0.20] |    -3.64 | 0.001 

# different p-value style in output
model_parameters(model, p_digits = 5)
#> Parameter   | Coefficient |   SE |         95% CI | t(29) |           p
#> -----------------------------------------------------------------------
#> (Intercept) |       39.69 | 1.71 | [36.18, 43.19] | 23.14 | 3.04318e-20
#> wt          |       -3.19 | 0.76 | [-4.74, -1.64] | -4.22 | 0.00022    
#> cyl         |       -1.51 | 0.41 | [-2.36, -0.66] | -3.64 | 0.00106    
model_parameters(model, digits = 3, ci_digits = 4, p_digits = "scientific")
#> Parameter   | Coefficient |    SE |             95% CI |  t(29) |           p
#> -----------------------------------------------------------------------------
#> (Intercept) |      39.686 | 1.715 | [36.1787, 43.1938] | 23.141 | 3.04318e-20
#> wt          |      -3.191 | 0.757 | [-4.7390, -1.6429] | -4.216 | 2.22020e-04
#> cyl         |      -1.508 | 0.415 | [-2.3559, -0.6597] | -3.636 | 1.06428e-03

# logistic regression model
model <- glm(vs ~ wt + cyl, data = mtcars, family = "binomial")
model_parameters(model)
#> Parameter   | Log-Odds |   SE |         95% CI |     z |     p
#> --------------------------------------------------------------
#> (Intercept) |    10.62 | 4.17 | [ 4.79, 22.66] |  2.55 | 0.011
#> wt          |     2.10 | 1.55 | [-0.53,  6.24] |  1.36 | 0.174
#> cyl         |    -2.93 | 1.38 | [-6.92, -1.07] | -2.12 | 0.034

# show odds ratio / exponentiated coefficients
model_parameters(model, exponentiate = TRUE)
#> Parameter   | Odds Ratio |       SE |             95% CI |     z |     p
#> ------------------------------------------------------------------------
#> (Intercept) |   40911.34 | 1.71e+05 | [120.16, 6.95e+09] |  2.55 | 0.011
#> wt          |       8.17 |    12.63 | [  0.59,   514.10] |  1.36 | 0.174
#> cyl         |       0.05 |     0.07 | [  0.00,     0.34] | -2.12 | 0.034