Parameters from (General) Linear Models

Extract and compute indices and measures to describe parameters of (general) linear models (GLMs).

# S3 method for default
model_parameters(
  model,
  ci = 0.95,
  ci_method = NULL,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  summary = FALSE,
  keep = NULL,
  drop = NULL,
  parameters = keep,
  verbose = TRUE,
  ...
)

# S3 method for glm
model_parameters(
  model,
  ci = 0.95,
  ci_method = NULL,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  summary = FALSE,
  verbose = TRUE,
  df_method = ci_method,
  ...
)

# S3 method for logitor
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = TRUE,
  robust = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for poissonmfx
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "marginal"),
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

# S3 method for betamfx
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "precision", "marginal"),
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

Arguments

model
Model object.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

ci_method
Method for computing degrees of freedom for confidence intervals (CI) and the related p-values. Allowed are following options (which vary depending on the model class): "residual", "normal", "likelihood", "satterthwaite", "kenward", "wald", "profile", "boot", "uniroot", "ml1", "betwithin", "hdi", "quantile", "ci", "eti", "si", "bci", or "bcai". See section Confidence intervals and approximation of degrees of freedom in model_parameters() for further details. When ci_method=NULL, in most cases "wald" is used then.

bootstrap
Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also bootstrap_parameters()).

iterations
The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

standardize
The method used for standardizing the parameters. Can be NULL (default; no standardization), "refit" (for re-fitting the model on standardized data) or one of "basic", "posthoc", "smart", "pseudo". See 'Details' in effectsize::standardize_parameters(). Important:

The "refit" method does not standardized categorical predictors (i.e. factors), which may be a different behaviour compared to other R packages (such as lm.beta) or other software packages (like SPSS). to mimic such behaviours, either use standardize="basic" or standardize the data with effectsize::standardize(force=TRUE) before fitting the model.

For mixed models, when using methods other than "refit", only the fixed effects will be returned.

Robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for logistic regression, or more generally speaking, for models with log or logit links. Note: Delta-method standard errors are also computed (by multiplying the standard errors by the transformed coefficients). This is to mimic behaviour of other software packages, such as Stata, but these standard errors poorly estimate uncertainty for the transformed coefficient. The transformed confidence interval more clearly captures this uncertainty. For compare_parameters(), exponentiate = "nongaussian" will only exponentiate coefficients from non-Gaussian families.

robust
Logical, if TRUE, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors. Additional arguments like vcov_estimation or vcov_type are passed down to other methods, see standard_error_robust() for details and this vignette for working examples.

p_adjust
Character vector, if not NULL, indicates the method to adjust p-values. See stats::p.adjust() for details. Further possible adjustment methods are "tukey", "scheffe", "sidak" and "none" to explicitly disable adjustment for emmGrid objects (from emmeans).

summary
Logical, if TRUE, prints summary information about the model (model formula, number of observations, residual standard deviation and more).

keep, drop
Character containing a regular expression pattern that describes the parameters that should be included (for keep) or excluded (for drop) in the returned data frame. keep may also be a named list of regular expressions. All non-matching parameters will be removed from the output. If keep is a character vector, every parameter name in the "Parameter" column that matches the regular expression in parameters will be selected from the returned data frame (and vice versa, all parameter names matching drop will be excluded). Furthermore, if keep has more than one element, these will be merged with an OR operator into a regular expression pattern like this: "(one|two|three)". If keep is a named list of regular expression patterns, the names of the list-element should equal the column name where selection should be applied. This is useful for model objects where model_parameters() returns multiple columns with parameter components, like in model_parameters.lavaan(). Note that the regular expression pattern should match the parameter names as they are stored in the returned data frame, which can be different from how they are printed. Inspect the $Parameter column of the parameters table to get the exact parameter names.

parameters
Deprecated, alias for keep.

verbose
Toggle warnings and messages.

...
Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like type or parallel are passed down to bootstrap_model(), and arguments like ci_method are passed down to bayestestR::describe_posterior().

df_method
Deprecated. Please use ci_method.

component
Model component for which parameters should be shown. May be one of "conditional", "precision" (betareg), "scale" (ordinal), "extra" (glmx), "marginal" (mfx), "conditional" or "full" (for MuMIn::model.avg()) or "all".

Value

A data frame of indices related to the model's parameters.

Confidence intervals and approximation of degrees of freedom

There are different ways of approximating the degrees of freedom depending on different assumptions about the nature of the model and its sampling distribution. The ci_method modulates the method for computing degrees of freedom (df) that are used to calculate confidence intervals (CI) and the related p-values. Following options are allowed, depending on the model class:

Classical methods:

Classical inference is generally based on the Wald method. The Wald approach to inference computes a test statistic by dividing the parameter estimate by its standard error (Coefficient / SE), then comparing this statistic against a t- or normal distribution. This approach can be used to compute CIs and p-values.

"wald"

Applies to non-Bayesian models

For linear models, CIs computed using the Wald method (SE and a t-distribution with residual df)

For linear models, p-values computed using the Wald method with a t-distribution with residual df

For other models, CIs computed using the Wald method (SE and a normal distribution)

For other models, p-values computed using the Wald method with a normal distribution

"normal"

Applies to non-Bayesian models

Compute Wald CIs and p values, but always use a normal distribution

CIs computed using the Wald method (SE and a normal distribution)

p-values computed using the Wald method with a normal distribution

"residual"

Applies to non-Bayesian models

Compute Wald CIs and p values, but always use a t-distribution with residual df when possible

CIs computed using the Wald method (SE and a t-distribution with residual df)

p-values computed using the Wald method with a t-distribution with residual df

If the residual df for a model cannot be determined, a normal distribution is used instead

Methods for mixed models:

Compared to fixed effects (or single-level) models, determining appropriate df for Wald-based inference in mixed models is more difficult. See the R GLMM FAQ for a discussion.

Several approximate methods for computing df are available, but you should also consider instead using profile likelihood ("profile") or bootstrap ("boot") CIs and p-values instead.

"satterthwaite"

Applies to linear mixed models

CIs computed using the Wald method (SE and a t-distribution with Satterthwaite df)

p-values computed using the Wald method with a t-distribution with Satterthwaite df

"kenward"

Applies to linear mixed models

CIs computed using the Wald method (Kenward-Roger SE and a t-distribution with Kenward-Roger df)

p-values computed using the Wald method with a Kenward-Roger SE and t-distribution with Kenward-Roger df

"ml1"

Applies to linear mixed models

CIs computed using the Wald method (SE and a t-distribution with m-l-1 approximated df)

p-values computed using the Wald method with a t-distribution with m-l-1 approximated df

See ci_ml1()

"betwithin"

Applies to linear mixed models and generalized linear mixed models

CIs computed using the Wald method (SE and a t-distribution with between-within df)

p-values computed using the Wald method with a t-distribution with between-within df

See ci_betwithin()

Likelihood-based methods:

Likelihood-based inference is based on comparing the likelihood for the maximum-likelihood estimate to the the likelihood for models with one or more parameter values changed (e.g., set to zero or a range of alternative values). Likelihood ratios for the maximum-likelihood and alternative models are compared to a \(\chi\)-squared distribution to compute CIs and p-values.

"profile"

Applies to non-Bayesian models of class glm, polr or glmmTMB

CIs computed by profiling the likelihood curve for a parameter, using linear interpolation to find where likelihood ratio equals a critical value

p-values computed using the Wald method with a normal-distribution) (note: this might change in a future update!)

"uniroot"

Applies to non-Bayesian models of class glmmTMB

CIs computed by profiling the likelihood curve for a parameter, using root finding to find where likelihood ratio equals a critical value;

p-values computed using the Wald method with a normal-distribution) (note: this might change in a future update!)

Methods for bootstrapped or Bayesian models:

Bootstrap-based inference is based on resampling and refitting the model to the resampled datasets. The distribution of parameter estimates across resampled datasets is used to approximate the parameter's sampling distribution. Depending on the type of model, several different methods for bootstrapping and constructing CIs and p-values from the bootstrap distribution are available.

For Bayesian models, inference is based on drawing samples from the model posterior distribution.

"quantile"

Applies to all models (including Bayesian models)

For non-Bayesian models, only applies if bootstrap = TRUE

CIs computed as equal tailed intervals using the quantiles of the bootstrap or posterior samples

p-values are based on the probability of direction

See bayestestR::eti()

"eti"

Same as "quantile"

"hdi"

Applies to all models (including Bayesian models)

For non-Bayesian models, only applies if bootstrap = TRUE

CIs computed as highest density intervals for the bootstrap or posterior samples

p-values are based on the probability of direction

See bayestestR::hdi()

"bci"

Applies to all models (including Bayesian models)

For non-Bayesian models, only applies if bootstrap = TRUE

CIs computed as bias corrected and accelerated intervals for the bootstrap or posterior samples

p-values are based on the probability of direction

See bayestestR::bci()

"bcai"

Same as "bci"

"si"

Applies to Bayesian models with proper priors

CIs computed as support intervals comparing the posterior samples against the prior samples

p-values are based on the probability of direction

See bayestestR::si()

"boot"

Applies to non-Bayesian models of class merMod

CIs computed using parametric bootstrapping (simulating data from the fitted model)

p-values computed using the Wald method with a normal-distribution) (note: this might change in a future update!)

For all iteration-based methods other than "boot" ("hdi", "quantile", "ci", "eti", "si", "bci", "bcai"), p-values are based on the probability of direction (bayestestR::p_direction()), which is converted into a p-value using bayestestR::pd_to_p().

See also

Examples

library(parameters)
model <- lm(mpg ~ wt + cyl, data = mtcars)

model_parameters(model)
#> Parameter   | Coefficient |   SE |         95% CI | t(29) |      p
#> ------------------------------------------------------------------
#> (Intercept) |       39.69 | 1.71 | [36.18, 43.19] | 23.14 | < .001
#> wt          |       -3.19 | 0.76 | [-4.74, -1.64] | -4.22 | < .001
#> cyl         |       -1.51 | 0.41 | [-2.36, -0.66] | -3.64 | 0.001 
#> 
#> Uncertainty intervals (equal-tailed) and p values (two-tailed) computed using a
#>   Wald t-distribution approximation.

# bootstrapped parameters
model_parameters(model, bootstrap = TRUE)
#> Parameter   | Coefficient |         95% CI |      p
#> ---------------------------------------------------
#> (Intercept) |       39.85 | [35.12, 43.99] | < .001
#> wt          |       -3.18 | [-4.86, -1.91] | < .001
#> cyl         |       -1.49 | [-2.22, -0.81] | < .001
#> 
#> Uncertainty intervals (equal-tailed) are naıve bootstrap intervals.

# standardized parameters
model_parameters(model, standardize = "refit")
#> Parameter   | Coefficient |   SE |         95% CI |    t(29) |      p
#> ---------------------------------------------------------------------
#> (Intercept) |    4.90e-17 | 0.08 | [-0.15,  0.15] | 6.50e-16 | > .999
#> wt          |       -0.52 | 0.12 | [-0.77, -0.27] |    -4.22 | < .001
#> cyl         |       -0.45 | 0.12 | [-0.70, -0.20] |    -3.64 | 0.001 
#> 
#> Uncertainty intervals (equal-tailed) and p values (two-tailed) computed using a
#>   Wald t-distribution approximation.

# different p-value style in output
model_parameters(model, p_digits = 5)
#> Parameter   | Coefficient |   SE |         95% CI | t(29) |           p
#> -----------------------------------------------------------------------
#> (Intercept) |       39.69 | 1.71 | [36.18, 43.19] | 23.14 | 3.04318e-20
#> wt          |       -3.19 | 0.76 | [-4.74, -1.64] | -4.22 | 0.00022    
#> cyl         |       -1.51 | 0.41 | [-2.36, -0.66] | -3.64 | 0.00106    
#> 
#> Uncertainty intervals (equal-tailed) and p values (two-tailed) computed using a
#>   Wald t-distribution approximation.
model_parameters(model, digits = 3, ci_digits = 4, p_digits = "scientific")
#> Parameter   | Coefficient |    SE |             95% CI |  t(29) |           p
#> -----------------------------------------------------------------------------
#> (Intercept) |      39.686 | 1.715 | [36.1787, 43.1938] | 23.141 | 3.04318e-20
#> wt          |      -3.191 | 0.757 | [-4.7390, -1.6429] | -4.216 | 2.22020e-04
#> cyl         |      -1.508 | 0.415 | [-2.3559, -0.6597] | -3.636 | 1.06428e-03
#> 
#> Uncertainty intervals (equal-tailed) and p values (two-tailed) computed using a
#>   Wald t-distribution approximation.

# logistic regression model
model <- glm(vs ~ wt + cyl, data = mtcars, family = "binomial")
model_parameters(model)
#> Parameter   | Log-Odds |   SE |         95% CI |     z |     p
#> --------------------------------------------------------------
#> (Intercept) |    10.62 | 4.17 | [ 4.79, 22.66] |  2.55 | 0.011
#> wt          |     2.10 | 1.55 | [-0.53,  6.24] |  1.36 | 0.174
#> cyl         |    -2.93 | 1.38 | [-6.92, -1.07] | -2.12 | 0.034
#> 
#> Uncertainty intervals (profile-likelihood) and p values (two-tailed) computed
#>   using a Wald z-distribution approximation.

# show odds ratio / exponentiated coefficients
model_parameters(model, exponentiate = TRUE)
#> Parameter   | Odds Ratio |       SE |             95% CI |     z |     p
#> ------------------------------------------------------------------------
#> (Intercept) |   40911.34 | 1.71e+05 | [120.16, 6.95e+09] |  2.55 | 0.011
#> wt          |       8.17 |    12.63 | [  0.59,   514.10] |  1.36 | 0.174
#> cyl         |       0.05 |     0.07 | [  0.00,     0.34] | -2.12 | 0.034
#> 
#> Uncertainty intervals (profile-likelihood) and p values (two-tailed) computed
#>   using a Wald z-distribution approximation.