Parameters from (General) Linear Models

Extract and compute indices and measures to describe parameters of (general) linear models (GLMs).

# S3 method for default
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  ...
)

# S3 method for betareg
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("conditional", "precision", "all"),
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  ...
)

# S3 method for clm2
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "scale"),
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  ...
)

# S3 method for glmx
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "extra"),
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  ...
)

Arguments

model
Model object.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

bootstrap
Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also bootstrap_parameters()).

iterations
The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

standardize
The method used for standardizing the parameters. Can be "refit", "posthoc", "smart", "basic" or NULL (default) for no standardization. See 'Details' in standardize_parameters. Note that robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for, say, logistic regressions, or more generally speaking: for models with log or logit link. Note: standard errors are also transformed (by multiplying the standard errors with the exponentiated coefficients), to mimic behaviour of other software packages, such as Stata.

robust
Logical, if TRUE, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors. Additional arguments like vcov_estimation or vcov_type are passed down to other methods, see standard_error_robust() for details.

p_adjust
Character vector, if not NULL, indicates the method to adjust p-values. See p.adjust for details.

...
Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like ci_method are passed down to describe_posterior.

component
Model component for which parameters should be shown. May be one of "conditional", "precision" (betareg), "scale" (ordinal), "extra" (glmx) or "all".

Value

A data frame of indices related to the model's parameters.

See also

Examples

library(parameters)
model <- lm(mpg ~ wt + cyl, data = mtcars)

model_parameters(model)

#> Parameter   | Coefficient |   SE |         95% CI |     t | df |      p
#> -----------------------------------------------------------------------
#> (Intercept) |       39.69 | 1.71 | [36.18, 43.19] | 23.14 | 29 | < .001
#> wt          |       -3.19 | 0.76 | [-4.74, -1.64] | -4.22 | 29 | < .001
#> cyl         |       -1.51 | 0.41 | [-2.36, -0.66] | -3.64 | 29 | 0.001 

# bootstrapped parameters
model_parameters(model, bootstrap = TRUE)

#> Parameter   | Coefficient |         95% CI |      p
#> ---------------------------------------------------
#> (Intercept) |       39.82 | [35.17, 44.06] | < .001
#> wt          |       -3.19 | [-4.93, -1.96] | < .001
#> cyl         |       -1.51 | [-2.25, -0.79] | < .001

# standardized parameters
model_parameters(model, standardize = "refit")

#> Parameter   | Coefficient |   SE |         95% CI |     t | df |      p
#> -----------------------------------------------------------------------
#> (Intercept) |        0.00 | 0.08 | [-0.15,  0.15] |  0.00 | 29 | 1.000 
#> wt          |       -0.52 | 0.12 | [-0.77, -0.27] | -4.22 | 29 | < .001
#> cyl         |       -0.45 | 0.12 | [-0.70, -0.20] | -3.64 | 29 | 0.001 

# different p-value style in output
model_parameters(model, p_digits = 5)

#> Parameter   | Coefficient |   SE |         95% CI |     t | df |       p
#> ------------------------------------------------------------------------
#> (Intercept) |       39.69 | 1.71 | [36.18, 43.19] | 23.14 | 29 | 0.00000
#> wt          |       -3.19 | 0.76 | [-4.74, -1.64] | -4.22 | 29 | 0.00022
#> cyl         |       -1.51 | 0.41 | [-2.36, -0.66] | -3.64 | 29 | 0.00106

model_parameters(model, digits = 3, ci_digits = 4, p_digits = "scientific")

#> Parameter   | Coefficient |    SE |             95% CI |      t | df |           p
#> ----------------------------------------------------------------------------------
#> (Intercept) |      39.686 | 1.715 | [36.1787, 43.1938] | 23.141 | 29 | 3.04318e-20
#> wt          |      -3.191 | 0.757 | [-4.7390, -1.6429] | -4.216 | 29 | 2.22020e-04
#> cyl         |      -1.508 | 0.415 | [-2.3559, -0.6597] | -3.636 | 29 | 1.06428e-03

# logistic regression model
model <- glm(vs ~ wt + cyl, data = mtcars, family = "binomial")
model_parameters(model)

#> Parameter   | Coefficient |   SE |         95% CI |     z | df |     p
#> ----------------------------------------------------------------------
#> (Intercept) |       10.62 | 4.17 | [ 4.79, 22.66] |  2.55 | 29 | 0.011
#> wt          |        2.10 | 1.55 | [-0.53,  6.24] |  1.36 | 29 | 0.174
#> cyl         |       -2.93 | 1.38 | [-6.92, -1.07] | -2.12 | 29 | 0.034

# show odds ratio / exponentiated coefficients
model_parameters(model, exponentiate = TRUE)

#> Parameter   | Coefficient |    SE |                  95% CI |     z | df |     p
#> --------------------------------------------------------------------------------
#> (Intercept) |    40911.34 | 64.58 | [120.16,   6.95162e+09] |  2.55 | 29 | 0.011
#> wt          |        8.17 |  4.69 | [  0.59,        514.10] |  1.36 | 29 | 0.174
#> cyl         |        0.05 |  3.97 | [  0.00,          0.34] | -2.12 | 29 | 0.034