Parameters from Mixed Models

Parameters from (linear) mixed models.

# S3 method for merMod
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  df_method = "wald",
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  details = FALSE,
  p_adjust = NULL,
  ...
)

# S3 method for glmmTMB
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "zi", "zero_inflated"),
  standardize = NULL,
  exponentiate = FALSE,
  df_method = NULL,
  details = FALSE,
  ...
)

# S3 method for mixor
model_parameters(
  model,
  ci = 0.95,
  effects = c("all", "fixed", "random"),
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  details = FALSE,
  ...
)

# S3 method for clmm
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  details = FALSE,
  df_method = NULL,
  ...
)

Arguments

model
A mixed model.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

bootstrap
Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also bootstrap_parameters()).

df_method
Method for computing degrees of freedom for p values, standard errors and confidence intervals (CI). May be "wald" (default, see degrees_of_freedom), "ml1" (see dof_ml1), "betwithin" (see dof_betwithin), "satterthwaite" (see dof_satterthwaite) or "kenward" (see dof_kenward). Note that when df_method is not "wald", robust standard errors etc. cannot be computed.

iterations
The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

standardize
The method used for standardizing the parameters. Can be "refit", "posthoc", "smart", "basic" or NULL (default) for no standardization. See 'Details' in standardize_parameters. Note that robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for, say, logistic regressions, or more generally speaking: for models with log or logit link. Note: standard errors are also transformed (by multiplying the standard errors with the exponentiated coefficients), to mimic behaviour of other software packages, such as Stata.

robust
Logical, if TRUE, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors. Additional arguments like vcov_estimation or vcov_type are passed down to other methods, see standard_error_robust() for details.

details
Logical, if TRUE, a summary of the random effects is included. See random_parameters for details.

p_adjust
Character vector, if not NULL, indicates the method to adjust p-values. See p.adjust for details.

...
Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like ci_method are passed down to describe_posterior.

component
Model component for which parameters should be shown. May be one of "conditional", "precision" (betareg), "scale" (ordinal), "extra" (glmx) or "all".

effects
Should parameters for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

Value

A data frame of indices related to the model's parameters.

See also

Examples

library(parameters)
if (require("lme4")) {
  data(mtcars)
  model <- lmer(mpg ~ wt + (1 | gear), data = mtcars)
  model_parameters(model)
}

#> Parameter   | Coefficient |   SE |         95% CI |     t | df |      p
#> -----------------------------------------------------------------------
#> (Intercept) |       36.19 | 2.19 | [31.90, 40.48] | 16.52 | 28 | < .001
#> wt          |       -5.05 | 0.64 | [-6.30, -3.79] | -7.89 | 28 | < .001

# \donttest{
if (require("glmmTMB")) {
  data(Salamanders)
  model <- glmmTMB(
    count ~ spp + mined + (1 | site),
    ziformula = ~mined,
    family = poisson(),
    data = Salamanders
  )
  model_parameters(model, details = TRUE)
}

#> # Fixed Effects
#> 
#> # Fixed Effects component
#> 
#> Parameter   | Coefficient |   SE |         95% CI |     z |  df |      p
#> ------------------------------------------------------------------------
#> (Intercept) |       -0.36 | 0.28 | [-0.90,  0.18] | -1.30 | 633 | 0.194 
#> spp [PR]    |       -1.27 | 0.24 | [-1.74, -0.80] | -5.27 | 633 | < .001
#> spp [DM]    |        0.27 | 0.14 | [ 0.00,  0.54] |  1.95 | 633 | 0.051 
#> spp [EC-A]  |       -0.57 | 0.21 | [-0.97, -0.16] | -2.75 | 633 | 0.006 
#> spp [EC-L]  |        0.67 | 0.13 | [ 0.41,  0.92] |  5.20 | 633 | < .001
#> spp [DES-L] |        0.63 | 0.13 | [ 0.38,  0.87] |  4.96 | 633 | < .001
#> spp [DF]    |        0.12 | 0.15 | [-0.17,  0.40] |  0.78 | 633 | 0.435 
#> mined [no]  |        1.27 | 0.27 | [ 0.74,  1.80] |  4.72 | 633 | < .001
#> 
#> # Zero-Inflated component
#> 
#> Parameter   | Coefficient |   SE |         95% CI |     z |  df |      p
#> ------------------------------------------------------------------------
#> (Intercept) |        0.79 | 0.27 | [ 0.26,  1.32] |  2.90 | 633 | 0.004 
#> mined [no]  |       -1.84 | 0.31 | [-2.46, -1.23] | -5.87 | 633 | < .001
#> 
#> 
#> # Random Effects
#> 
#> Within-Group Variance      0.51 (0.71)
#> Between-Group Variance
#>   Random Intercept (site)  0.11 (0.33)
#> N (groups per factor)
#>   site                       23
#> Observations                644

if (require("lme4")) {
  model <- lmer(mpg ~ wt + (1 | gear), data = mtcars)
  model_parameters(model, bootstrap = TRUE, iterations = 50)
}

#> Parameter   | Coefficient |         95% CI |      p
#> ---------------------------------------------------
#> (Intercept) |       35.85 | [31.90, 40.49] | < .001
#> wt          |       -4.93 | [-6.26, -3.52] | < .001

# }