Parameters from multiply imputed repeated analyses

Format models of class mira, obtained from mice::width.mids().

# S3 method for mira
model_parameters(model, ci = 0.95, exponentiate = FALSE, p_adjust = NULL, ...)

Arguments

model
An object of class mira.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for, say, logistic regressions, or more generally speaking: for models with log or logit link. Note: standard errors are also transformed (by multiplying the standard errors with the exponentiated coefficients), to mimic behaviour of other software packages, such as Stata.

p_adjust
Character vector, if not NULL, indicates the method to adjust p-values. See p.adjust for details.

...
Arguments passed to or from other methods.

Details

model_parameters() for objects of class mira works similar to mice::pool(), i.e. it generates the pooled summary of multiple imputed repeated regression analyses.

Examples

library(parameters)
if (require("mice")) {
  data(nhanes2)
  imp <- mice(nhanes2)
  fit <- with(data = imp, exp = lm(bmi ~ age + hyp + chl))
  model_parameters(fit)
}

#> Loading required package: mice

#> 
#> Attaching package: ‘mice’

#> The following objects are masked from ‘package:base’:
#> 
#>     cbind, rbind

#> 
#>  iter imp variable
#>   1   1  bmi  hyp  chl
#>   1   2  bmi  hyp  chl
#>   1   3  bmi  hyp  chl
#>   1   4  bmi  hyp  chl
#>   1   5  bmi  hyp  chl
#>   2   1  bmi  hyp  chl
#>   2   2  bmi  hyp  chl
#>   2   3  bmi  hyp  chl
#>   2   4  bmi  hyp  chl
#>   2   5  bmi  hyp  chl
#>   3   1  bmi  hyp  chl
#>   3   2  bmi  hyp  chl
#>   3   3  bmi  hyp  chl
#>   3   4  bmi  hyp  chl
#>   3   5  bmi  hyp  chl
#>   4   1  bmi  hyp  chl
#>   4   2  bmi  hyp  chl
#>   4   3  bmi  hyp  chl
#>   4   4  bmi  hyp  chl
#>   4   5  bmi  hyp  chl
#>   5   1  bmi  hyp  chl
#>   5   2  bmi  hyp  chl
#>   5   3  bmi  hyp  chl
#>   5   4  bmi  hyp  chl
#>   5   5  bmi  hyp  chl

#> Parameter   | Coefficient |   SE |          95% CI |     t |    df |      p
#> ---------------------------------------------------------------------------
#> (Intercept) |       18.26 | 3.41 | [ 11.10, 25.41] |  5.35 | 18.26 | < .001
#> age [40-59] |       -5.49 | 1.68 | [ -9.01, -1.97] | -3.27 | 18.26 | 0.004 
#> age [60-99] |       -7.16 | 2.47 | [-12.35, -1.98] | -2.90 | 18.26 | 0.009 
#> hyp [yes]   |        2.80 | 1.75 | [ -0.87,  6.46] |  1.60 | 18.26 | 0.126 
#> chl         |        0.06 | 0.02 | [  0.01,  0.10] |  2.85 | 18.26 | 0.011 

# model_parameters() also works for models that have no "tidy"-method in mice
if (require("mice") && require("gee")) {
  data(warpbreaks)
  set.seed(1234)
  warpbreaks$tension[sample(1:nrow(warpbreaks), size = 10)] <- NA
  imp <- mice(warpbreaks)
  fit <- with(data = imp, expr = gee(breaks ~ tension, id = wool))

  # does not work:
  # summary(pool(fit))

  model_parameters(fit)
}

#> Loading required package: gee

#> 
#>  iter imp variable
#>   1   1  tension
#>   1   2  tension
#>   1   3  tension
#>   1   4  tension
#>   1   5  tension
#>   2   1  tension
#>   2   2  tension
#>   2   3  tension
#>   2   4  tension
#>   2   5  tension
#>   3   1  tension
#>   3   2  tension
#>   3   3  tension
#>   3   4  tension
#>   3   5  tension
#>   4   1  tension
#>   4   2  tension
#>   4   3  tension
#>   4   4  tension
#>   4   5  tension
#>   5   1  tension
#>   5   2  tension
#>   5   3  tension
#>   5   4  tension
#>   5   5  tension

#> Beginning Cgee S-function, @(#) geeformula.q 4.13 98/01/27

#> running glm to get initial regression estimate

#> (Intercept)    tensionM    tensionH 
#>    36.04762   -12.26984   -13.71429 

#> Beginning Cgee S-function, @(#) geeformula.q 4.13 98/01/27

#> running glm to get initial regression estimate

#> (Intercept)    tensionM    tensionH 
#>    35.04545   -10.29545   -12.98295 

#> Beginning Cgee S-function, @(#) geeformula.q 4.13 98/01/27

#> running glm to get initial regression estimate

#> (Intercept)    tensionM    tensionH 
#>   35.150000   -8.973529  -13.267647 

#> Beginning Cgee S-function, @(#) geeformula.q 4.13 98/01/27

#> running glm to get initial regression estimate

#> (Intercept)    tensionM    tensionH 
#>    36.66667   -13.26667   -14.50000 

#> Beginning Cgee S-function, @(#) geeformula.q 4.13 98/01/27

#> running glm to get initial regression estimate

#> (Intercept)    tensionM    tensionH 
#>    36.15000   -11.37222   -14.21250 
#> Parameters can't be retrieved for objects of class 'mira'.
#> 

#> Parameter   | Coefficient |   SE |          95% CI |     z |      p
#> -------------------------------------------------------------------
#> (Intercept) |       35.81 | 2.71 | [ 30.50, 41.12] | 13.21 | < .001
#> tensionM    |      -11.24 | 4.31 | [-19.67, -2.80] | -2.61 | 0.009 
#> tensionH    |      -13.74 | 3.98 | [-21.54, -5.93] | -3.45 | < .001