Parameters from multiply imputed repeated analyses

Format models of class mira, obtained from mice::width.mids().

# S3 method for mira
model_parameters(
  model,
  ci = 0.95,
  exponentiate = FALSE,
  p_adjust = NULL,
  verbose = TRUE,
  ...
)

Arguments

model
An object of class mira.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for, say, logistic regressions, or more generally speaking: for models with log or logit link. Note: standard errors are also transformed (by multiplying the standard errors with the exponentiated coefficients), to mimic behaviour of other software packages, such as Stata. For compare_parameters(), exponentiate = "nongaussian" will only exponentiate coefficients for all models except those from Gaussian family.

p_adjust
Character vector, if not NULL, indicates the method to adjust p-values. See p.adjust for details.

verbose
Toggle warnings and messages.

...
Arguments passed to or from other methods.

Details

model_parameters() for objects of class mira works similar to summary(mice::pool()), i.e. it generates the pooled summary of multiple imputed repeated regression analyses.

Examples

library(parameters)
if (require("mice")) {
  data(nhanes2)
  imp <- mice(nhanes2)
  fit <- with(data = imp, exp = lm(bmi ~ age + hyp + chl))
  model_parameters(fit)
}

#> Loading required package: mice

#> 
#> Attaching package: ‘mice’

#> The following object is masked from ‘package:stats’:
#> 
#>     filter

#> The following objects are masked from ‘package:base’:
#> 
#>     cbind, rbind

#> 
#>  iter imp variable
#>   1   1  bmi  hyp  chl
#>   1   2  bmi  hyp  chl
#>   1   3  bmi  hyp  chl
#>   1   4  bmi  hyp  chl
#>   1   5  bmi  hyp  chl
#>   2   1  bmi  hyp  chl
#>   2   2  bmi  hyp  chl
#>   2   3  bmi  hyp  chl
#>   2   4  bmi  hyp  chl
#>   2   5  bmi  hyp  chl
#>   3   1  bmi  hyp  chl
#>   3   2  bmi  hyp  chl
#>   3   3  bmi  hyp  chl
#>   3   4  bmi  hyp  chl
#>   3   5  bmi  hyp  chl
#>   4   1  bmi  hyp  chl
#>   4   2  bmi  hyp  chl
#>   4   3  bmi  hyp  chl
#>   4   4  bmi  hyp  chl
#>   4   5  bmi  hyp  chl
#>   5   1  bmi  hyp  chl
#>   5   2  bmi  hyp  chl
#>   5   3  bmi  hyp  chl
#>   5   4  bmi  hyp  chl
#>   5   5  bmi  hyp  chl

#> Parameter   | Coefficient |   SE |          95% CI | Statistic |    df |      p
#> -------------------------------------------------------------------------------
#> (Intercept) |       18.06 | 3.24 | [ 11.21, 24.91] |      5.57 | 16.56 | < .001
#> age40-59    |       -5.34 | 1.87 | [ -9.40, -1.28] |     -2.85 | 12.60 | 0.014 
#> age60-99    |       -7.59 | 2.34 | [-12.96, -2.22] |     -3.24 |  8.27 | 0.011 
#> hypyes      |        2.36 | 2.00 | [ -2.28,  7.01] |      1.18 | 16.60 | 0.273 
#> chl         |        0.06 | 0.02 | [  0.02,  0.10] |      3.28 |  7.65 | 0.005 

if (FALSE) {
# model_parameters() also works for models that have no "tidy"-method in mice
if (require("mice") && require("gee")) {
  data(warpbreaks)
  set.seed(1234)
  warpbreaks$tension[sample(1:nrow(warpbreaks), size = 10)] <- NA
  imp <- mice(warpbreaks)
  fit <- with(data = imp, expr = gee(breaks ~ tension, id = wool))

  # does not work:
  # summary(pool(fit))

  model_parameters(fit)
}
}

# and it works with pooled results
if (require("mice")) {
  data("nhanes2")
  imp <- mice(nhanes2)
  fit <- with(data = imp, exp = lm(bmi ~ age + hyp + chl))
  pooled <- pool(fit)

  model_parameters(pooled)
}

#> 
#>  iter imp variable
#>   1   1  bmi  hyp  chl
#>   1   2  bmi  hyp  chl
#>   1   3  bmi  hyp  chl
#>   1   4  bmi  hyp  chl
#>   1   5  bmi  hyp  chl
#>   2   1  bmi  hyp  chl
#>   2   2  bmi  hyp  chl
#>   2   3  bmi  hyp  chl
#>   2   4  bmi  hyp  chl
#>   2   5  bmi  hyp  chl
#>   3   1  bmi  hyp  chl
#>   3   2  bmi  hyp  chl
#>   3   3  bmi  hyp  chl
#>   3   4  bmi  hyp  chl
#>   3   5  bmi  hyp  chl
#>   4   1  bmi  hyp  chl
#>   4   2  bmi  hyp  chl
#>   4   3  bmi  hyp  chl
#>   4   4  bmi  hyp  chl
#>   4   5  bmi  hyp  chl
#>   5   1  bmi  hyp  chl
#>   5   2  bmi  hyp  chl
#>   5   3  bmi  hyp  chl
#>   5   4  bmi  hyp  chl
#>   5   5  bmi  hyp  chl

#> Parameter   | Coefficient |   SE |          95% CI | Statistic |    df |      p
#> -------------------------------------------------------------------------------
#> (Intercept) |       19.89 | 3.47 | [ 12.44, 27.33] |      5.74 | 13.78 | < .001
#> age40-59    |       -5.39 | 2.26 | [-10.78,  0.01] |     -2.38 |  6.67 | 0.050 
#> age60-99    |       -6.99 | 2.74 | [-13.95, -0.02] |     -2.55 |  5.16 | 0.050 
#> hypyes      |        2.23 | 2.03 | [ -2.26,  6.71] |      1.10 |  9.83 | 0.297 
#> chl         |        0.05 | 0.02 | [  0.00,  0.09] |      2.38 | 10.80 | 0.039