Parameters from multinomial or cumulative link models

Parameters from multinomial or cumulative link models

# S3 method for mlm
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  ...
)

# S3 method for multinom
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  ...
)

# S3 method for bracl
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  standardize = NULL,
  exponentiate = FALSE,
  p_adjust = NULL,
  ...
)

# S3 method for DirichletRegModel
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "precision"),
  standardize = NULL,
  exponentiate = FALSE,
  ...
)

Arguments

model
A model with multinomial or categorical response value.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

bootstrap
Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also bootstrap_parameters()).

iterations
The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

standardize
The method used for standardizing the parameters. Can be "refit", "posthoc", "smart", "basic" or NULL (default) for no standardization. See 'Details' in standardize_parameters. Note that robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for, say, logistic regressions, or more generally speaking: for models with log or logit link. Note: standard errors are also transformed (by multiplying the standard errors with the exponentiated coefficients), to mimic behaviour of other software packages, such as Stata.

p_adjust
Character vector, if not NULL, indicates the method to adjust p-values. See p.adjust for details.

...
Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like ci_method are passed down to describe_posterior.

component
Model component for which parameters should be shown. May be one of "conditional", "precision" (betareg), "scale" (ordinal), "extra" (glmx) or "all".

Value

A data frame of indices related to the model's parameters.

Details

Multinomial or cumulative link models, i.e. models where the response value (dependent variable) is categorical and has more than two levels, usually return coefficients for each response level. Hence, the output from model_parameters() will split the coefficient tables by the different levels of the model's response.

See also

Examples

library(parameters)
if (require("brglm2")) {
  data("stemcell")
  model <- bracl(
    research ~ as.numeric(religion) + gender,
    weights = frequency,
    data = stemcell,
    type = "ML"
  )
  model_parameters(model)
}

#> Loading required package: brglm2

#> # Response level: definitely
#> 
#> Parameter       | Coefficient |   SE |         95% CI |     z | df |      p
#> ---------------------------------------------------------------------------
#> (Intercept)     |       -1.25 | 0.26 | [-1.76, -0.73] | -4.76 | 63 | < .001
#> religion        |        0.44 | 0.10 | [ 0.23,  0.64] |  4.20 | 63 | < .001
#> gender [female] |       -0.14 | 0.17 | [-0.47,  0.19] | -0.82 | 63 | 0.414 
#> 
#> # Response level: probably
#> 
#> Parameter       | Coefficient |   SE |        95% CI |    z | df |     p
#> ------------------------------------------------------------------------
#> (Intercept)     |        0.47 | 0.29 | [-0.10, 1.04] | 1.62 | 63 | 0.105
#> religion        |        0.26 | 0.13 | [ 0.01, 0.51] | 2.01 | 63 | 0.044
#> gender [female] |        0.19 | 0.21 | [-0.22, 0.60] | 0.90 | 63 | 0.370
#> 
#> # Response level: probably not
#> 
#> Parameter       | Coefficient |   SE |        95% CI |     z | df |     p
#> -------------------------------------------------------------------------
#> (Intercept)     |        0.43 | 0.39 | [-0.33, 1.18] |  1.11 | 63 | 0.268
#> religion        |        0.01 | 0.17 | [-0.33, 0.35] |  0.07 | 63 | 0.945
#> gender [female] |       -0.16 | 0.28 | [-0.71, 0.39] | -0.57 | 63 | 0.566
#>