Format PCA/FA from the psych package

Format PCA/FA objects from the psych package (Revelle, 2016).

# S3 method for principal
model_parameters(model, sort = FALSE,
  threshold = NULL, ...)

Arguments

model
PCA or FA created by the psych::principal or psych::fa functions.

sort
Sort the loadings.

threshold
A value between 0 and 1 indicates which (absolute) values from the loadings should be removed. Can also be "max", in which case it will only display the maximum loading per veriable (the most simple structure).

...
Arguments passed to or from other methods.

Details

Complexity (Hoffman's, 1978; Pettersson and Turkheimer, 2010) represents the number of latent components needed to account for the observed variables. Whereas a perfect simple structure solution has a complexity of 1 in that each item would only load on one factor, a solution with evenly distributed items has a complexity greater than 1.

Uniqueness represents the variance that is 'unique' to the variable and not shared with other variables. It is equal to 1 – communality (variance that is shared with other variables). A uniqueness of 0.20 suggests that 20% or that variable is not shared with other variables in the overall factor model. The greater 'uniqueness' the lower the relevance of the variable in the factor model.

References

Pettersson, E., \& Turkheimer, E. (2010). Item selection, evaluation, and simple structure in personality data. Journal of research in personality, 44(4), 407-420.

Revelle, W. (2016). How To: Use the psych package for Factor Analysis and data reduction.

Examples

library(parameters)
# NOT RUN {
library(psych)

# Principal Component Analysis (PCA) ---------
pca <- psych::principal(attitude)
model_parameters(pca)

pca <- psych::principal(attitude, nfactors = 3, rotate = "none")
model_parameters(pca, sort = TRUE, threshold = 0.2)
# }
# Note that the latter is identical to the 'principal_components' function available in parameters:
principal_components(attitude, n = 3, sort = TRUE, threshold = 0.2)

#> The 3 principal components accounted for 81.49% of the total variance of the original data (PC1 = 53.09%, PC2 = 16.30%, PC3 = 12.10%).
#> 
#>     Variable       PC1        PC2        PC3 Complexity
#> 1     raises 0.8619753         NA         NA   1.102249
#> 2 complaints 0.8492648 -0.3563602  0.2076599   1.476308
#> 3   learning 0.8261764         NA -0.2994027   1.265768
#> 4     rating 0.7961866 -0.4239721  0.2424866   1.739422
#> 5 privileges 0.6839314         NA         NA   1.175153
#> 6    advance 0.5832712  0.6086213 -0.4562244   2.849165
#> 7   critical 0.3573175  0.6436979  0.6450449   2.542072
# NOT RUN {
# Exploratory Factor Analysis (EFA) ---------
efa <- psych::fa(attitude, nfactors = 3)
model_parameters(efa, threshold = "max", sort = TRUE)
# }