Bayesian Models Parameters

Parameters of Bayesian models.

# S3 method for stanreg
model_parameters(model, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi", test = c("pd",
  "rope"), rope_range = "default", rope_ci = 1, bf_prior = NULL,
  diagnostic = c("ESS", "Rhat"), priors = TRUE, standardize = FALSE,
  standardize_robust = FALSE, iterations = 1000, ...)

Arguments

model
Bayesian model.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

dispersion
Logical, if TRUE, computes indices of dispersion related to the estimate(s) (SD and MAD for mean and median, respectively).

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

ci_method
The type of index used for Credible Interval. Can be "hdi" (default, see hdi) or "quantile" (see ci).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests.#' For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

rope_range
ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.1, 0.1)) or "default". If "default", the bounds are set to x +- 0.1*SD(response).

rope_ci
The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use for the percentage in ROPE.

bf_prior
Distribution representing a prior for the computation of Bayes factors. Used if the input is a posterior, otherwise (in the case of models) ignored.

diagnostic
Diagnostic metrics to compute. Character (vector) or list with one or more of these options: "ESS", "Rhat", "MCSE" or "all".

priors
Add the prior used for each parameter.

standardize
Add standardized parameters. Can be FALSE or a character indicating the standardization method (see parameters_standardize).

standardize_robust
Robust standardization. See parameters_standardize.

iterations
The number of bootstrap replicates. This only apply in the case of bootsrapped frequentist models.

...
Arguments passed to or from other methods (e.g., to standardize).

Examples

# NOT RUN {
library(rstanarm)
model <- stan_glm(Sepal.Length ~ Species, data = iris)
model_parameters(model, standardize = "full")

library(brms)
model <- brm(Sepal.Length ~ Species, data = iris)
model_parameters(model)
# }