Parameters from Zero-Inflated Models

Parameters from zero-inflated models.

# S3 method for zeroinfl
model_parameters(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "zi", "zero_inflated"),
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  ...
)

Arguments

model
A model with zero-inflation component.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

bootstrap
Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also bootstrap_parameters()).

iterations
The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

component
Model component for which parameters should be shown. May be one of "conditional", "precision" (betareg), "scale" (ordinal), "extra" (glmx) or "all".

standardize
The method used for standardizing the parameters. Can be "refit", "posthoc", "smart", "basic" or NULL (default) for no standardization. See 'Details' in standardize_parameters. Note that robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for, say, logistic regressions, or more generally speaking: for models with log or logit link. Note: standard errors are also transformed (by multiplying the standard errors with the exponentiated coefficients), to mimic behaviour of other software packages, such as Stata.

robust
Logical, if TRUE, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors. Additional arguments like vcov_estimation or vcov_type are passed down to other methods, see standard_error_robust() for details.

p_adjust
Character vector, if not NULL, indicates the method to adjust p-values. See p.adjust for details.

...
Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like ci_method are passed down to describe_posterior.

Value

A data frame of indices related to the model's parameters.

See also

Examples

library(parameters)
if (require("pscl")) {
  data("bioChemists")
  model <- zeroinfl(art ~ fem + mar + kid5 + ment | kid5 + phd, data = bioChemists)
  model_parameters(model)
}

#> Loading required package: pscl

#> Classes and Methods for R developed in the
#> Political Science Computational Laboratory
#> Department of Political Science
#> Stanford University
#> Simon Jackman
#> hurdle and zeroinfl functions by Achim Zeileis

#> # Fixed Effects component
#> 
#> Parameter     | Coefficient |   SE |         95% CI |     z |  df |      p
#> --------------------------------------------------------------------------
#> (Intercept)   |        0.56 | 0.07 | [ 0.43,  0.69] |  8.26 | 907 | < .001
#> fem [Women]   |       -0.23 | 0.06 | [-0.34, -0.11] | -3.91 | 907 | < .001
#> mar [Married] |        0.14 | 0.07 | [ 0.01,  0.27] |  2.07 | 907 | 0.038 
#> kid5          |       -0.17 | 0.05 | [-0.26, -0.07] | -3.43 | 907 | < .001
#> ment          |        0.02 | 0.00 | [ 0.02,  0.03] | 10.05 | 907 | < .001
#> 
#> # Zero-Inflated component
#> 
#> Parameter   | Coefficient |   SE |         95% CI |     z |  df |     p
#> -----------------------------------------------------------------------
#> (Intercept) |       -0.93 | 0.43 | [-1.78, -0.08] | -2.14 | 907 | 0.032
#> kid5        |        0.05 | 0.22 | [-0.38,  0.47] |  0.21 | 907 | 0.831
#> phd         |       -0.25 | 0.14 | [-0.51,  0.02] | -1.84 | 907 | 0.065
#>