Simulated draws from model coefficients

Simulate draws a statistical model n times to return a data.frame of estimates.

model_simulate(model, n_sims = 1000, ...)

# S3 method for glmmTMB
model_simulate(model, n_sims = 1000,
  component = c("all", "conditional", "zi", "zero_inflated"), ...)

Arguments

model
Statistical model.

n_sims
The number of simulation draws to create.

...
Arguments passed to or from other methods.

component
Should all parameters, parameters for the conditional model, or for the zero-inflated part of the model be returned? Applies to models with zero-inflated component. component may be one of "conditional", "zi", "zero-inflated" or "all" (default). May be abbreviated.

Value

A data.frame.

Details

model_simulate() is a computationally faster alternative to model_bootstrap(). Simulated draws for coefficients are based on a multivariate normal distribution (MASS::mvrnorm() with mean coef(model) and variance vcov(model).

See also

Examples

library(parameters)

model <- lm(Sepal.Length ~ Species * Petal.Width + Petal.Length, data = iris)
head(model_simulate(model))

#>   (Intercept) Speciesversicolor Speciesvirginica Petal.Width Petal.Length
#> 1    3.490212        -1.3821275        -2.175870   0.4619212    0.9495768
#> 2    3.809263        -1.6443443        -2.357484   0.1530034    0.8738285
#> 3    3.535979        -0.5986886        -2.571125   0.8282128    0.8782181
#> 4    3.452466        -0.9171241        -1.590942   0.6519571    0.9361557
#> 5    3.566214        -1.3826144        -1.598236   0.1875140    0.8911452
#> 6    3.729271        -1.2732424        -2.495559  -0.3373356    0.9359377
#>   Speciesversicolor:Petal.Width Speciesvirginica:Petal.Width
#> 1                    -0.6480091                  -0.44946886
#> 2                    -0.1400418                  -0.02490191
#> 3                    -1.3360465                  -0.44865600
#> 4                    -1.1155244                  -0.89753792
#> 5                    -0.2319216                  -0.33975393
#> 6                    -0.1188659                   0.42877840