Simulated draws from model coefficients

Simulate draws from a statistical model to return a data.frame of estimates.

model_simulate(model, iterations = 1000, ...)

# S3 method for glmmTMB
model_simulate(model, iterations = 1000,
  component = c("all", "conditional", "zi", "zero_inflated"), ...)

Arguments

model
Statistical model (no Bayesian models).

iterations
The number of draws to simulate/bootstrap.

...
Arguments passed to or from other methods.

component
Should all parameters, parameters for the conditional model, or for the zero-inflated part of the model be returned? Applies to models with zero-inflated component. component may be one of "conditional", "zi", "zero-inflated" or "all" (default). May be abbreviated.

Value

A data frame.

Details

Technical Details

model_simulate() is a computationally faster alternative to model_bootstrap(). Simulated draws for coefficients are based on a multivariate normal distribution (MASS::mvrnorm()) with mean mu = coef(model) and variance Sigma = vcov(model).

Models with Zero-Inflation Component

For models from packages glmmTMB, pscl, GLMMadaptive and countreg, the component argument can be used to specify which parameters should be simulated. For all other models, parameters from the conditional component (fixed effects) are simulated. This may include smooth terms, but not random effects.

See also

Examples

library(parameters)
library(glmmTMB)

model <- lm(Sepal.Length ~ Species * Petal.Width + Petal.Length, data = iris)
head(model_simulate(model))

#>   (Intercept) Speciesversicolor Speciesvirginica Petal.Width Petal.Length
#> 1    3.308727        -0.6387331        -1.840927  0.94266501    0.9470016
#> 2    3.521971        -1.2831925        -1.791598  0.72667515    0.8774803
#> 3    3.444105        -1.3885282        -2.417014  0.01642164    1.0377796
#> 4    3.387368        -1.2203630        -2.483280  0.45719124    1.0107853
#> 5    3.203380        -0.6703644        -2.085831  1.53968442    0.9767090
#> 6    3.531947        -0.8260595        -2.430570  0.24200952    0.9206217
#>   Speciesversicolor:Petal.Width Speciesvirginica:Petal.Width
#> 1                    -1.5596791                  -1.00068844
#> 2                    -0.7984172                  -0.70368837
#> 3                    -0.3411097                  -0.08857689
#> 4                    -0.8414908                  -0.41697867
#> 5                    -2.1571169                  -1.53455671
#> 6                    -0.7493993                  -0.07573117

model <- glmmTMB(
  count ~ spp + mined + (1 | site),
  ziformula = ~mined,
  family = poisson(),
  data = Salamanders
)
head(model_simulate(model))

#>   (Intercept)      sppPR     sppDM    sppEC-A   sppEC-L  sppDES-L      sppDF
#> 1 -0.62585809 -1.0861403 0.4518512 -0.5144004 0.9109420 0.8144825 0.01697147
#> 2 -0.62362943 -0.8791991 0.4132972 -0.1616461 0.8517232 0.8238346 0.15985582
#> 3 -0.40082321 -0.9449677 0.4874868 -0.4567502 0.7975132 0.8055994 0.28576736
#> 4  0.09584125 -0.9658859 0.2107973 -0.3845486 0.7282838 0.6185811 0.07465655
#> 5 -0.15879481 -0.9701387 0.4457972 -0.4228344 0.7701103 0.6931179 0.01904309
#> 6 -0.54573026 -1.3562299 0.1682816 -0.5530070 0.5694093 0.5727591 0.21732044
#>     minedno (Intercept)_zi minedno_zi
#> 1 1.4448942      0.5015027  -1.345750
#> 2 1.4541614      0.8639666  -1.955520
#> 3 1.0674028      0.6575522  -1.693844
#> 4 0.7850103      0.5192618  -1.505744
#> 5 1.0564476      0.8970211  -1.807129
#> 6 1.6319444      0.0484331  -1.366121

head(model_simulate(model, component = "zero_inflated"))

#>   (Intercept)   minedno
#> 1   0.7328976 -2.009657
#> 2   0.6803349 -1.669484
#> 3   0.8078101 -1.885715
#> 4   0.6757801 -1.663919
#> 5   0.7259168 -1.849500
#> 6   0.9124648 -1.995910