Number of Components/Factors to Retain

This function runs many existing procedures for determining how many factors to retain for your factor analysis (FA) or dimension reduction (PCA). It returns the number of factors based on the maximum consensus. In case of ties, it will select the solution with the less factors.

n_factors(x, type = "FA", rotation = "varimax",
  algorithm = "default", package = "all", safe = TRUE, ...)

Arguments

x
A dataframe.

type
Can be "FA" or "PCA", depending on what you want to do.

rotation
Only used for VSS (Very Simple Structure criterion, see VSS. The rotation to apply. Can be "none", "varimax", "quartimax", "bentlerT", "equamax", "varimin", "geominT" and "bifactor" for orthogonal rotations, and "promax", "oblimin", "simplimax", "bentlerQ", "geominQ", "biquartimin" and "cluster" for oblique transformations.

algorithm
Factoring method used by VSS. Can be "pa" for Principal Axis Factor Analysis, "minres" for minimum residual (OLS) factoring, "mle" for Maximum Likelihood FA and "pc" for Principal Components. "default" will select "minres" if type = "FA" and "pc" if type = "PCA".

package
Can be "all" or a vector containing "nFactors", "EGAnet" and "psych". These are the packages from which methods are used.

safe
If TRUE, will run all the procedures in try blocks, and will only return those that work and silently skip the ones that may fail.

...
Arguments passed to or from other methods.

Value

A data.frame.

References

Bartlett, M. S. (1950). Tests of significance in factor analysis. British Journal of statistical psychology, 3(2), 77-85.

Bentler, P. M., & Yuan, K. H. (1996). Test of linear trend in eigenvalues of a covariance matrix with application to data analysis. British Journal of Mathematical and Statistical Psychology, 49(2), 299-312.

Cattell, R. B. (1966). The scree test for the number of factors. Multivariate behavioral research, 1(2), 245-276.

Zoski, K. W., & Jurs, S. (1996). An objective counterpart to the visual scree test for factor analysis: The standard error scree. Educational and Psychological Measurement, 56(3), 443-451.

Zoski, K., & Jurs, S. (1993). Using multiple regression to determine the number of factors to retain in factor analysis. Multiple Linear Regression Viewpoints, 20(1), 5-9.

Nasser, F., Benson, J., & Wisenbaker, J. (2002). The performance of regression-based variations of the visual scree for determining the number of common factors. Educational and psychological measurement, 62(3), 397-419.

Golino, H., Shi, D., Garrido, L. E., Christensen, A. P., Nieto, M. D., Sadana, R., & Thiyagarajan, J. A. (2018). Investigating the performance of Exploratory Graph Analysis and traditional techniques to identify the number of latent factors: A simulation and tutorial.

Golino, H. F., & Epskamp, S. (2017). Exploratory graph analysis: A new approach for estimating the number of dimensions in psychological research. PloS one, 12(6), e0174035.

Revelle, W., & Rocklin, T. (1979). Very simple structure: An alternative procedure for estimating the optimal number of interpretable factors. Multivariate Behavioral Research, 14(4), 403-414.

Velicer, W. F. (1976). Determining the number of components from the matrix of partial correlations. Psychometrika, 41(3), 321-327.

Examples

library(parameters)

n_factors(mtcars, type = "PCA", package = "nFactors")

#> # Method Agreement Procedure:
#> 
#> The choice of 3 dimensions is supported by 4 (28.57%) methods out of 14 (Bartlett, CNG, SE Scree, R2).

result <- n_factors(mtcars[, 1:5], type = "FA", package = "psych")
as.numeric(result)

#> [1] 1

summary(result)

#>   n_Factors n_Methods
#> 1         1         3
#> 2         2         1
#> 3         3         1

n_factors(mtcars, type = "PCA")

#> # Method Agreement Procedure:
#> 
#> The choice of 2 dimensions is supported by 5 (31.25%) methods out of 16 (Optimal coordinates, Parallel analysis, Kaiser criterion, EGA (glasso), EGA (TMFG)).

n_factors(mtcars, type = "FA", algorithm = "mle")

#> # Method Agreement Procedure:
#> 
#> The choice of 2 dimensions is supported by 6 (28.57%) methods out of 21 (Optimal coordinates, Parallel analysis, Kaiser criterion, EGA (glasso), EGA (TMFG), VSS complexity 2).