Satterthwaite approximation for SEs, CIs and p-values

An approximate F-test based on the Satterthwaite (1946) approach.

ci_satterthwaite(model, ci = 0.95)

dof_satterthwaite(model)

p_value_satterthwaite(model, dof = NULL)

se_satterthwaite(model)

Arguments

model
A statistical model.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

dof
Degrees of Freedom.

Value

A data frame.

Details

Inferential statistics (like p-values, confidence intervals and standard errors) may be biased in mixed models when the number of clusters is small (even if the sample size of level-1 units is high). In such cases it is recommended to approximate a more accurate number of degrees of freedom for such inferential statitics. Unlike simpler approximation heuristics like the "m-l-1" rule (dof_ml1), the Satterthwaite approximation is also applicable in more complex multilevel designs. However, the "m-l-1" heuristic also applies to generalized mixed models, while approaches like Kenward-Roger or Satterthwaite are limited to linear mixed models only.

References

Satterthwaite FE (1946) An approximate distribution of estimates of variance components. Biometrics Bulletin 2 (6):110–4.

See also

Examples

# \donttest{
if (require("lme4")) {
  model <- lmer(Petal.Length ~ Sepal.Length + (1 | Species), data = iris)
  p_value_satterthwaite(model)
}

#>      Parameter            p
#> 1  (Intercept) 9.605150e-01
#> 2 Sepal.Length 7.881906e-29

# }