Wald-test approximation for CIs and p-values

The Wald-test approximation treats t-values as Wald z. Since the t distribution converges to the z distribution as degrees of freedom increase, this is like assuming infinite degrees of freedom. While this is unambiguously anti-conservative, this approximation appears as reasonable for reasonable sample sizes (Barr et al., 2013). That is, if we take the p-value to measure the probability of a false positive, this approximation produces a higher false positive rate than the nominal 5% at p = 0.05.

ci_wald(
  model,
  ci = 0.95,
  dof = NULL,
  effects = c("fixed", "random", "all"),
  component = c("all", "conditional", "zi", "zero_inflated", "precision", "scale",
    "smooth_terms"),
  robust = FALSE,
  ...
)

p_value_wald(model, ...)

# S3 method for merMod
p_value_wald(model, dof = Inf, ...)

Arguments

model
A statistical model.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

dof
Degrees of Freedom. If not specified, for ci_wald(), defaults to model's residual degrees of freedom (i.e. n-k, where n is the number of observations and k is the number of parameters). For p_value_wald(), defaults to Inf.

effects
Should standard errors for fixed effects or random effects be returned? Only applies to mixed models. May be abbreviated. When standard errors for random effects are requested, for each grouping factor a list of standard errors (per group level) for random intercepts and slopes is returned.

component
Should all parameters, parameters for the conditional model, or for the zero-inflated part of the model be returned? Applies to models with zero-inflated component. component may be one of "conditional", "zi", "zero-inflated" or "all" (default). May be abbreviated.

robust
Logical, if TRUE, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors. Additional arguments like vcov_estimation or vcov_type are passed down to other methods, see standard_error_robust() for details.

...
Arguments passed down to standard_error_robust() when confidence intervals or p-values based on robust standard errors should be computed.

Value

A data frame.

References

Barr, D. J. (2013). Random effects structure for testing interactions in linear mixed-effects models. Frontiers in psychology, 4, 328.

Examples

# \donttest{
if (require("lme4")) {
  model <- lmer(Petal.Length ~ Sepal.Length + (1 | Species), data = iris)
  p_value_wald(model)
  ci_wald(model, ci = c(0.90, 0.95))
}

#>      Parameter CI     CI_low   CI_high
#> 1  (Intercept) 90 -1.5296293 1.6346793
#> 2 Sepal.Length 90  0.5592337 0.7090407
#> 3  (Intercept) 95 -1.8364229 1.9414729
#> 4 Sepal.Length 95  0.5447092 0.7235652

# }