Parameters Selection

This function performs an automated selection of the 'best' parameters, updating and returning the "best" model. For frequentist simple GLMs, it performs an AIC-based stepwise selection. For Bayesian models, it uses the projpred package.

parameters_selection(model, ...)

# S3 method for lm
parameters_selection(model, direction = "both",
  steps = 1000, k = 2, ...)

# S3 method for merMod
parameters_selection(model, direction = "backward",
  steps = 1000, ...)

# S3 method for stanreg
parameters_selection(model, method = NULL,
  cross_validation = FALSE, ...)

Arguments

model
A statistical model.

...
Arguments passed to or from other methods.

direction
the mode of stepwise search, can be one of "both", "backward", or "forward", with a default of "both". If the scope argument is missing the default for direction is "backward". Values can be abbreviated.

steps
the maximum number of steps to be considered. The default is 1000 (essentially as many as required). It is typically used to stop the process early.

k
the multiple of the number of degrees of freedom used for the penalty. Only k = 2 gives the genuine AIC: k = log(n) is sometimes referred to as BIC or SBC.

method
The method used in the variable selection. Can be NULL (default), "forward" or "L1". See projpred::varsel.

cross_validation
Select with cross-validation.

Value

The model refitted with optimal number of parameters.

Examples

model <- lm(mpg ~ ., data = mtcars)
parameters_selection(model)

#> 
#> Call:
#> lm(formula = mpg ~ wt + qsec + am, data = mtcars)
#> 
#> Coefficients:
#> (Intercept)           wt         qsec           am  
#>       9.618       -3.917        1.226        2.936  
#> 

model <- lm(mpg ~ cyl * disp * hp * wt, data = mtcars)
parameters_selection(model)

#> 
#> Call:
#> lm(formula = mpg ~ cyl + disp + hp + wt + cyl:disp + cyl:hp + 
#>     disp:hp + cyl:wt + disp:wt + hp:wt + cyl:disp:hp + cyl:hp:wt, 
#>     data = mtcars)
#> 
#> Coefficients:
#> (Intercept)          cyl         disp           hp           wt     cyl:disp  
#>  49.1436077   -3.6167276   -1.2955318   -0.0004854   58.8328841    0.1704703  
#>      cyl:hp      disp:hp       cyl:wt      disp:wt        hp:wt  cyl:disp:hp  
#>  -0.0134573    0.0132124   -7.4915051   -0.0167172   -0.6524341   -0.0016542  
#>   cyl:hp:wt  
#>   0.0850798  
#> 

# \donttest{
# lme4 -------------------------------------------
library(lme4)
model <- lmer(Sepal.Width ~ Sepal.Length * Petal.Width * Petal.Length + (1 | Species), data = iris)
parameters_selection(model)

#> Linear mixed model fit by REML ['lmerMod']
#> Formula: Sepal.Width ~ Sepal.Length * Petal.Width * Petal.Length + (1 |  
#>     Species)
#>    Data: iris
#> REML criterion at convergence: 50.9896
#> Random effects:
#>  Groups   Name        Std.Dev.
#>  Species  (Intercept) 0.8259  
#>  Residual             0.2536  
#> Number of obs: 150, groups:  Species, 3
#> Fixed Effects:
#>                           (Intercept)                           Sepal.Length  
#>                             -2.000245                               0.936730  
#>                           Petal.Width                           Petal.Length  
#>                              1.575534                               0.265564  
#>              Sepal.Length:Petal.Width              Sepal.Length:Petal.Length  
#>                             -0.282960                              -0.088409  
#>              Petal.Width:Petal.Length  Sepal.Length:Petal.Width:Petal.Length  
#>                              0.001863                               0.023319  

# rstanarm -------------------------------------------
library(rstanarm)
model <- stan_glm(mpg ~ ., data = mtcars, iter = 500, refresh = 0)

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

parameters_selection(model, cross_validation = TRUE)

#> Cross-validating best parameters...

#> Warning: Some Pareto k diagnostic values are too high. See help('pareto-k-diagnostic') for details.

#> stan_glm
#>  family:       gaussian [identity]
#>  formula:      mpg ~ wt
#>  observations: 32
#>  predictors:   2
#> ------
#>             Median MAD_SD
#> (Intercept) 37.4    2.1  
#> wt          -5.4    0.6  
#> 
#> Auxiliary parameter(s):
#>       Median MAD_SD
#> sigma 3.1    0.4   
#> 
#> ------
#> * For help interpreting the printed output see ?print.stanreg
#> * For info on the priors used see ?prior_summary.stanreg

model <- stan_glm(mpg ~ cyl * disp * hp, data = mtcars)

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.131 seconds (Warm-up)
#> Chain 1:                1.486 seconds (Sampling)
#> Chain 1:                2.617 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.095 seconds (Warm-up)
#> Chain 2:                0.953 seconds (Sampling)
#> Chain 2:                2.048 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.231 seconds (Warm-up)
#> Chain 3:                1.186 seconds (Sampling)
#> Chain 3:                2.417 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.08 seconds (Warm-up)
#> Chain 4:                0.98 seconds (Sampling)
#> Chain 4:                2.06 seconds (Total)
#> Chain 4: 

parameters_selection(model, cross_validation = FALSE)

#> stan_glm
#>  family:       gaussian [identity]
#>  formula:      mpg ~ cyl + disp + hp + cyl:disp + cyl:hp
#>  observations: 32
#>  predictors:   6
#> ------
#>             Median MAD_SD
#> (Intercept) 50.9    5.6  
#> cyl         -3.2    0.9  
#> disp        -0.1    0.0  
#> hp          -0.1    0.1  
#> cyl:disp     0.0    0.0  
#> cyl:hp       0.0    0.0  
#> 
#> Auxiliary parameter(s):
#>       Median MAD_SD
#> sigma 2.6    0.4   
#> 
#> ------
#> * For help interpreting the printed output see ?print.stanreg
#> * For info on the priors used see ?prior_summary.stanreg

# }