Parameters simulation

Compute simulated draws of parameters and their related indices such as Confidence Intervals (CI) and p-values. Simulating parameter draws can be seen as a (computationally faster) alternative to bootstrapping.

parameters_simulate(model, ...)

# S3 method for default
parameters_simulate(model, iterations = 1000,
  centrality = "median", ci = 0.95, ci_method = "quantile",
  test = "p-value", ...)

Arguments

model
Statistical model (no Bayesian models).

...
Arguments passed to or from other methods.

iterations
The number of draws to simulate/bootstrap.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

ci
Value or vector of probability of the CI (between 0 and 1) to be estimated. Default to .89 (89%) for Bayesian models and .95 (95%) for frequentist models.

ci_method
The type of index used for Credible Interval. Can be "HDI" (default, see hdi) or "ETI" (see eti).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests. For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

Value

A data frame with simulated parameters.

Details

Technical Details

model_simulate() is a computationally faster alternative to model_bootstrap(). Simulated draws for coefficients are based on a multivariate normal distribution (MASS::mvrnorm()) with mean mu = coef(model) and variance Sigma = vcov(model).

Models with Zero-Inflation Component

For models from packages glmmTMB, pscl, GLMMadaptive and countreg, the component argument can be used to specify which parameters should be simulated. For all other models, parameters from the conditional component (fixed effects) are simulated. This may include smooth terms, but not random effects.

References

Gelman A, Hill J. Data analysis using regression and multilevel/hierarchical models. Cambridge; New York: Cambridge University Press 2007: 140-143

See also

Examples

library(parameters)
library(glmmTMB)

model <- lm(Sepal.Length ~ Species * Petal.Width + Petal.Length, data = iris)
parameters_simulate(model)

#>                       Parameter Coefficient     CI_low    CI_high           p
#> 1                   (Intercept)   3.5193526  3.2096629  3.8292219 0.000000000
#> 2             Speciesversicolor  -1.1754302 -1.8317978 -0.4755456 0.001998002
#> 3              Speciesvirginica  -2.2499403 -3.0368273 -1.3843288 0.000000000
#> 4                   Petal.Width   0.4251690 -0.4475586  1.3323834 0.337662338
#> 5                  Petal.Length   0.9393009  0.7920573  1.0943804 0.000000000
#> 6 Speciesversicolor:Petal.Width  -0.7250165 -1.8430475  0.1990345 0.123876124
#> 7  Speciesvirginica:Petal.Width  -0.3682232 -1.4148144  0.5244613 0.427572428

model <- glmmTMB(
  count ~ spp + mined + (1 | site),
  ziformula = ~mined,
  family = poisson(),
  data = Salamanders
)
parameters_simulate(model, centrality = "mean")

#>         Parameter Coefficient        CI_low    CI_high           p
#> 1     (Intercept)  -0.3475189 -0.9036283180  0.1966030 0.217782218
#> 2           sppPR  -1.2770436 -1.7600600446 -0.8112499 0.000000000
#> 3           sppDM   0.2704482  0.0003048768  0.5590856 0.049950050
#> 4         sppEC-A  -0.5628975 -0.9515413056 -0.1895242 0.001998002
#> 5         sppEC-L   0.6653810  0.4069161227  0.9167313 0.000000000
#> 6        sppDES-L   0.6244235  0.3793519782  0.8674894 0.000000000
#> 7           sppDF   0.1138468 -0.1845096024  0.3945827 0.455544456
#> 8         minedno   1.2532244  0.7413075702  1.7711593 0.000000000
#> 9  (Intercept)_zi   0.7971973  0.2631760856  1.3386471 0.000000000
#> 10     minedno_zi  -1.8578715 -2.4369912137 -1.2184980 0.000000000

parameters_simulate(model, ci = c(.8, .95), component = "zero_inflated")

#>     Parameter Coefficient CI     CI_low   CI_high           p
#> 1 (Intercept)   0.7957343 80  0.4508656  1.138326 0.007992008
#> 2 (Intercept)   0.7957343 95  0.2686625  1.307365 0.007992008
#> 3     minedno  -1.8523712 80 -2.2446124 -1.428804 0.000000000
#> 4     minedno  -1.8523712 95 -2.4656001 -1.213513 0.000000000