Pool Model Parameters

This function "pools" (i.e. combines) model parameters in a similar fashion as mice::pool(). However, this function pools parameters from parameters_model objects, as returned by model_parameters.

pool_parameters(
  x,
  exponentiate = FALSE,
  effects = "fixed",
  component = "conditional",
  verbose = TRUE,
  ...
)

Arguments

x
A list of parameters_model objects, as returned by model_parameters, or a list of model-objects that is supported by model_parameters().

exponentiate
Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for, say, logistic regressions, or more generally speaking: for models with log or logit link. Note: standard errors are also transformed (by multiplying the standard errors with the exponentiated coefficients), to mimic behaviour of other software packages, such as Stata. For compare_parameters(), exponentiate = "nongaussian" will only exponentiate coefficients for all models except those from Gaussian family.

effects
Should parameters for fixed effects ("fixed"), random effects ("random"), or both ("all") be returned? Only applies to mixed models. May be abbreviated.

component
Model component for which parameters should be shown. May be one of "conditional", "precision" (betareg), "scale" (ordinal), "extra" (glmx), "marginal" (mfx), "conditional" or "full" (for MuMIn::model.avg()) or "all".

verbose
Toggle warnings and messages.

...
Currently not used.

Value

A data frame of indices related to the model's parameters.

Details

Averaging of parameters follows Rubin's rules (Rubin, 1987, p. 76). The pooled degrees of freedom is based on the Barnard-Rubin adjustment for small samples (Barnard and Rubin, 1999).

Note

Models with multiple components, (for instance, models with zero-inflation, where predictors appear in the count and zero-inflated part) may fail in case of identical names for coefficients in the different model components, since the coefficient table is grouped by coefficient names for pooling. In such cases, coefficients of count and zero-inflated model parts would be combined. Therefore, the component argument defaults to "conditional" to avoid this.

References

Barnard, J. and Rubin, D.B. (1999). Small sample degrees of freedom with multiple imputation. Biometrika, 86, 948-955. Rubin, D.B. (1987). Multiple Imputation for Nonresponse in Surveys. New York: John Wiley and Sons.

Examples

# example for multiple imputed datasets
if (require("mice")) {
  data("nhanes2")
  imp <- mice(nhanes2, printFlag = FALSE)
  models <- lapply(1:5, function(i) {
    lm(bmi ~ age + hyp + chl, data = complete(imp, action = i))
  })
  pool_parameters(models)

  # should be identical to:
  m <- with(data = imp, exp = lm(bmi ~ age + hyp + chl))
  summary(pool(m))
}

#>          term    estimate  std.error statistic        df      p.value
#> 1 (Intercept) 18.14256305 3.54562901  5.116881 11.625350 0.0002813879
#> 2    age40-59 -6.15715380 2.19792337 -2.801351  5.969005 0.0312810653
#> 3    age60-99 -7.72866592 2.45997959 -3.141760  6.642762 0.0174969755
#> 4      hypyes  2.46673562 2.07396774  1.189380  6.147560 0.2781911871
#> 5         chl  0.06028557 0.02078061  2.901050 10.206904 0.0154916183