Principal Component Analysis (PCA)

This function performs a principal component analysis (PCA) and returns the loadings as a dataframe.

principal_components(x, n = "auto", rotation = "none", sort = FALSE,
  threshold = NULL, standardize = TRUE, ...)

Arguments

x
A dataframe or a statistical model.

n
Number of components to extract. If n = "all" =, then n is set as the number of variables minus 1 (ncol(x)-1). If n = "auto" (default) or n = NULL, the number of components is selected through n_factors. In parameters_reduction, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

rotation
If not "none", the PCA will be computed using the psych package. Possible options include "varimax", "quartimax", "promax", "oblimin", "simplimax", and "cluster". See fa for details.

sort
Sort the loadings.

threshold
A value between 0 and 1 indicates which (absolute) values from the loadings should be removed. An integer higher than 1 indicates the n strongest loadings to retain. Can also be "max", in which case it will only display the maximum loading per variable (the most simple structure).

standardize
A logical value indicating whether the variables should be standardized (centred and scaled) to have unit variance before the analysis takes place (in general, such scaling is advisable).

...
Arguments passed to or from other methods.

Value

A data.frame of loadings.

Details

Complexity (Hoffman's, 1978; Pettersson and Turkheimer, 2010) represents the number of latent components needed to account for the observed variables. Whereas a perfect simple structure solution has a complexity of 1 in that each item would only load on one factor, a solution with evenly distributed items has a complexity greater than 1.

Note

There is a summary()-method that prints the Eigenvalues and (explained) variance for each extracted component.

References

Pettersson, E., & Turkheimer, E. (2010). Item selection, evaluation, and simple structure in personality data. Journal of research in personality, 44(4), 407-420.

Examples

library(parameters)

principal_components(mtcars[, 1:7], n = "all", threshold = 0.2)

#> # Loadings from Principal Component Analysis (no rotation)
#> 
#> Variable |   PC1 |   PC2 |  PC3 |   PC4 |  PC5 |   PC6 | Complexity
#> -------------------------------------------------------------------
#> mpg      | -0.93 |       |      | -0.30 |      |       |       1.30
#> cyl      |  0.96 |       |      |       |      | -0.21 |       1.18
#> disp     |  0.95 |       |      | -0.23 |      |       |       1.16
#> hp       |  0.87 |  0.36 |      |       | 0.30 |       |       1.64
#> drat     | -0.75 |  0.48 | 0.44 |       |      |       |       2.47
#> wt       |  0.88 | -0.35 | 0.26 |       |      |       |       1.54
#> qsec     | -0.54 | -0.81 |      |       |      |       |       1.96
#> 
#> The 6 principal components accounted for 99.30% of the total variance of the original data (PC1 = 72.66%, PC2 = 16.52%, PC3 = 4.93%, PC4 = 2.26%, PC5 = 1.85%, PC6 = 1.08%).

principal_components(mtcars[, 1:7], n = 2, rotation = "oblimin", threshold = "max", sort = TRUE)

#> # Rotated loadings from Principal Component Analysis (oblimin-rotation)
#> 
#> Variable |   TC1 |   TC2 | Complexity | Uniqueness
#> --------------------------------------------------
#> wt       |  0.98 |       |       1.03 |       0.10
#> drat     | -0.95 |       |       1.19 |       0.21
#> disp     |  0.89 |       |       1.07 |       0.08
#> mpg      | -0.87 |       |       1.07 |       0.13
#> cyl      |  0.78 |       |       1.38 |       0.08
#> qsec     |       | -0.98 |       1.00 |       0.06
#> hp       |       |  0.61 |       1.97 |       0.10
#> 
#> The 2 principal components (oblimin rotation) accounted for 89.18% of the total variance of the original data (TC1 = 63.90%, TC2 = 25.28%).

principal_components(mtcars[, 1:7], n = 2, threshold = 2, sort = TRUE)

#> # Loadings from Principal Component Analysis (no rotation)
#> 
#> Variable |  PC1 |   PC2 | Complexity
#> ------------------------------------
#> cyl      | 0.96 |       |       1.02
#> disp     | 0.95 |       |       1.02
#> mpg      |      |       |       1.02
#> wt       |      |       |       1.30
#> hp       |      |       |       1.33
#> drat     |      |  0.48 |       1.71
#> qsec     |      | -0.81 |       1.75
#> 
#> The 2 principal components accounted for 89.18% of the total variance of the original data (PC1 = 72.66%, PC2 = 16.52%).

pca <- principal_components(mtcars[, 1:5], n = 2)
summary(pca)

#> # (Explained) Variance of Components
#> 
#> Parameter                       |   PC1 |   PC2
#> -----------------------------------------------
#> Eigenvalues                     | 4.038 | 0.562
#> Variance Explained              | 0.808 | 0.112
#> Variance Explained (Cumulative) | 0.808 | 0.920
#> Variance Explained (Proportion) | 0.808 | 0.112

predict(pca)

#>            PC1         PC2
#> 1  -0.83958112 -0.12417860
#> 2  -0.83958112 -0.12417860
#> 3  -1.78076155  0.20357157
#> 4   0.09834476  1.10598482
#> 5   1.58552749  0.32873919
#> 6   0.42628595  1.61371418
#> 7   2.31798915 -0.38579468
#> 8  -1.83392218  0.71554314
#> 9  -1.69378203  0.07797675
#> 10 -0.60554684 -0.27702188
#> 11 -0.49812301 -0.28570556
#> 12  1.52990575  0.40014511
#> 13  1.46084757  0.40572748
#> 14  1.62198332  0.39270195
#> 15  2.99730327  0.34308959
#> 16  2.96336452  0.15336639
#> 17  2.48424862 -0.29287006
#> 18 -2.96506160  0.14911053
#> 19 -3.52699413 -1.02550272
#> 20 -3.21693607 -0.04349485
#> 21 -1.50101897  0.38651333
#> 22  1.79695027  1.11397253
#> 23  1.48376642  0.52524948
#> 24  1.97953272 -1.17526822
#> 25  1.75060571  0.43376680
#> 26 -2.57258674  0.11744981
#> 27 -2.41238961 -0.63531022
#> 28 -2.22991874  0.20202457
#> 29  1.55523735 -2.06076910
#> 30 -0.18704672 -0.26276186
#> 31  2.36338454 -1.64100466
#> 32 -1.71202696 -0.33478623

# \donttest{
# Automated number of components
principal_components(mtcars[, 1:4], n = "auto")

#> # Loadings from Principal Component Analysis (no rotation)
#> 
#> Variable |   PC1 | Complexity
#> -----------------------------
#> mpg      | -0.93 |       1.00
#> cyl      |  0.96 |       1.00
#> disp     |  0.95 |       1.00
#> hp       |  0.91 |       1.00
#> 
#> The unique principal component accounted for 87.55% of the total variance of the original data.

# }