Principal Component Analysis (PCA)

This function performs a principal component analysis (PCA) and returns the loadings (of the unrotated matrix) as dataframe.

principal_components(x, n = NULL, sort = FALSE, threshold = NULL,
  standardize = TRUE, ...)

PCA(x, n = NULL, sort = FALSE, threshold = NULL,
  standardize = TRUE, ...)

Arguments

x
A dataframe.

n
Number of components to extract. If n = NULL, the number of components is selected through n_factors.

sort
Sort the loadings.

threshold
A value between 0 and 1 indicates which (absolute) values from the loadings should be removed. Can also be "max", in which case it will only display the maximum loading per veriable (the most simple structure).

standardize
A logical value indicating whether the variables should be standardized (centred and scaled) to have unit variance before the analysis takes place (in general, such scaling is advisable).

...
Arguments passed to or from other methods.

Details

Complexity (Hoffman's, 1978; Pettersson and Turkheimer, 2010) represents the number of latent components needed to account for the observed variables. Whereas a perfect simple structure solution has a complexity of 1 in that each item would only load on one factor, a solution with evenly distributed items has a complexity greater than 1.

References

Pettersson, E., \& Turkheimer, E. (2010). Item selection, evaluation, and simple structure in personality data. Journal of research in personality, 44(4), 407-420.

Examples

library(parameters)

# principal_components(mtcars[, 1:4])
principal_components(mtcars[, 1:7], n = "all", threshold = 0.2)

#> The 7 principal components accounted for 100.00% of the total variance of the original data (PC1 = 72.66%, PC2 = 16.52%, PC3 = 4.93%, PC4 = 2.26%, PC5 = 1.85%, PC6 = 1.08%, PC7 = 0.70%).
#> 
#>   Variable        PC1        PC2       PC3        PC4       PC5        PC6 PC7
#> 1      mpg -0.9308661         NA        NA -0.3047324        NA         NA  NA
#> 2      cyl  0.9578708         NA        NA         NA        NA -0.2149809  NA
#> 3     disp  0.9528463         NA        NA -0.2335489        NA         NA  NA
#> 4       hp  0.8744936  0.3623835        NA         NA 0.2992024         NA  NA
#> 5     drat -0.7468681  0.4824288 0.4435378         NA        NA         NA  NA
#> 6       wt  0.8825092 -0.3466793 0.2587107         NA        NA         NA  NA
#> 7     qsec -0.5410937 -0.8058477        NA         NA        NA         NA  NA
#>   Complexity
#> 1   1.315779
#> 2   1.184449
#> 3   1.208044
#> 4   1.638575
#> 5   2.474732
#> 6   1.597572
#> 7   1.965888

principal_components(mtcars[, 1:7], n = 2, threshold = "max", sort = TRUE)

#> The 2 principal components accounted for 89.18% of the total variance of the original data (PC1 = 72.66%, PC2 = 16.52%).
#> 
#>   Variable        PC1        PC2 Complexity
#> 1      cyl  0.9578708         NA   1.015511
#> 2     disp  0.9528463         NA   1.017297
#> 3      mpg -0.9308661         NA   1.018371
#> 4       wt  0.8825092         NA   1.301458
#> 5       hp  0.8744936         NA   1.333605
#> 6     drat -0.7468681         NA   1.710738
#> 7     qsec         NA -0.8058477   1.749385

pca <- principal_components(mtcars[, 1:5], n = 2)
summary(pca)

#>   Component Eigenvalues  Variance Variance_Cumulative
#> 1       PC1   4.0379396 0.8075879           0.8075879
#> 2       PC2   0.5622352 0.1124470           0.9200349

predict(pca)

#>            PC1         PC2
#> 1  -0.83958112 -0.12417860
#> 2  -0.83958112 -0.12417860
#> 3  -1.78076155  0.20357157
#> 4   0.09834476  1.10598482
#> 5   1.58552749  0.32873919
#> 6   0.42628595  1.61371418
#> 7   2.31798915 -0.38579468
#> 8  -1.83392218  0.71554314
#> 9  -1.69378203  0.07797675
#> 10 -0.60554684 -0.27702188
#> 11 -0.49812301 -0.28570556
#> 12  1.52990575  0.40014511
#> 13  1.46084757  0.40572748
#> 14  1.62198332  0.39270195
#> 15  2.99730327  0.34308959
#> 16  2.96336452  0.15336639
#> 17  2.48424862 -0.29287006
#> 18 -2.96506160  0.14911053
#> 19 -3.52699413 -1.02550272
#> 20 -3.21693607 -0.04349485
#> 21 -1.50101897  0.38651333
#> 22  1.79695027  1.11397253
#> 23  1.48376642  0.52524948
#> 24  1.97953272 -1.17526822
#> 25  1.75060571  0.43376680
#> 26 -2.57258674  0.11744981
#> 27 -2.41238961 -0.63531022
#> 28 -2.22991874  0.20202457
#> 29  1.55523735 -2.06076910
#> 30 -0.18704672 -0.26276186
#> 31  2.36338454 -1.64100466
#> 32 -1.71202696 -0.33478623