Principal Component Analysis (PCA)

This function performs a principal component analysis (PCA) and returns the loadings as a data frame.

principal_components(
  x,
  n = "auto",
  rotation = "none",
  sort = FALSE,
  threshold = NULL,
  standardize = TRUE,
  ...
)

closest_component(x)

Arguments

x
A data frame or a statistical model.

n
Number of components to extract. If n="all", then n is set as the number of variables minus 1 (ncol(x)-1). If n="auto" (default) or n=NULL, the number of components is selected through n_factors. In reduce_parameters, can also be "max", in which case it will select all the components that are maximally pseudo-loaded (i.e., correlated) by at least one variable.

rotation
If not "none", the PCA / FA will be computed using the psych package. Possible options include "varimax", "quartimax", "promax", "oblimin", "simplimax", and "cluster". See fa for details.

sort
Sort the loadings.

threshold
A value between 0 and 1 indicates which (absolute) values from the loadings should be removed. An integer higher than 1 indicates the n strongest loadings to retain. Can also be "max", in which case it will only display the maximum loading per variable (the most simple structure).

standardize
A logical value indicating whether the variables should be standardized (centered and scaled) to have unit variance before the analysis takes place (in general, such scaling is advisable).

...
Arguments passed to or from other methods.

Value

A data frame of loadings.

Details

Complexity

Complexity represents the number of latent components needed to account for the observed variables. Whereas a perfect simple structure solution has a complexity of 1 in that each item would only load on one factor, a solution with evenly distributed items has a complexity greater than 1 (Hofman, 1978; Pettersson and Turkheimer, 2010) .

Uniqueness

Uniqueness represents the variance that is 'unique' to the variable and not shared with other variables. It is equal to 1 – communality (variance that is shared with other variables). A uniqueness of 0.20 suggests that 20% or that variable's variance is not shared with other variables in the overall factor model. The greater 'uniqueness' the lower the relevance of the variable in the factor model.

MSA

MSA represents the Kaiser-Meyer-Olkin Measure of Sampling Adequacy (Kaiser and Rice, 1974) for each item. It indicates whether there is enough data for each factor give reliable results for the PCA. The value should be > 0.6, and desirable values are > 0.8 (Tabachnick and Fidell, 2013).

PCA or FA?

There is a simplified rule of thumb that may help do decide whether to run a factor analysis or a principal component analysis:

Run factor analysis if you assume or wish to test a theoretical model of latent factors causing observed variables.

Run principal component analysis If you want to simply reduce your correlated observed variables to a smaller set of important independent composite variables.

(Source: CrossValidated)

Note

There is a summary()-method that prints the Eigenvalues and (explained) variance for each extracted component. closest_component() will return a numeric vector with the assigned component index for each column from the original data frame. There is also a plot()-method implemented in the see-package.

References

Kaiser, H.F. and Rice. J. (1974). Little jiffy, mark iv. Educational and Psychological Measurement, 34(1):111–117

Hofmann, R. (1978). Complexity and simplicity as objective indices descriptive of factor solutions. Multivariate Behavioral Research, 13:2, 247-250, doi: 10.1207/s15327906mbr1302_9

Pettersson, E., & Turkheimer, E. (2010). Item selection, evaluation, and simple structure in personality data. Journal of research in personality, 44(4), 407-420, doi: 10.1016/j.jrp.2010.03.002

Tabachnick, B. G., and Fidell, L. S. (2013). Using multivariate statistics (6th ed.). Boston: Pearson Education.

See also

Examples

library(parameters)

principal_components(mtcars[, 1:7], n = "all", threshold = 0.2)

#> # Loadings from Principal Component Analysis (no rotation)
#> 
#> Variable |   PC1 |   PC2 |  PC3 |   PC4 |  PC5 |   PC6 | Complexity
#> -------------------------------------------------------------------
#> mpg      | -0.93 |       |      | -0.30 |      |       |       1.30
#> cyl      |  0.96 |       |      |       |      | -0.21 |       1.18
#> disp     |  0.95 |       |      | -0.23 |      |       |       1.16
#> hp       |  0.87 |  0.36 |      |       | 0.30 |       |       1.64
#> drat     | -0.75 |  0.48 | 0.44 |       |      |       |       2.47
#> wt       |  0.88 | -0.35 | 0.26 |       |      |       |       1.54
#> qsec     | -0.54 | -0.81 |      |       |      |       |       1.96
#> 
#> The 6 principal components accounted for 99.30% of the total variance of the original data (PC1 = 72.66%, PC2 = 16.52%, PC3 = 4.93%, PC4 = 2.26%, PC5 = 1.85%, PC6 = 1.08%).

principal_components(mtcars[, 1:7], n = 2, rotation = "oblimin", threshold = "max", sort = TRUE)

#> # Rotated loadings from Principal Component Analysis (oblimin-rotation)
#> 
#> Variable |   TC1 |   TC2 | Complexity | Uniqueness |  MSA
#> ---------------------------------------------------------
#> wt       |  0.98 |       |       1.03 |       0.10 | 0.77
#> drat     | -0.95 |       |       1.19 |       0.21 | 0.85
#> disp     |  0.89 |       |       1.07 |       0.08 | 0.85
#> mpg      | -0.87 |       |       1.07 |       0.13 | 0.87
#> cyl      |  0.78 |       |       1.38 |       0.08 | 0.87
#> qsec     |       | -0.98 |       1.00 |       0.06 | 0.61
#> hp       |       |  0.61 |       1.97 |       0.10 | 0.90
#> 
#> The 2 principal components (oblimin rotation) accounted for 89.18% of the total variance of the original data (TC1 = 63.90%, TC2 = 25.28%).

principal_components(mtcars[, 1:7], n = 2, threshold = 2, sort = TRUE)

#> # Loadings from Principal Component Analysis (no rotation)
#> 
#> Variable |  PC1 |   PC2 | Complexity
#> ------------------------------------
#> cyl      | 0.96 |       |       1.02
#> disp     | 0.95 |       |       1.02
#> mpg      |      |       |       1.02
#> wt       |      |       |       1.30
#> hp       |      |       |       1.33
#> drat     |      |  0.48 |       1.71
#> qsec     |      | -0.81 |       1.75
#> 
#> The 2 principal components accounted for 89.18% of the total variance of the original data (PC1 = 72.66%, PC2 = 16.52%).

pca <- principal_components(mtcars[, 1:5], n = 2, rotation = "varimax")
summary(pca)

#> # (Explained) Variance of Components
#> 
#> Parameter                       |   RC1 |   RC2
#> -----------------------------------------------
#> Eigenvalues                     | 4.038 | 0.562
#> Variance Explained              | 0.565 | 0.355
#> Variance Explained (Cumulative) | 0.565 | 0.920
#> Variance Explained (Proportion) | 0.614 | 0.386

predict(pca)

#>            RC1         RC2
#> 1  -0.23906186  0.38058456
#> 2  -0.23906186  0.38058456
#> 3  -0.87523403  0.30496365
#> 4  -0.83256507 -1.21853890
#> 5   0.37716377 -0.82007976
#> 6  -1.10127278 -1.86114104
#> 7   1.23453249 -0.26702364
#> 8  -1.30024506 -0.23007562
#> 9  -0.74129623  0.41446358
#> 10 -0.02461777  0.47612740
#> 11  0.02534454  0.45386195
#> 12  0.29853804 -0.88051969
#> 13  0.26641941 -0.86620618
#> 14  0.34136287 -0.89960437
#> 15  0.93247714 -1.25088059
#> 16  1.06844670 -1.03682781
#> 17  1.22799380 -0.41588998
#> 18 -1.30762202  0.71198099
#> 19 -0.60701035  2.14073148
#> 20 -1.25685030  0.99325379
#> 21 -0.90720118  0.02588588
#> 22 -0.15711335 -1.72688554
#> 23  0.18137850 -1.00150788
#> 24  1.72116683  0.68171810
#> 25  0.36060813 -0.98161661
#> 26 -1.12513535  0.63056351
#> 27 -0.46730967  1.39310467
#> 28 -1.05428733  0.43877634
#> 29  2.24906373  1.75900256
#> 30  0.13210722  0.33766027
#> 31  2.24244855  1.06973229
#> 32 -0.42316749  0.86380202

# which variables from the original data belong to which extracted component?
closest_component(pca)

#>  mpg  cyl disp   hp drat 
#>    1    1    1    1    2 

# \donttest{
# Automated number of components
principal_components(mtcars[, 1:4], n = "auto")

#> # Loadings from Principal Component Analysis (no rotation)
#> 
#> Variable |   PC1 | Complexity
#> -----------------------------
#> mpg      | -0.93 |       1.00
#> cyl      |  0.96 |       1.00
#> disp     |  0.95 |       1.00
#> hp       |  0.91 |       1.00
#> 
#> The unique principal component accounted for 87.55% of the total variance of the original data.

# }