Simulated draws from model coefficients

Simulate draws from a statistical model to return a data frame of estimates.

simulate_model(model, iterations = 1000, ...)

model_simulate(model, iterations = 1000, ...)

# S3 method for glmmTMB
simulate_model(
  model,
  iterations = 1000,
  component = c("all", "conditional", "zi", "zero_inflated"),
  ...
)

Arguments

model
Statistical model (no Bayesian models).

iterations
The number of draws to simulate/bootstrap.

...
Arguments passed to or from other methods.

component
Should all parameters, parameters for the conditional model, or for the zero-inflated part of the model be returned? Applies to models with zero-inflated component. component may be one of "conditional", "zi", "zero-inflated" or "all" (default). May be abbreviated.

Value

A data frame.

Details

Technical Details

simulate_model() is a computationally faster alternative to bootstrap_model(). Simulated draws for coefficients are based on a multivariate normal distribution (MASS::mvrnorm()) with mean mu = coef(model) and variance Sigma = vcov(model).

Models with Zero-Inflation Component

For models from packages glmmTMB, pscl, GLMMadaptive and countreg, the component argument can be used to specify which parameters should be simulated. For all other models, parameters from the conditional component (fixed effects) are simulated. This may include smooth terms, but not random effects.

See also

Examples

library(parameters)
library(glmmTMB)

model <- lm(Sepal.Length ~ Species * Petal.Width + Petal.Length, data = iris)
head(simulate_model(model))

#>   (Intercept) Speciesversicolor Speciesvirginica Petal.Width Petal.Length
#> 1    3.404895        -1.4495523        -1.658974  1.04226519    0.8944799
#> 2    3.594144        -1.3131367        -1.982839  0.52434634    0.8687698
#> 3    3.285108        -0.8198478        -2.809915  0.66981567    1.0335770
#> 4    3.543699        -0.9742509        -2.158027 -0.01023858    1.0307308
#> 5    3.712911        -0.7844119        -3.267886 -0.42696354    0.9873675
#> 6    3.626768        -1.3156968        -1.590338 -0.14994659    0.9139153
#>   Speciesversicolor:Petal.Width Speciesvirginica:Petal.Width
#> 1                    -0.8729099                  -1.12953933
#> 2                    -0.5673161                  -0.42988776
#> 3                    -1.3881523                  -0.45778575
#> 4                    -0.7451538                  -0.19654116
#> 5                    -0.4686398                   0.70172417
#> 6                    -0.1066046                  -0.08454052

model <- glmmTMB(
  count ~ spp + mined + (1 | site),
  ziformula = ~mined,
  family = poisson(),
  data = Salamanders
)
head(simulate_model(model))

#>   (Intercept)     sppPR     sppDM    sppEC-A   sppEC-L  sppDES-L      sppDF
#> 1 -0.04358525 -1.323612 0.2021646 -0.8251109 0.7606523 0.6202670 0.05885506
#> 2 -0.55857284 -1.371607 0.3515352 -0.5458674 0.6728705 0.5908258 0.11513497
#> 3 -0.41281277 -1.304294 0.1976165 -0.3772979 0.7708132 0.5506967 0.15609011
#> 4 -0.58463323 -1.371095 0.2457783 -0.3849654 0.8851888 0.7818880 0.14630199
#> 5  0.14979521 -1.308963 0.1975455 -0.4483306 0.6163035 0.6089289 0.13172565
#> 6 -0.66886572 -1.330268 0.3436324 -0.4475572 0.6902804 0.6183474 0.15638409
#>    minedno (Intercept)_zi minedno_zi
#> 1 1.025334      0.8213169  -1.528439
#> 2 1.440005      0.4374825  -1.605004
#> 3 1.254055      0.8576511  -1.778367
#> 4 1.542409      0.7179802  -1.680469
#> 5 0.962958      0.8669426  -1.767918
#> 6 1.441789      1.0695663  -1.985709

head(simulate_model(model, component = "zero_inflated"))

#>   (Intercept)   minedno
#> 1   0.9302574 -1.902361
#> 2   0.8314003 -1.924865
#> 3   0.1053317 -1.078260
#> 4   0.6856044 -1.833492
#> 5   0.7881163 -2.011245
#> 6   0.8491423 -2.075195