Simulate Model Parameters

Compute simulated draws of parameters and their related indices such as Confidence Intervals (CI) and p-values. Simulating parameter draws can be seen as a (computationally faster) alternative to bootstrapping.

simulate_parameters(model, ...)

# S3 method for default
simulate_parameters(
  model,
  iterations = 1000,
  centrality = "median",
  ci = 0.95,
  ci_method = "quantile",
  test = "p-value",
  ...
)

# S3 method for glmmTMB
simulate_parameters(
  model,
  iterations = 1000,
  centrality = "median",
  ci = 0.95,
  ci_method = "quantile",
  test = "p-value",
  ...
)

Arguments

model
Statistical model (no Bayesian models).

...
Arguments passed to or from other methods.

iterations
The number of draws to simulate/bootstrap.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

ci
Value or vector of probability of the CI (between 0 and 1) to be estimated. Default to .89 (89%) for Bayesian models and .95 (95%) for frequentist models.

ci_method
The type of index used for Credible Interval. Can be "HDI" (default, see hdi), "ETI" (see eti) or "SI" (see si).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests. For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

Value

A data frame with simulated parameters.

Details

Technical Details

simulate_parameters() is a computationally faster alternative to bootstrap_parameters(). Simulated draws for coefficients are based on a multivariate normal distribution (MASS::mvrnorm()) with mean mu = coef(model) and variance Sigma = vcov(model).

Models with Zero-Inflation Component

For models from packages glmmTMB, pscl, GLMMadaptive and countreg, the component argument can be used to specify which parameters should be simulated. For all other models, parameters from the conditional component (fixed effects) are simulated. This may include smooth terms, but not random effects.

Note

There is also a plot()-method implemented in the see-package.

References

Gelman A, Hill J. Data analysis using regression and multilevel/hierarchical models. Cambridge; New York: Cambridge University Press 2007: 140-143

See also

Examples

library(parameters)

model <- lm(Sepal.Length ~ Species * Petal.Width + Petal.Length, data = iris)
simulate_parameters(model)

#> Parameter                     | Coefficient |         95% CI |      p
#> ---------------------------------------------------------------------
#> (Intercept)                   |        3.53 | [ 3.21,  3.86] | < .001
#> Speciesversicolor             |       -1.17 | [-1.83, -0.52] | < .001
#> Speciesvirginica              |       -2.26 | [-3.08, -1.47] | < .001
#> Petal.Width                   |        0.42 | [-0.52,  1.29] | 0.392 
#> Petal.Length                  |        0.94 | [ 0.79,  1.10] | < .001
#> Speciesversicolor:Petal.Width |       -0.75 | [-1.73,  0.33] | 0.172 
#> Speciesvirginica:Petal.Width  |       -0.39 | [-1.30,  0.61] | 0.470 

# \donttest{
if (require("glmmTMB")) {
  model <- glmmTMB(
    count ~ spp + mined + (1 | site),
    ziformula = ~mined,
    family = poisson(),
    data = Salamanders
  )
  simulate_parameters(model, centrality = "mean")
  simulate_parameters(model, ci = c(.8, .95), component = "zero_inflated")
}

#> Parameter   | Coefficient |    CI |         80% CI | 95% CI |      p
#> --------------------------------------------------------------------
#> (Intercept) |        0.78 | 80.00 | [ 0.44,  1.12] |        | 0.006 
#> (Intercept) |        0.78 | 95.00 | [ 0.25,  1.34] |        | 0.006 
#> minedno     |       -1.83 | 80.00 | [-2.26, -1.46] |        | < .001
#> minedno     |       -1.83 | 95.00 | [-2.47, -1.24] |        | < .001

# }