Extract standard errors

This function attempts to return standard errors of model parameters.

standard_error(model, ...)

se(model, ...)

# S3 method for factor
standard_error(model, force = FALSE, verbose = TRUE,
  ...)

# S3 method for glmmTMB
standard_error(model, component = c("all",
  "conditional", "zi", "zero_inflated"), ...)

# S3 method for MixMod
standard_error(model, component = c("all",
  "conditional", "zi", "zero_inflated"), ...)

Arguments

model
A model.

...
Arguments passed to or from other methods.

force
Logical, if TRUE, factors are converted to numerical values to calculate the standard error, with the lowest level being the value 1 (unless the factor has numeric levels, which are converted to the corresponding numeric value). By default, NA is returned for factors or character vectors.

verbose
Toggle warnings on or off.

component
Should all parameters, parameters for the conditional model, or for the zero-inflated part of the model be returned? Applies to models with zero-inflated component. component may be one of "conditional", "zi", "zero-inflated" or "all" (default). May be abbreviated.

Value

A data frame.

Examples

model <- lm(Petal.Length ~ Sepal.Length * Species, data = iris)
standard_error(model)

#>                        Parameter        SE
#> 1                    (Intercept) 0.5310388
#> 2                   Sepal.Length 0.1058237
#> 3              Speciesversicolor 0.6836543
#> 4               Speciesvirginica 0.6578142
#> 5 Sepal.Length:Speciesversicolor 0.1281447
#> 6  Sepal.Length:Speciesvirginica 0.1209952