Standard Errors

standard_error() attempts to return standard errors of model parameters, while standard_error_robust() attempts to return robust standard errors.

standard_error(model, ...)

# S3 method for factor
standard_error(model, force = FALSE, verbose = TRUE, ...)

# S3 method for default
standard_error(model, method = NULL, ...)

# S3 method for merMod
standard_error(model, effects = c("fixed", "random"), method = NULL, ...)

# S3 method for glmmTMB
standard_error(
  model,
  effects = c("fixed", "random"),
  component = c("all", "conditional", "zi", "zero_inflated"),
  ...
)

# S3 method for MixMod
standard_error(
  model,
  effects = c("fixed", "random"),
  component = c("all", "conditional", "zi", "zero_inflated"),
  ...
)

# S3 method for zeroinfl
standard_error(
  model,
  component = c("all", "conditional", "zi", "zero_inflated"),
  method = NULL,
  ...
)

# S3 method for coxph
standard_error(model, method = NULL, ...)

# S3 method for mixor
standard_error(model, effects = c("all", "fixed", "random"), ...)

# S3 method for clm2
standard_error(model, component = c("all", "conditional", "scale"), ...)

# S3 method for betareg
standard_error(model, component = c("all", "conditional", "precision"), ...)

# S3 method for DirichletRegModel
standard_error(model, component = c("all", "conditional", "precision"), ...)

Arguments

model
A model.

...
Arguments passed to or from other methods. For standard_error(), if method = "robust", arguments vcov_estimation, vcov_type and vcov_args can be passed down to standard_error_robust().

force
Logical, if TRUE, factors are converted to numerical values to calculate the standard error, with the lowest level being the value 1 (unless the factor has numeric levels, which are converted to the corresponding numeric value). By default, NA is returned for factors or character vectors.

verbose
Toggle off warnings.

method
If "robust", robust standard errors are computed by calling standard_error_robust(). standard_error_robust(), in turn, calls one of the vcov*()-functions from the sandwich or clubSandwich package for robust covariance matrix estimators. For certain mixed models, method may also be one of "wald", "ml1", "betwithin", "satterthwaite" or "kenward".

effects
Should standard errors for fixed effects or random effects be returned? Only applies to mixed models. May be abbreviated. When standard errors for random effects are requested, for each grouping factor a list of standard errors (per group level) for random intercepts and slopes is returned.

component
Should all parameters, parameters for the conditional model, or for the zero-inflated part of the model be returned? Applies to models with zero-inflated component. component may be one of "conditional", "zi", "zero-inflated" or "all" (default). May be abbreviated.

Value

A data frame.

Note

For Bayesian models (from rstanarm or brms), the standard error is the SD of the posterior samples.

Examples

model <- lm(Petal.Length ~ Sepal.Length * Species, data = iris)
standard_error(model)

#>                        Parameter        SE
#> 1                    (Intercept) 0.5310388
#> 2                   Sepal.Length 0.1058237
#> 3              Speciesversicolor 0.6836543
#> 4               Speciesvirginica 0.6578142
#> 5 Sepal.Length:Speciesversicolor 0.1281447
#> 6  Sepal.Length:Speciesvirginica 0.1209952